Split video/window.rs into submodules
window.rs was 10.8k lines mixing the winit App with drawing and
mapping helpers, plus a 2.5k-line test module. Same module family, no
functional change: helpers move to child modules as pub(super) (the
drawing primitives ui.rs shares stay video-visible via re-exports),
and the test suite moves out-of-line as the same window::tests module:

- window/statusbar.rs   status-bar layout, hit rectangles, and glyph
                        rasterisers
- window/present.rs     render-job post-processing, texture scaling,
                        present-frame copies, TV mask, PNG capture
- window/host_input.rs  host keycode -> Amiga rawkey mapping, key-name
                        parsing, calibration view
- window/tests.rs       the unit suite, via use super::*

Unit suite and raw-screenshot byte-identity verified against the
pre-split tree.
