+++ src/video/window.rs
@@ -1910,238 +1910,6 @@ fn set_mouse_button(emu: &mut Emulator, button: MouseButtonKind, pressed: bool) 
     emu.tt_note_input(crate::inputsched::ReplayAction::MouseButton { index, pressed });
 }
 
-fn render_job_to_presentation(
-    generation: u64,
-    input: &bitplane::RenderInput,
-    fb: &mut [u32],
-    deinterlacer: &mut Deinterlacer,
-    h_shift: usize,
-    overscan: Overscan,
-    presentation_fb: &mut Vec<u32>,
-) -> RenderWorkerResult {
-    let render_result = bitplane::render_from_input(input, fb);
-    let geometry = input.geometry();
-    let visible_start_vpos = input.visible_start_vpos();
-    let field_rows =
-        post_process_rendered_field(fb, geometry, visible_start_vpos, h_shift, overscan);
-    let base = input.render_base();
-    deinterlacer.push_field(
-        fb,
-        field_rows,
-        base.bplcon0 & 0x0004 != 0,
-        base.long_field,
-        !geometry.programmable,
-    );
-    let present_rows = deinterlacer.output_rows();
-    let active = present_rows * FB_WIDTH;
-    presentation_fb.resize(active, 0);
-    presentation_fb.copy_from_slice(&deinterlacer.output()[..active]);
-    RenderWorkerResult {
-        generation,
-        emulated_frame: input.emulated_frames(),
-        timing: render_result.timing,
-        presentation_fb: std::mem::take(presentation_fb),
-        present_rows,
-        standard_tv_aperture: uses_standard_pal_tv_aperture(geometry, present_rows, &base),
-    }
-}
-
-fn post_process_rendered_field(
-    fb: &mut [u32],
-    geometry: FrameGeometry,
-    visible_start_vpos: u32,
-    h_shift: usize,
-    overscan: Overscan,
-) -> usize {
-    let field_rows = geometry.visible_lines.min(fb.len() / FB_WIDTH);
-    // Vertical centring, optional full-overscan horizontal recentring, and the
-    // TV bezel mask are 15 kHz CRT concepts anchored to the standard PAL/NTSC
-    // window; a programmable scan defines its own window and presents in full,
-    // like a multisync monitor.
-    if !geometry.programmable {
-        center_present_frame_for_visible_start(fb, visible_start_vpos);
-        center_present_frame_horizontally(fb, h_shift);
-        if overscan == Overscan::Tv {
-            mask_present_frame_to_tv(fb, h_shift, standard_window_top_row(visible_start_vpos));
-        }
-    } else if geometry.line_cck != 227 {
-        // A multisync monitor's horizontal deflection is time-linear:
-        // each colour clock of this scan's shorter/longer line covers
-        // 227/line_cck of the glass a standard line's clock would.
-        screenshot::stretch_rows_x(fb, FB_WIDTH, field_rows, geometry.line_cck, 227);
-    }
-    field_rows
-}
-
-fn log_frame_dump_metadata(index: u32, emu: &Emulator) {
-    let bus = emu.bus();
-    let base = bus.frame_render_base();
-    let (top, bottom, bottom_valid) = bus.frame_palette_split();
-    let events = bus.frame_render_events();
-    let captured_rows = bus
-        .frame_captured_bitplane_rows()
-        .iter()
-        .filter(|row| row.is_some())
-        .count();
-    let mut cpu_colors = 0usize;
-    let mut copper_colors = 0usize;
-    let mut control_events = 0usize;
-    for event in events {
-        let off = event.offset & 0x01FE;
-        if matches!(off, 0x180..=0x1BE) {
-            match event.source {
-                BeamWriteSource::Cpu | BeamWriteSource::CpuCopperIrq => cpu_colors += 1,
-                BeamWriteSource::Copper => copper_colors += 1,
-            }
-        }
-        if matches!(off, 0x100 | 0x102 | 0x104) {
-            control_events += 1;
-        }
-    }
-    info!(
-        "frame-meta idx={} emu_frame={} emu_secs={:.3} pc={:#08X} beam=({}, {}) dmacon={:#06X} bplcon0={:#06X} bplcon1={:#06X} bplcon2={:#06X} bplpt={:06X?} sprpt={:06X?} sprctl={:04X?} bplmod=({},{}) ddf=({:#05X},{:#05X}) diw=({:#06X},{:#06X}) base_pal={:03X?} top_pal={:03X?} bottom_valid={} bottom_pal={:03X?} events={} cpu_colors={} copper_colors={} controls={} captured_rows={}",
-        index,
-        bus.emulated_frames(),
-        bus.emulated_seconds(),
-        emu.machine.pc(),
-        bus.agnus.vpos,
-        bus.agnus.hpos,
-        base.dmacon,
-        base.bplcon0,
-        base.bplcon1,
-        base.bplcon2,
-        base.bplpt,
-        base.sprpt,
-        base.sprctl,
-        base.bpl1mod,
-        base.bpl2mod,
-        base.ddfstrt,
-        base.ddfstop,
-        base.diwstrt,
-        base.diwstop,
-        &base.palette.hi_words()[..16],
-        &top.hi_words()[..16],
-        bottom_valid,
-        &bottom.hi_words()[..16],
-        events.len(),
-        cpu_colors,
-        copper_colors,
-        control_events,
-        captured_rows
-    );
-    if crate::envcfg::flag("COPPERLINE_DUMP_RENDER_META_VERBOSE") {
-        let render_events: Vec<_> = events
-            .iter()
-            .map(|event| {
-                (
-                    event.vpos,
-                    event.hpos,
-                    event.offset & 0x01FE,
-                    event.value,
-                    event.source,
-                )
-            })
-            .collect();
-        info!(
-            "frame-meta-events idx={} events={:03X?}",
-            index, render_events
-        );
-        let color_events: Vec<_> = events
-            .iter()
-            .filter_map(|event| {
-                let off = event.offset & 0x01FE;
-                matches!(off, 0x180..=0x1BE).then_some((
-                    event.vpos,
-                    event.hpos,
-                    off,
-                    event.value & 0x0FFF,
-                    event.source,
-                ))
-            })
-            .collect();
-        info!(
-            "frame-meta-colors idx={} events={:03X?}",
-            index, color_events
-        );
-        let setup_events: Vec<_> = events
-            .iter()
-            .filter_map(|event| {
-                let off = event.offset & 0x01FE;
-                matches!(
-                    off,
-                    0x08E | 0x090 | 0x092 | 0x094 | 0x100..=0x10A | 0x0E0..=0x0F7
-                )
-                .then_some((
-                    event.vpos,
-                    event.hpos,
-                    off,
-                    event.value,
-                    event.source,
-                ))
-            })
-            .collect();
-        info!(
-            "frame-meta-setup idx={} events={:03X?}",
-            index, setup_events
-        );
-        let mut nonzero_ranges: [(Option<usize>, Option<usize>); 6] =
-            std::array::from_fn(|_| (None, None));
-        let mut row_shape_ranges: Vec<(usize, usize, usize, usize)> = Vec::new();
-        for (y, row) in bus.frame_captured_bitplane_rows().iter().enumerate() {
-            let Some(row) = row else {
-                continue;
-            };
-            match row_shape_ranges.last_mut() {
-                Some((_, end, nplanes, words_per_row))
-                    if *end + 1 == y
-                        && *nplanes == row.nplanes
-                        && *words_per_row == row.words_per_row =>
-                {
-                    *end = y;
-                }
-                _ => row_shape_ranges.push((y, y, row.nplanes, row.words_per_row)),
-            }
-            for (plane, range) in nonzero_ranges.iter_mut().enumerate() {
-                if row.planes[plane].iter().any(|word| *word != 0) {
-                    if range.0.is_none() {
-                        range.0 = Some(y);
-                    }
-                    range.1 = Some(y);
-                }
-            }
-        }
-        info!(
-            "frame-meta-bitplanes idx={} nonzero_ranges={:?} row_shape_ranges={:?}",
-            index, nonzero_ranges, row_shape_ranges
-        );
-        let bottom_events: Vec<_> = bus
-            .frame_bottom_palette_events()
-            .iter()
-            .map(|event| {
-                (
-                    event.vpos,
-                    event.hpos,
-                    event.offset & 0x01FE,
-                    event.value & 0x0FFF,
-                    event.source,
-                )
-            })
-            .collect();
-        info!(
-            "frame-meta-bottom-events idx={} events={:03X?}",
-            index, bottom_events
-        );
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(super) struct Rect {
-    pub(super) x: usize,
-    pub(super) y: usize,
-    pub(super) w: usize,
-    pub(super) h: usize,
-}
-
 impl Rect {
     pub(super) fn contains(self, pos: (i32, i32)) -> bool {
         let (x, y) = pos;
@@ -2266,2311 +2034,6 @@ fn decode_embedded_png(bytes: &[u8]) -> Result<EmbeddedRgbaImage> {
         height,
         rgba,
     })
-}
-
-fn texture_scale_for_window(window: &Window) -> usize {
-    texture_scale_for_factor(window.scale_factor())
-}
-
-/// Integer supersample factor for the backing texture at a given host DPI
-/// scale factor. The texture is rendered at this multiple of the logical
-/// FB_WIDTH x window-height size so a 2x display stays crisp.
-fn texture_scale_for_factor(scale_factor: f64) -> usize {
-    (scale_factor.round() as usize).clamp(1, MAX_TEXTURE_SCALE)
-}
-
-/// React to a host DPI scale-factor change for one window's pixel surface.
-///
-/// pixels' `window_pos_to_pixel` maps a host click into texture space using
-/// both the surface size (which the following Resized event updates) and the
-/// texture extent (which nothing updated before this). When the supersample
-/// factor changes -- e.g. dragging between a 1x and a 2x monitor -- the texture
-/// must be rebuilt to the new size, otherwise the two halves of the mapping
-/// disagree and clicks land in the wrong region. The rebuild reallocates a GPU
-/// texture, so it is skipped when the rounded factor is unchanged (a slow drag
-/// across a fractional-scale monitor seam can emit many events); the surface
-/// itself is re-synced by the Resized event that always follows.
-fn resync_render_scale(pixels: &mut Pixels<'static>, texture_scale: &mut usize, scale_factor: f64) {
-    let new_scale = texture_scale_for_factor(scale_factor);
-    if new_scale == *texture_scale {
-        return;
-    }
-    match pixels.resize_buffer(
-        texture_width(new_scale) as u32,
-        texture_height(new_scale) as u32,
-    ) {
-        Ok(()) => *texture_scale = new_scale,
-        Err(e) => warn!("resize texture buffer for scale {scale_factor} failed: {e}"),
-    }
-}
-
-fn build_pixels_for_window(
-    window: Arc<Window>,
-    texture_scale: usize,
-    vsync: bool,
-) -> std::result::Result<Pixels<'static>, pixels::Error> {
-    let inner = window.inner_size();
-    let surface = SurfaceTexture::new(inner.width.max(1), inner.height.max(1), window);
-    let builder = PixelsBuilder::new(
-        texture_width(texture_scale) as u32,
-        texture_height(texture_scale) as u32,
-        surface,
-    )
-    .enable_vsync(vsync);
-    let builder = if cfg!(target_os = "linux") {
-        builder.wgpu_backend(
-            pixels::wgpu::Backends::from_env().unwrap_or(pixels::wgpu::Backends::VULKAN),
-        )
-    } else {
-        builder
-    };
-    let mut pixels = builder.build()?;
-    pixels.set_scaling_mode(ScalingMode::Fill);
-    Ok(pixels)
-}
-
-pub(super) fn texture_width(scale: usize) -> usize {
-    FB_WIDTH * scale
-}
-
-pub(super) fn texture_height(scale: usize) -> usize {
-    window_present_height() * scale
-}
-
-pub(super) fn scale_rect(rect: Rect, scale: usize) -> Rect {
-    Rect {
-        x: rect.x * scale,
-        y: rect.y * scale,
-        w: rect.w * scale,
-        h: rect.h * scale,
-    }
-}
-
-fn cursor_texture_position(
-    pixels: &Pixels<'_>,
-    position: winit::dpi::PhysicalPosition<f64>,
-    texture_scale: usize,
-) -> Option<(i32, i32)> {
-    let (x, y) = pixels
-        .window_pos_to_pixel((position.x as f32, position.y as f32))
-        .ok()?;
-    Some(((x / texture_scale) as i32, (y / texture_scale) as i32))
-}
-
-fn cursor_in_status_bar(pos: (i32, i32)) -> bool {
-    pos.1 >= present_height() as i32 && pos.1 < window_present_height() as i32
-}
-
-fn cursor_in_display(pos: (i32, i32)) -> bool {
-    pos.0 >= 0 && pos.0 < FB_WIDTH as i32 && pos.1 >= 0 && pos.1 < present_height() as i32
-}
-
-fn volume_percent_from_pos(pos: (i32, i32)) -> u8 {
-    let track = volume_slider_track_rect();
-    let x = pos.0.clamp(track.x as i32, (track.x + track.w - 1) as i32) as usize;
-    let range = track.w.saturating_sub(1).max(1);
-    (((x - track.x) * 100 + range / 2) / range) as u8
-}
-
-fn volume_scroll_steps(delta: MouseScrollDelta) -> Option<i16> {
-    let amount = match delta {
-        MouseScrollDelta::LineDelta(_, y) => f64::from(y),
-        MouseScrollDelta::PixelDelta(pos) => pos.y,
-    };
-    if amount > 0.0 {
-        Some(1)
-    } else if amount < 0.0 {
-        Some(-1)
-    } else {
-        None
-    }
-}
-
-fn owner_name_from_code(code: u8) -> &'static str {
-    match code {
-        b'R' => "refresh",
-        b'B' => "bitplane",
-        b'S' => "sprite",
-        b'D' => "disk",
-        b'A' => "audio",
-        b'C' => "copper",
-        b'L' => "blitter",
-        b'P' => "cpu",
-        _ => "idle",
-    }
-}
-
-fn copy_present_frame(src_fb: &[u32], src_rows: usize, frame: &mut [u8], texture_scale: usize) {
-    debug_assert!(src_fb.len() >= src_rows * FB_WIDTH);
-    debug_assert_eq!(
-        frame.len(),
-        texture_width(texture_scale) * texture_height(texture_scale) * 4
-    );
-    let dst_stride = texture_width(texture_scale) * 4;
-    let out_rows = present_height() * texture_scale;
-    // The 570 woven scanlines map onto the 537-row 4:3 presentation (times
-    // the HiDPI texture scale). Select whole source rows instead of blending
-    // adjacent Amiga scanlines; normal presentation should not synthesize
-    // intermediate colours from line-to-line dithering.
-    for y in 0..out_rows {
-        let src_y = screenshot::scaled_source_row(y, src_rows, out_rows);
-        let row = &src_fb[src_y * FB_WIDTH..(src_y + 1) * FB_WIDTH];
-
-        let dst_off = y * dst_stride;
-        match texture_scale {
-            1 => unsafe {
-                std::ptr::copy_nonoverlapping(
-                    row.as_ptr() as *const u8,
-                    frame.as_mut_ptr().add(dst_off),
-                    FB_WIDTH * 4,
-                );
-            },
-            2 => {
-                for (x, &pixel) in row.iter().enumerate() {
-                    let pair = pixel as u64 | ((pixel as u64) << 32);
-                    unsafe {
-                        (frame.as_mut_ptr().add(dst_off + x * 8) as *mut u64).write_unaligned(pair);
-                    }
-                }
-            }
-            _ => {
-                let dst = &mut frame[dst_off..dst_off + dst_stride];
-                for x in 0..FB_WIDTH * texture_scale {
-                    dst[x * 4..x * 4 + 4].copy_from_slice(&row[x / texture_scale].to_le_bytes());
-                }
-            }
-        }
-    }
-}
-
-fn copy_window_present_frame(
-    src_fb: &[u32],
-    src_rows: usize,
-    frame: &mut [u8],
-    texture_scale: usize,
-    overscan: Overscan,
-    standard_tv_aperture: bool,
-) {
-    if overscan == Overscan::Tv && standard_tv_aperture {
-        copy_tv_aperture_to_window(src_fb, src_rows, frame, texture_scale);
-    } else {
-        copy_present_frame(src_fb, src_rows, frame, texture_scale);
-    }
-}
-
-fn copy_tv_aperture_to_window(
-    src_fb: &[u32],
-    src_rows: usize,
-    frame: &mut [u8],
-    texture_scale: usize,
-) {
-    debug_assert!(src_fb.len() >= src_rows * FB_WIDTH);
-    debug_assert_eq!(
-        frame.len(),
-        texture_width(texture_scale) * texture_height(texture_scale) * 4
-    );
-    let dst_stride = texture_width(texture_scale) * 4;
-    let present_rows = present_height();
-    let out_rows = present_rows * texture_scale;
-    let black = rgba(0, 0, 0).to_le_bytes();
-    for y in 0..out_rows {
-        let Some(crop_y) = tv_aperture_source_row(y, present_rows, texture_scale) else {
-            let dst = &mut frame[y * dst_stride..(y + 1) * dst_stride];
-            for px in dst.chunks_exact_mut(4) {
-                px.copy_from_slice(&black);
-            }
-            continue;
-        };
-        let src_y = (TV_PAL_PRESENT_SOURCE_Y + crop_y).min(src_rows - 1);
-        let row = &src_fb[src_y * FB_WIDTH..(src_y + 1) * FB_WIDTH];
-        let dst_off = y * dst_stride;
-        match texture_scale {
-            1 => {
-                let dst = &mut frame[dst_off..dst_off + dst_stride];
-                for x in 0..FB_WIDTH {
-                    let crop_x = x
-                        .saturating_sub(TV_PAL_LIVE_PAD_X)
-                        .min(TV_PAL_PRESENT_WIDTH - 1);
-                    let src_x = (TV_PAL_PRESENT_SOURCE_X + crop_x).min(FB_WIDTH - 1);
-                    dst[x * 4..x * 4 + 4].copy_from_slice(&row[src_x].to_le_bytes());
-                }
-            }
-            2 => {
-                for x in 0..FB_WIDTH {
-                    let crop_x = x
-                        .saturating_sub(TV_PAL_LIVE_PAD_X)
-                        .min(TV_PAL_PRESENT_WIDTH - 1);
-                    let src_x = (TV_PAL_PRESENT_SOURCE_X + crop_x).min(FB_WIDTH - 1);
-                    let pixel = row[src_x];
-                    let pair = pixel as u64 | ((pixel as u64) << 32);
-                    unsafe {
-                        (frame.as_mut_ptr().add(dst_off + x * 8) as *mut u64).write_unaligned(pair);
-                    }
-                }
-            }
-            _ => {
-                let dst = &mut frame[dst_off..dst_off + dst_stride];
-                for x in 0..FB_WIDTH * texture_scale {
-                    let out_x = x / texture_scale;
-                    let crop_x = out_x
-                        .saturating_sub(TV_PAL_LIVE_PAD_X)
-                        .min(TV_PAL_PRESENT_WIDTH - 1);
-                    let src_x = (TV_PAL_PRESENT_SOURCE_X + crop_x).min(FB_WIDTH - 1);
-                    dst[x * 4..x * 4 + 4].copy_from_slice(&row[src_x].to_le_bytes());
-                }
-            }
-        }
-    }
-}
-
-/// Map an output texture row to a row of the 540-line TV aperture crop,
-/// or None for rows that fall on the black bezel of the square-pixel
-/// presentation. The square canvas (570 rows) is taller than the
-/// aperture, so the crop is centred between black bands -- the vertical
-/// counterpart of the TV_PAL_LIVE_PAD_X side bands -- and its rows map
-/// 1:1. The default 4:3 canvas (537 rows) is shorter than the aperture
-/// and keeps mapping all 540 rows onto the output, as before.
-fn tv_aperture_source_row(y: usize, present_rows: usize, texture_scale: usize) -> Option<usize> {
-    let out_rows = present_rows * texture_scale;
-    let pad_rows = present_rows.saturating_sub(TV_PAL_PRESENT_HEIGHT) / 2 * texture_scale;
-    let content_rows = out_rows - 2 * pad_rows;
-    if y < pad_rows || y >= pad_rows + content_rows {
-        return None;
-    }
-    Some(screenshot::scaled_source_row(
-        y - pad_rows,
-        TV_PAL_PRESENT_HEIGHT,
-        content_rows,
-    ))
-}
-
-/// Paint a TV-style test pattern into the presentation framebuffer,
-/// shown while the machine is powered off. SMPTE-style colour bars over
-/// a grayscale step wedge: instantly readable as "no signal", and handy
-/// for setting up video capture levels before the machine boots.
-fn paint_test_screen(fb: &mut [u32]) {
-    debug_assert!(fb.len() >= FB_PIXELS);
-    const BARS: [u32; 7] = [
-        rgba(192, 192, 192), // grey
-        rgba(192, 192, 0),   // yellow
-        rgba(0, 192, 192),   // cyan
-        rgba(0, 192, 0),     // green
-        rgba(192, 0, 192),   // magenta
-        rgba(192, 0, 0),     // red
-        rgba(0, 0, 192),     // blue
-    ];
-    const STEPS: usize = 8;
-    let bars_h = FB_HEIGHT * 4 / 5;
-    for y in 0..FB_HEIGHT {
-        let row = &mut fb[y * FB_WIDTH..(y + 1) * FB_WIDTH];
-        if y < bars_h {
-            for (x, px) in row.iter_mut().enumerate() {
-                *px = BARS[x * BARS.len() / FB_WIDTH];
-            }
-        } else {
-            for (x, px) in row.iter_mut().enumerate() {
-                let level = (x * STEPS / FB_WIDTH) as u32 * 255 / (STEPS as u32 - 1);
-                *px = rgba(level, level, level);
-            }
-        }
-    }
-    draw_test_screen_logo(fb, bars_h);
-}
-
-fn draw_test_screen_logo(fb: &mut [u32], bars_h: usize) {
-    let Some(image) = copperline_logo_image() else {
-        return;
-    };
-    let x = FB_WIDTH.saturating_sub(image.width) / 2;
-    let y = bars_h.saturating_sub(image.height) / 2;
-    alpha_blit_rgba(fb, FB_WIDTH, FB_HEIGHT, x, y, image);
-}
-
-fn alpha_blit_rgba(
-    dst: &mut [u32],
-    dst_w: usize,
-    dst_h: usize,
-    x0: usize,
-    y0: usize,
-    src: &EmbeddedRgbaImage,
-) {
-    for sy in 0..src.height {
-        let dy = y0 + sy;
-        if dy >= dst_h {
-            break;
-        }
-        for sx in 0..src.width {
-            let dx = x0 + sx;
-            if dx >= dst_w {
-                break;
-            }
-            let src_off = (sy * src.width + sx) * 4;
-            let sr = src.rgba[src_off] as u32;
-            let sg = src.rgba[src_off + 1] as u32;
-            let sb = src.rgba[src_off + 2] as u32;
-            let sa = src.rgba[src_off + 3] as u32;
-            if sa == 0 {
-                continue;
-            }
-            let dst_px = &mut dst[dy * dst_w + dx];
-            *dst_px = if sa == 0xFF {
-                rgba(sr, sg, sb)
-            } else {
-                blend_rgba_over_opaque(*dst_px, sr, sg, sb, sa)
-            };
-        }
-    }
-}
-
-fn blend_rgba_over_opaque(dst: u32, sr: u32, sg: u32, sb: u32, sa: u32) -> u32 {
-    let [dr, dg, db, _] = dst.to_le_bytes();
-    let inv = 0xFF - sa;
-    let r = (sr * sa + u32::from(dr) * inv + 127) / 0xFF;
-    let g = (sg * sa + u32::from(dg) * inv + 127) / 0xFF;
-    let b = (sb * sa + u32::from(db) * inv + 127) / 0xFF;
-    rgba(r, g, b)
-}
-
-#[allow(clippy::too_many_arguments)]
-fn draw_status_bar(frame: &mut [u8], view: &StatusBarView, texture_scale: usize) {
-    let status = view.status;
-    let layout = bar_layout(&view.media);
-    let hover = view.hover;
-    fill_rect(
-        frame,
-        scale_rect(status_bar_rect(), texture_scale),
-        STATUS_BG,
-        texture_scale,
-    );
-    draw_hline(
-        frame,
-        present_height() * texture_scale,
-        STATUS_TOP,
-        texture_scale,
-    );
-    draw_hline(
-        frame,
-        window_present_height() * texture_scale - 1,
-        STATUS_BOTTOM,
-        texture_scale,
-    );
-    let rows = led_rows(&status, view.powered_on);
-    for (row, spec) in rows.iter().enumerate() {
-        draw_text(
-            frame,
-            STATUS_LABEL_X * texture_scale,
-            (present_height() + led_row_label_y(row, rows.len())) * texture_scale,
-            spec.label,
-            STATUS_TEXT,
-            texture_scale,
-        );
-        draw_led(
-            frame,
-            scale_rect(led_row_rect(row, rows.len()), texture_scale),
-            spec.on,
-            spec.on_color,
-            spec.off_color,
-            spec.highlight_on,
-            spec.highlight_off,
-            texture_scale,
-        );
-    }
-    draw_fdd_track_counter(frame, status.fdd_track, texture_scale);
-    for idx in 0..4 {
-        let drive = view.media.drives[idx];
-        if let Some(rect) = layout.drive_load[idx] {
-            draw_disk_button(
-                frame,
-                scale_rect(rect, texture_scale),
-                idx,
-                hover == Some(BarControl::DriveLoad(idx)),
-                texture_scale,
-            );
-        }
-        if let Some(rect) = layout.drive_swap[idx] {
-            draw_swap_button(
-                frame,
-                scale_rect(rect, texture_scale),
-                drive.multi,
-                hover == Some(BarControl::DriveSwap(idx)),
-                texture_scale,
-            );
-        }
-        if let Some(rect) = layout.drive_eject[idx] {
-            draw_eject_button(
-                frame,
-                scale_rect(rect, texture_scale),
-                drive.inserted,
-                hover == Some(BarControl::DriveEject(idx)),
-                texture_scale,
-            );
-        }
-    }
-    if let Some(rect) = layout.cd_load {
-        draw_cd_button(
-            frame,
-            scale_rect(rect, texture_scale),
-            hover == Some(BarControl::CdLoad),
-            texture_scale,
-        );
-    }
-    if let Some(rect) = layout.cd_eject {
-        draw_eject_button(
-            frame,
-            scale_rect(rect, texture_scale),
-            view.media.cd == Some(true),
-            hover == Some(BarControl::CdEject),
-            texture_scale,
-        );
-    }
-    draw_joystick_button(
-        frame,
-        scale_rect(joystick_toggle_rect(), texture_scale),
-        view.joystick_input_mode,
-        hover == Some(BarControl::Joystick),
-        texture_scale,
-    );
-    draw_volume_control(frame, status.output_volume_percent, texture_scale);
-    draw_menu_button(
-        frame,
-        scale_rect(menu_button_rect(), texture_scale),
-        hover == Some(BarControl::Menu),
-        texture_scale,
-    );
-    draw_shot_button(
-        frame,
-        scale_rect(shot_button_rect(), texture_scale),
-        hover == Some(BarControl::Screenshot),
-        texture_scale,
-    );
-    draw_pause_button(
-        frame,
-        scale_rect(pause_button_rect(), texture_scale),
-        view.paused,
-        hover == Some(BarControl::Pause),
-        texture_scale,
-    );
-    draw_power_button(
-        frame,
-        scale_rect(power_button_rect(), texture_scale),
-        view.powered_on,
-        hover == Some(BarControl::Power),
-        texture_scale,
-    );
-    draw_reboot_button(
-        frame,
-        scale_rect(reboot_button_rect(), texture_scale),
-        hover == Some(BarControl::Reboot),
-        texture_scale,
-    );
-}
-
-/// Per-drive status feeding the media controls in the status bar.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-struct DriveBar {
-    /// Drive is wired up this session; unconnected drives get no controls.
-    connected: bool,
-    /// A disk is currently inserted (enables the eject button).
-    inserted: bool,
-    /// More than one image is queued for this drive (enables swap).
-    multi: bool,
-}
-
-/// Removable-media status for the bar: the floppy drives plus the CD
-/// drive (None on machines without one, Some(disc inserted) otherwise).
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct MediaBar {
-    drives: [DriveBar; 4],
-    cd: Option<bool>,
-}
-
-/// Everything draw_status_bar needs for one frame.
-struct StatusBarView {
-    status: FrontPanelStatus,
-    powered_on: bool,
-    paused: bool,
-    media: MediaBar,
-    /// Active host joystick source, shown by the status-bar toggle icon.
-    joystick_input_mode: JoystickInputMode,
-    hover: Option<BarControl>,
-}
-
-/// A clickable status-bar control, used for hit-testing and hover.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum BarControl {
-    Power,
-    Pause,
-    Reboot,
-    Screenshot,
-    Menu,
-    Joystick,
-    Volume,
-    DriveLoad(usize),
-    DriveSwap(usize),
-    DriveEject(usize),
-    CdLoad,
-    CdEject,
-}
-
-/// Computed positions of the variable (media) part of the status bar.
-/// The fixed controls (volume, screenshot, pause, power, reboot) keep
-/// their own rect functions.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct BarLayout {
-    drive_load: [Option<Rect>; 4],
-    drive_swap: [Option<Rect>; 4],
-    drive_eject: [Option<Rect>; 4],
-    cd_load: Option<Rect>,
-    cd_eject: Option<Rect>,
-}
-
-/// Lay out the media controls left to right after the track counter.
-/// One or two drives sit in a single full-height row; three or four
-/// stack two-up in shorter rows, so even the worst case (four drives
-/// plus CD) keeps the counter and ends clear of the volume control.
-fn bar_layout(media: &MediaBar) -> BarLayout {
-    let mut layout = BarLayout {
-        drive_load: [None; 4],
-        drive_swap: [None; 4],
-        drive_eject: [None; 4],
-        cd_load: None,
-        cd_eject: None,
-    };
-    // At most 4 drives, so track membership/position without allocating a
-    // Vec on every call (this runs on every mouse-move and frame redraw).
-    let connected_count = (0..4).filter(|&idx| media.drives[idx].connected).count();
-    let stacked = connected_count > 2;
-
-    let cluster = |x: usize, y: usize, h: usize| {
-        let button = |x: usize, w: usize| Rect { x, y, w, h };
-        (
-            button(x, MEDIA_LOAD_W),
-            button(x + MEDIA_LOAD_W + MEDIA_INNER_GAP, MEDIA_SMALL_W),
-            button(
-                x + MEDIA_LOAD_W + 2 * MEDIA_INNER_GAP + MEDIA_SMALL_W,
-                MEDIA_SMALL_W,
-            ),
-        )
-    };
-
-    let mut drives_end_x = MEDIA_CLUSTER_X;
-    let mut pos = 0usize;
-    for idx in 0..4 {
-        if !media.drives[idx].connected {
-            continue;
-        }
-        let (x, y, h) = if stacked {
-            // Row-major two-column grid: DF0 DF1 over DF2 DF3.
-            let col = pos % 2;
-            let row = pos / 2;
-            (
-                MEDIA_CLUSTER_X + col * (MEDIA_CLUSTER_W + MEDIA_CLUSTER_GAP),
-                present_height() + MEDIA_STACKED_ROW0_Y + row * MEDIA_STACKED_PITCH,
-                MEDIA_STACKED_H,
-            )
-        } else {
-            (
-                MEDIA_CLUSTER_X + pos * (MEDIA_CLUSTER_W + MEDIA_CLUSTER_GAP),
-                present_height() + STATUS_CONTROL_Y,
-                STATUS_CONTROL_H,
-            )
-        };
-        let (load, swap, eject) = cluster(x, y, h);
-        layout.drive_load[idx] = Some(load);
-        layout.drive_swap[idx] = Some(swap);
-        layout.drive_eject[idx] = Some(eject);
-        drives_end_x = drives_end_x.max(x + MEDIA_CLUSTER_W);
-        pos += 1;
-    }
-
-    if media.cd.is_some() {
-        let x = if connected_count == 0 {
-            MEDIA_CLUSTER_X
-        } else {
-            drives_end_x + MEDIA_CD_GAP
-        };
-        // The CD cluster is load plus eject only; eject takes the slot a
-        // drive cluster gives to swap.
-        let (load, eject, _) = cluster(x, present_height() + STATUS_CONTROL_Y, STATUS_CONTROL_H);
-        layout.cd_load = Some(load);
-        layout.cd_eject = Some(eject);
-    }
-    layout
-}
-
-/// Map a cursor position to the status-bar control under it.
-fn control_at(pos: (i32, i32), layout: &BarLayout) -> Option<BarControl> {
-    for idx in 0..4 {
-        if layout.drive_load[idx].is_some_and(|r| r.contains(pos)) {
-            return Some(BarControl::DriveLoad(idx));
-        }
-        if layout.drive_swap[idx].is_some_and(|r| r.contains(pos)) {
-            return Some(BarControl::DriveSwap(idx));
-        }
-        if layout.drive_eject[idx].is_some_and(|r| r.contains(pos)) {
-            return Some(BarControl::DriveEject(idx));
-        }
-    }
-    if layout.cd_load.is_some_and(|r| r.contains(pos)) {
-        return Some(BarControl::CdLoad);
-    }
-    if layout.cd_eject.is_some_and(|r| r.contains(pos)) {
-        return Some(BarControl::CdEject);
-    }
-    if shot_button_rect().contains(pos) {
-        return Some(BarControl::Screenshot);
-    }
-    if menu_button_rect().contains(pos) {
-        return Some(BarControl::Menu);
-    }
-    if pause_button_rect().contains(pos) {
-        return Some(BarControl::Pause);
-    }
-    if power_button_rect().contains(pos) {
-        return Some(BarControl::Power);
-    }
-    if reboot_button_rect().contains(pos) {
-        return Some(BarControl::Reboot);
-    }
-    if joystick_toggle_rect().contains(pos) {
-        return Some(BarControl::Joystick);
-    }
-    if volume_control_hit_rect().contains(pos) {
-        return Some(BarControl::Volume);
-    }
-    None
-}
-
-fn status_bar_rect() -> Rect {
-    Rect {
-        x: 0,
-        y: present_height(),
-        w: FB_WIDTH,
-        h: STATUS_BAR_HEIGHT,
-    }
-}
-
-/// One LED row of the front-panel block (label plus LED palette).
-struct LedRowSpec {
-    label: &'static str,
-    on: bool,
-    on_color: u32,
-    off_color: u32,
-    highlight_on: u32,
-    highlight_off: u32,
-}
-
-/// The LED rows present this session: PWR and FDD always, HDD on IDE
-/// machines, CD on CDTV/CD32.
-fn led_rows(status: &FrontPanelStatus, powered_on: bool) -> Vec<LedRowSpec> {
-    let mut rows = vec![
-        LedRowSpec {
-            label: "PWR",
-            on: powered_on && status.power_led_on,
-            on_color: POWER_LED_ON,
-            off_color: POWER_LED_OFF,
-            highlight_on: rgba(255, 91, 82),
-            highlight_off: rgba(90, 27, 24),
-        },
-        LedRowSpec {
-            label: "FDD",
-            on: status.fdd_led_on,
-            on_color: FDD_LED_ON,
-            off_color: FDD_LED_OFF,
-            highlight_on: rgba(255, 190, 70),
-            highlight_off: rgba(100, 58, 18),
-        },
-    ];
-    if let Some(on) = status.hdd_led {
-        rows.push(LedRowSpec {
-            label: "HDD",
-            on,
-            on_color: HDD_LED_ON,
-            off_color: HDD_LED_OFF,
-            highlight_on: rgba(120, 255, 150),
-            highlight_off: rgba(26, 88, 40),
-        });
-    }
-    if let Some(on) = status.cd_led {
-        rows.push(LedRowSpec {
-            label: "CD",
-            on,
-            on_color: CD_LED_ON,
-            off_color: CD_LED_OFF,
-            highlight_on: rgba(140, 214, 255),
-            highlight_off: rgba(32, 74, 104),
-        });
-    }
-    rows
-}
-
-/// Label y (bar-local) for LED row `row` of `count`. Up to three rows
-/// use the classic spacing; four rows pack tighter to stay inside the
-/// bar.
-fn led_row_label_y(row: usize, count: usize) -> usize {
-    if count <= 3 {
-        LED_ROW_START_Y + row * LED_ROW_PITCH
-    } else {
-        LED_ROW_START_Y_TIGHT + row * LED_ROW_PITCH_TIGHT
-    }
-}
-
-fn led_row_rect(row: usize, count: usize) -> Rect {
-    Rect {
-        x: STATUS_LED_X,
-        y: present_height() + led_row_label_y(row, count) + STATUS_LED_Y_OFFSET,
-        w: STATUS_LED_W,
-        h: STATUS_LED_H,
-    }
-}
-
-fn fdd_track_counter_rect() -> Rect {
-    Rect {
-        x: 132,
-        y: present_height() + STATUS_CONTROL_Y,
-        w: 58,
-        h: STATUS_CONTROL_H,
-    }
-}
-
-fn fdd_track_digit_rect(index: usize) -> Rect {
-    let display = fdd_track_counter_rect();
-    Rect {
-        x: display.x + 5 + index * 17,
-        y: display.y + 3,
-        w: 12,
-        h: 16,
-    }
-}
-
-fn shot_button_rect() -> Rect {
-    Rect {
-        x: SHOT_BUTTON_X,
-        y: present_height() + STATUS_CONTROL_Y,
-        w: SHOT_BUTTON_W,
-        h: STATUS_CONTROL_H,
-    }
-}
-
-fn menu_button_rect() -> Rect {
-    Rect {
-        x: ui::MENU_BUTTON_X,
-        y: present_height() + STATUS_CONTROL_Y,
-        w: ui::MENU_BUTTON_W,
-        h: STATUS_CONTROL_H,
-    }
-}
-
-fn volume_control_hit_rect() -> Rect {
-    Rect {
-        x: VOLUME_SLIDER_X - 8,
-        y: present_height() + STATUS_CONTROL_Y,
-        w: VOLUME_SLIDER_W + 16,
-        h: STATUS_CONTROL_H,
-    }
-}
-
-fn joystick_toggle_rect() -> Rect {
-    Rect {
-        x: JOY_TOGGLE_X,
-        y: present_height() + STATUS_CONTROL_Y,
-        w: JOY_TOGGLE_W,
-        h: STATUS_CONTROL_H,
-    }
-}
-
-fn volume_slider_track_rect() -> Rect {
-    Rect {
-        x: VOLUME_SLIDER_X,
-        y: present_height() + VOLUME_SLIDER_Y,
-        w: VOLUME_SLIDER_W,
-        h: VOLUME_SLIDER_H,
-    }
-}
-
-fn volume_slider_knob_rect(percent: u8) -> Rect {
-    let track = volume_slider_track_rect();
-    let range = track.w.saturating_sub(1).max(1);
-    let center = track.x + range * usize::from(percent.min(100)) / 100;
-    Rect {
-        x: center.saturating_sub(VOLUME_KNOB_W / 2),
-        y: present_height() + STATUS_CONTROL_Y + (STATUS_CONTROL_H - VOLUME_KNOB_H) / 2,
-        w: VOLUME_KNOB_W,
-        h: VOLUME_KNOB_H,
-    }
-}
-
-fn reboot_button_rect() -> Rect {
-    Rect {
-        x: FB_WIDTH - 58,
-        y: present_height() + STATUS_CONTROL_Y,
-        w: 42,
-        h: STATUS_CONTROL_H,
-    }
-}
-
-fn power_button_rect() -> Rect {
-    Rect {
-        x: FB_WIDTH - 108,
-        y: present_height() + STATUS_CONTROL_Y,
-        w: 42,
-        h: STATUS_CONTROL_H,
-    }
-}
-
-fn pause_button_rect() -> Rect {
-    Rect {
-        x: FB_WIDTH - 158,
-        y: present_height() + STATUS_CONTROL_Y,
-        w: 42,
-        h: STATUS_CONTROL_H,
-    }
-}
-
-fn bar_hover_changed(
-    layout: &BarLayout,
-    previous: Option<(i32, i32)>,
-    current: Option<(i32, i32)>,
-) -> bool {
-    previous.and_then(|pos| control_at(pos, layout))
-        != current.and_then(|pos| control_at(pos, layout))
-}
-
-fn draw_fdd_track_counter(frame: &mut [u8], track: Option<u8>, texture_scale: usize) {
-    let rect = scale_rect(fdd_track_counter_rect(), texture_scale);
-    fill_rect(frame, rect, LED_BEZEL_DARK, texture_scale);
-    draw_rect_bevel(frame, rect, LED_BEZEL_LIGHT, STATUS_BOTTOM, texture_scale);
-    let inset = 2 * texture_scale;
-    fill_rect(
-        frame,
-        Rect {
-            x: rect.x + inset,
-            y: rect.y + inset,
-            w: rect.w.saturating_sub(inset * 2),
-            h: rect.h.saturating_sub(inset * 2),
-        },
-        TRACK_DISPLAY_BG,
-        texture_scale,
-    );
-
-    let digits = track.map_or([b'-', b'-', b'-'], |track| {
-        [
-            b'0' + track / 100,
-            b'0' + (track / 10) % 10,
-            b'0' + track % 10,
-        ]
-    });
-    for (idx, ch) in digits.into_iter().enumerate() {
-        draw_seven_segment_digit(
-            frame,
-            scale_rect(fdd_track_digit_rect(idx), texture_scale),
-            ch as char,
-            texture_scale,
-        );
-    }
-}
-
-fn draw_volume_control(frame: &mut [u8], percent: u8, texture_scale: usize) {
-    let percent = percent.min(100);
-    draw_speaker_glyph(frame, texture_scale);
-
-    let rect = scale_rect(volume_slider_track_rect(), texture_scale);
-    fill_rect(frame, rect, LED_BEZEL_DARK, texture_scale);
-    draw_rect_bevel(frame, rect, LED_BEZEL_LIGHT, STATUS_BOTTOM, texture_scale);
-
-    let inset = 2 * texture_scale;
-    let inner = Rect {
-        x: rect.x + inset,
-        y: rect.y + inset,
-        w: rect.w.saturating_sub(inset * 2),
-        h: rect.h.saturating_sub(inset * 2),
-    };
-    fill_rect(frame, inner, TRACK_DISPLAY_BG, texture_scale);
-
-    let fill_w = inner.w * usize::from(percent) / 100;
-    if fill_w != 0 {
-        let filled = Rect {
-            x: inner.x,
-            y: inner.y,
-            w: fill_w,
-            h: inner.h,
-        };
-        fill_rect(frame, filled, VOLUME_FILL, texture_scale);
-        draw_hline_span(
-            frame,
-            filled.y,
-            filled.x,
-            filled.x + filled.w,
-            VOLUME_FILL_HIGHLIGHT,
-            texture_scale,
-        );
-    }
-
-    let knob = scale_rect(volume_slider_knob_rect(percent), texture_scale);
-    fill_rect(frame, knob, BUTTON_FACE, texture_scale);
-    draw_rect_bevel(
-        frame,
-        knob,
-        BUTTON_EDGE_LIGHT,
-        BUTTON_EDGE_DARK,
-        texture_scale,
-    );
-}
-
-fn draw_button_base(frame: &mut [u8], rect: Rect, hover: bool, texture_scale: usize) {
-    let face = if hover {
-        BUTTON_FACE_HOVER
-    } else {
-        BUTTON_FACE
-    };
-    fill_rect(frame, rect, face, texture_scale);
-    draw_rect_bevel(
-        frame,
-        rect,
-        BUTTON_EDGE_LIGHT,
-        BUTTON_EDGE_DARK,
-        texture_scale,
-    );
-}
-
-fn draw_disk_button(
-    frame: &mut [u8],
-    rect: Rect,
-    drive_idx: usize,
-    hover: bool,
-    texture_scale: usize,
-) {
-    draw_button_base(frame, rect, hover, texture_scale);
-    draw_disk_glyph(frame, rect, drive_idx, texture_scale);
-}
-
-/// Swap button: two opposed horizontal arrows (cycle to the next queued
-/// disk). Drawn dim when there is nothing to swap to.
-fn draw_swap_button(
-    frame: &mut [u8],
-    rect: Rect,
-    enabled: bool,
-    hover: bool,
-    texture_scale: usize,
-) {
-    draw_button_base(frame, rect, hover && enabled, texture_scale);
-    let color = if enabled {
-        BUTTON_GLYPH
-    } else {
-        BUTTON_GLYPH_DISABLED
-    };
-    let s = texture_scale;
-    // Glyph coordinates are designed for a full-height (22) button;
-    // recentre vertically for the shorter stacked buttons.
-    let dy = glyph_dy(rect, s);
-    let fx = rect.x as f32;
-    let fy = rect.y as f32 + dy as f32 * s as f32;
-    let fs = s as f32;
-    let uy = |v: i32| (rect.y as i32 + (v + dy) * s as i32) as usize;
-    // Top arrow pointing right.
-    fill_rect(
-        frame,
-        Rect {
-            x: rect.x + 2 * s,
-            y: uy(7),
-            w: 8 * s,
-            h: 2 * s,
-        },
-        color,
-        texture_scale,
-    );
-    fill_triangle(
-        frame,
-        [
-            (fx + 10.0 * fs, fy + 5.0 * fs),
-            (fx + 10.0 * fs, fy + 11.0 * fs),
-            (fx + 13.5 * fs, fy + 8.0 * fs),
-        ],
-        color,
-        texture_scale,
-    );
-    // Bottom arrow pointing left.
-    fill_rect(
-        frame,
-        Rect {
-            x: rect.x + 6 * s,
-            y: uy(13),
-            w: 8 * s,
-            h: 2 * s,
-        },
-        color,
-        texture_scale,
-    );
-    fill_triangle(
-        frame,
-        [
-            (fx + 6.0 * fs, fy + 11.0 * fs),
-            (fx + 6.0 * fs, fy + 17.0 * fs),
-            (fx + 2.5 * fs, fy + 14.0 * fs),
-        ],
-        color,
-        texture_scale,
-    );
-}
-
-/// Vertical recentring (in unscaled pixels) for glyph art designed for a
-/// full-height control drawn in a shorter (stacked) button.
-fn glyph_dy(rect: Rect, texture_scale: usize) -> i32 {
-    ((rect.h / texture_scale) as i32 - STATUS_CONTROL_H as i32) / 2
-}
-
-/// Eject button: up triangle over a bar. Drawn dim when no media is in.
-fn draw_eject_button(
-    frame: &mut [u8],
-    rect: Rect,
-    enabled: bool,
-    hover: bool,
-    texture_scale: usize,
-) {
-    draw_button_base(frame, rect, hover && enabled, texture_scale);
-    let color = if enabled {
-        BUTTON_GLYPH
-    } else {
-        BUTTON_GLYPH_DISABLED
-    };
-    let s = texture_scale;
-    let dy = glyph_dy(rect, s);
-    let fx = rect.x as f32;
-    let fy = rect.y as f32 + dy as f32 * s as f32;
-    let fs = s as f32;
-    fill_triangle(
-        frame,
-        [
-            (fx + 8.0 * fs, fy + 5.0 * fs),
-            (fx + 2.5 * fs, fy + 12.0 * fs),
-            (fx + 13.5 * fs, fy + 12.0 * fs),
-        ],
-        color,
-        texture_scale,
-    );
-    fill_rect(
-        frame,
-        Rect {
-            x: rect.x + 3 * s,
-            y: (rect.y as i32 + (14 + dy) * s as i32) as usize,
-            w: 10 * s,
-            h: 2 * s,
-        },
-        color,
-        texture_scale,
-    );
-}
-
-/// CD load/swap button: a compact disc.
-fn draw_cd_button(frame: &mut [u8], rect: Rect, hover: bool, texture_scale: usize) {
-    draw_button_base(frame, rect, hover, texture_scale);
-    let s = texture_scale;
-    // Disc centre and radii in unscaled button-local pixels.
-    let cx = (rect.x + 11 * s) as f32;
-    let cy = rect.y as f32 + rect.h as f32 / 2.0;
-    let fs = s as f32;
-    for py in rect.y..rect.y + rect.h {
-        for px in rect.x..rect.x + rect.w {
-            let dx = (px as f32 + 0.5 - cx) / fs;
-            let dy = (py as f32 + 0.5 - cy) / fs;
-            let r2 = dx * dx + dy * dy;
-            let color = if r2 <= 2.2 {
-                CD_HOLE
-            } else if r2 <= 6.2 {
-                CD_HUB
-            } else if r2 <= 64.0 {
-                // A sheen wedge across the upper-left of the data area.
-                if r2 >= 30.0 && dx + dy < -3.0 {
-                    CD_SHEEN
-                } else {
-                    CD_BODY
-                }
-            } else {
-                continue;
-            };
-            put_pixel(frame, px, py, color, texture_scale);
-        }
-    }
-}
-
-/// Menu button: three stacked bars (opens the pop-up menu).
-fn draw_menu_button(frame: &mut [u8], rect: Rect, hover: bool, texture_scale: usize) {
-    draw_button_base(frame, rect, hover, texture_scale);
-    let s = texture_scale;
-    for row in 0..3 {
-        fill_rect(
-            frame,
-            Rect {
-                x: rect.x + 4 * s,
-                y: rect.y + (6 + row * 4) * s,
-                w: 14 * s,
-                h: 2 * s,
-            },
-            BUTTON_GLYPH,
-            texture_scale,
-        );
-    }
-}
-
-/// Screenshot button: a small camera.
-fn draw_shot_button(frame: &mut [u8], rect: Rect, hover: bool, texture_scale: usize) {
-    draw_button_base(frame, rect, hover, texture_scale);
-    let s = texture_scale;
-    // Viewfinder bump, then the body, then the lens.
-    fill_rect(
-        frame,
-        Rect {
-            x: rect.x + 8 * s,
-            y: rect.y + 5 * s,
-            w: 6 * s,
-            h: 3 * s,
-        },
-        CAMERA_BODY,
-        texture_scale,
-    );
-    fill_rect(
-        frame,
-        Rect {
-            x: rect.x + 3 * s,
-            y: rect.y + 7 * s,
-            w: 16 * s,
-            h: 10 * s,
-        },
-        CAMERA_BODY,
-        texture_scale,
-    );
-    let cx = (rect.x + 11 * s) as f32;
-    let cy = (rect.y + 12 * s) as f32;
-    let fs = s as f32;
-    for py in rect.y + 7 * s..rect.y + 17 * s {
-        for px in rect.x + 5 * s..rect.x + 17 * s {
-            let dx = (px as f32 + 0.5 - cx) / fs;
-            let dy = (py as f32 + 0.5 - cy) / fs;
-            let r2 = dx * dx + dy * dy;
-            let color = if r2 <= 5.5 {
-                CAMERA_LENS
-            } else if r2 <= 12.5 {
-                BUTTON_GLYPH
-            } else {
-                continue;
-            };
-            put_pixel(frame, px, py, color, texture_scale);
-        }
-    }
-}
-
-/// Speaker glyph labelling the volume slider: a driver box, a cone, and
-/// two sound arcs.
-fn draw_speaker_glyph(frame: &mut [u8], texture_scale: usize) {
-    let s = texture_scale;
-    let x = VOLUME_GLYPH_X * s;
-    let y = (present_height() + STATUS_CONTROL_Y) * s;
-    let fs = s as f32;
-    fill_rect(
-        frame,
-        Rect {
-            x: x + s,
-            y: y + 9 * s,
-            w: 3 * s,
-            h: 5 * s,
-        },
-        STATUS_TEXT,
-        texture_scale,
-    );
-    fill_triangle(
-        frame,
-        [
-            (x as f32 + 4.0 * fs, y as f32 + 11.5 * fs),
-            (x as f32 + 8.0 * fs, y as f32 + 5.5 * fs),
-            (x as f32 + 8.0 * fs, y as f32 + 17.5 * fs),
-        ],
-        STATUS_TEXT,
-        texture_scale,
-    );
-    draw_vline_span(
-        frame,
-        x + 10 * s,
-        y + 9 * s,
-        y + 14 * s,
-        STATUS_TEXT,
-        texture_scale,
-    );
-    draw_vline_span(
-        frame,
-        x + 12 * s,
-        y + 6 * s,
-        y + 17 * s,
-        STATUS_TEXT,
-        texture_scale,
-    );
-}
-
-fn draw_disk_glyph(frame: &mut [u8], rect: Rect, drive_idx: usize, texture_scale: usize) {
-    let s = texture_scale;
-    // Centre the 16px disk body vertically (full-height buttons give the
-    // original 3px margin; stacked buttons less).
-    let body_margin_y = (rect.h / s).saturating_sub(16) / 2;
-    let body = Rect {
-        x: rect.x + 3 * s,
-        y: rect.y + body_margin_y * s,
-        w: 16 * s,
-        h: 16 * s,
-    };
-    fill_rect(frame, body, DISK_BODY_SHADOW, texture_scale);
-    fill_rect(
-        frame,
-        Rect {
-            x: body.x + s,
-            y: body.y + s,
-            w: body.w.saturating_sub(2 * s),
-            h: body.h.saturating_sub(2 * s),
-        },
-        DISK_BODY,
-        texture_scale,
-    );
-    fill_rect(
-        frame,
-        Rect {
-            x: body.x + s,
-            y: body.y + s,
-            w: body.w.saturating_sub(2 * s),
-            h: s,
-        },
-        DISK_BODY_HIGHLIGHT,
-        texture_scale,
-    );
-    fill_rect(
-        frame,
-        Rect {
-            x: body.x + s,
-            y: body.y + s,
-            w: s,
-            h: body.h.saturating_sub(2 * s),
-        },
-        DISK_BODY_HIGHLIGHT,
-        texture_scale,
-    );
-    fill_rect(
-        frame,
-        Rect {
-            x: body.x + 5 * s,
-            y: body.y + 2 * s,
-            w: 8 * s,
-            h: 5 * s,
-        },
-        DISK_SHUTTER,
-        texture_scale,
-    );
-    fill_rect(
-        frame,
-        Rect {
-            x: body.x + 6 * s,
-            y: body.y + 3 * s,
-            w: 5 * s,
-            h: s,
-        },
-        DISK_LABEL,
-        texture_scale,
-    );
-    fill_rect(
-        frame,
-        Rect {
-            x: body.x + 5 * s,
-            y: body.y + 6 * s,
-            w: 8 * s,
-            h: s,
-        },
-        DISK_SHUTTER_DARK,
-        texture_scale,
-    );
-    fill_rect(
-        frame,
-        Rect {
-            x: body.x + 3 * s,
-            y: body.y + 9 * s,
-            w: 10 * s,
-            h: 6 * s,
-        },
-        DISK_LABEL,
-        texture_scale,
-    );
-    fill_rect(
-        frame,
-        Rect {
-            x: body.x + 4 * s,
-            y: body.y + 11 * s,
-            w: 5 * s,
-            h: s,
-        },
-        DISK_LABEL_LINE,
-        texture_scale,
-    );
-    fill_rect(
-        frame,
-        Rect {
-            x: body.x + 4 * s,
-            y: body.y + 13 * s,
-            w: 4 * s,
-            h: s,
-        },
-        DISK_LABEL_LINE,
-        texture_scale,
-    );
-    // The drive number, written on the right of the disk label.
-    draw_tiny_digit(
-        frame,
-        body.x + 9 * s,
-        body.y + 9 * s,
-        drive_idx as u8,
-        DISK_BODY_SHADOW,
-        texture_scale,
-    );
-}
-
-/// 3x5 pixel digits 0-3 for the drive number on the disk-button label.
-fn draw_tiny_digit(
-    frame: &mut [u8],
-    x: usize,
-    y: usize,
-    digit: u8,
-    color: u32,
-    texture_scale: usize,
-) {
-    const GLYPHS: [[u8; 5]; 4] = [
-        [0b111, 0b101, 0b101, 0b101, 0b111],
-        [0b010, 0b110, 0b010, 0b010, 0b111],
-        [0b111, 0b001, 0b111, 0b100, 0b111],
-        [0b111, 0b001, 0b011, 0b001, 0b111],
-    ];
-    let Some(rows) = GLYPHS.get(usize::from(digit)) else {
-        return;
-    };
-    let s = texture_scale;
-    for (row, bits) in rows.iter().enumerate() {
-        for col in 0..3 {
-            if bits & (0b100 >> col) != 0 {
-                fill_rect(
-                    frame,
-                    Rect {
-                        x: x + col * s,
-                        y: y + row * s,
-                        w: s,
-                        h: s,
-                    },
-                    color,
-                    texture_scale,
-                );
-            }
-        }
-    }
-}
-
-fn draw_seven_segment_digit(frame: &mut [u8], rect: Rect, ch: char, texture_scale: usize) {
-    const SEG_A: u8 = 1 << 0;
-    const SEG_B: u8 = 1 << 1;
-    const SEG_C: u8 = 1 << 2;
-    const SEG_D: u8 = 1 << 3;
-    const SEG_E: u8 = 1 << 4;
-    const SEG_F: u8 = 1 << 5;
-    const SEG_G: u8 = 1 << 6;
-
-    let mask = match ch {
-        '0' => SEG_A | SEG_B | SEG_C | SEG_D | SEG_E | SEG_F,
-        '1' => SEG_B | SEG_C,
-        '2' => SEG_A | SEG_B | SEG_D | SEG_E | SEG_G,
-        '3' => SEG_A | SEG_B | SEG_C | SEG_D | SEG_G,
-        '4' => SEG_B | SEG_C | SEG_F | SEG_G,
-        '5' => SEG_A | SEG_C | SEG_D | SEG_F | SEG_G,
-        '6' => SEG_A | SEG_C | SEG_D | SEG_E | SEG_F | SEG_G,
-        '7' => SEG_A | SEG_B | SEG_C,
-        '8' => SEG_A | SEG_B | SEG_C | SEG_D | SEG_E | SEG_F | SEG_G,
-        '9' => SEG_A | SEG_B | SEG_C | SEG_D | SEG_F | SEG_G,
-        '-' => SEG_G,
-        _ => 0,
-    };
-    let thickness = 2 * texture_scale;
-    let short = 5 * texture_scale;
-    let horizontal = 8 * texture_scale;
-
-    let segments = [
-        (
-            SEG_A,
-            Rect {
-                x: rect.x + thickness,
-                y: rect.y,
-                w: horizontal,
-                h: thickness,
-            },
-        ),
-        (
-            SEG_B,
-            Rect {
-                x: rect.x + rect.w - thickness,
-                y: rect.y + thickness,
-                w: thickness,
-                h: short,
-            },
-        ),
-        (
-            SEG_C,
-            Rect {
-                x: rect.x + rect.w - thickness,
-                y: rect.y + rect.h - thickness - short,
-                w: thickness,
-                h: short,
-            },
-        ),
-        (
-            SEG_D,
-            Rect {
-                x: rect.x + thickness,
-                y: rect.y + rect.h - thickness,
-                w: horizontal,
-                h: thickness,
-            },
-        ),
-        (
-            SEG_E,
-            Rect {
-                x: rect.x,
-                y: rect.y + rect.h - thickness - short,
-                w: thickness,
-                h: short,
-            },
-        ),
-        (
-            SEG_F,
-            Rect {
-                x: rect.x,
-                y: rect.y + thickness,
-                w: thickness,
-                h: short,
-            },
-        ),
-        (
-            SEG_G,
-            Rect {
-                x: rect.x + thickness,
-                y: rect.y + rect.h / 2 - thickness / 2,
-                w: horizontal,
-                h: thickness,
-            },
-        ),
-    ];
-
-    for (segment, segment_rect) in segments {
-        let lit = mask & segment != 0;
-        fill_rect(
-            frame,
-            segment_rect,
-            if lit {
-                TRACK_SEGMENT_ON
-            } else {
-                TRACK_SEGMENT_OFF
-            },
-            texture_scale,
-        );
-        if lit {
-            draw_hline_span(
-                frame,
-                segment_rect.y,
-                segment_rect.x,
-                segment_rect.x + segment_rect.w,
-                TRACK_SEGMENT_HIGHLIGHT,
-                texture_scale,
-            );
-        }
-    }
-}
-
-fn draw_led(
-    frame: &mut [u8],
-    rect: Rect,
-    on: bool,
-    on_color: u32,
-    off_color: u32,
-    on_highlight: u32,
-    off_highlight: u32,
-    texture_scale: usize,
-) {
-    fill_rect(frame, rect, LED_BEZEL_DARK, texture_scale);
-    draw_rect_bevel(frame, rect, LED_BEZEL_LIGHT, STATUS_BOTTOM, texture_scale);
-    let inset = 2 * texture_scale;
-    let inner = Rect {
-        x: rect.x + inset,
-        y: rect.y + inset,
-        w: rect.w.saturating_sub(inset * 2),
-        h: rect.h.saturating_sub(inset * 2),
-    };
-    fill_rect(
-        frame,
-        inner,
-        if on { on_color } else { off_color },
-        texture_scale,
-    );
-    for dy in 0..texture_scale {
-        draw_hline_span(
-            frame,
-            inner.y + dy,
-            inner.x,
-            inner.x + inner.w,
-            if on { on_highlight } else { off_highlight },
-            texture_scale,
-        );
-    }
-}
-
-fn draw_reboot_button(frame: &mut [u8], rect: Rect, hover: bool, texture_scale: usize) {
-    fill_rect(
-        frame,
-        rect,
-        if hover {
-            BUTTON_FACE_HOVER
-        } else {
-            BUTTON_FACE
-        },
-        texture_scale,
-    );
-    draw_rect_bevel(
-        frame,
-        rect,
-        BUTTON_EDGE_LIGHT,
-        BUTTON_EDGE_DARK,
-        texture_scale,
-    );
-    let cx = rect.x + rect.w / 2;
-    let cy = rect.y + rect.h / 2;
-    draw_reset_glyph(frame, cx, cy, texture_scale);
-}
-
-fn draw_power_button(
-    frame: &mut [u8],
-    rect: Rect,
-    powered_on: bool,
-    hover: bool,
-    texture_scale: usize,
-) {
-    fill_rect(
-        frame,
-        rect,
-        if hover {
-            BUTTON_FACE_HOVER
-        } else {
-            BUTTON_FACE
-        },
-        texture_scale,
-    );
-    draw_rect_bevel(
-        frame,
-        rect,
-        BUTTON_EDGE_LIGHT,
-        BUTTON_EDGE_DARK,
-        texture_scale,
-    );
-    let cx = rect.x + rect.w / 2;
-    let cy = rect.y + rect.h / 2;
-    let color = if powered_on {
-        POWER_GLYPH_ON
-    } else {
-        POWER_GLYPH_OFF
-    };
-    draw_power_glyph(frame, cx, cy, color, texture_scale);
-}
-
-fn draw_pause_button(
-    frame: &mut [u8],
-    rect: Rect,
-    paused: bool,
-    hover: bool,
-    texture_scale: usize,
-) {
-    fill_rect(
-        frame,
-        rect,
-        if hover {
-            BUTTON_FACE_HOVER
-        } else {
-            BUTTON_FACE
-        },
-        texture_scale,
-    );
-    draw_rect_bevel(
-        frame,
-        rect,
-        BUTTON_EDGE_LIGHT,
-        BUTTON_EDGE_DARK,
-        texture_scale,
-    );
-    let cx = rect.x + rect.w / 2;
-    let cy = rect.y + rect.h / 2;
-    // Show the action the button performs: a play triangle while paused
-    // (click to resume), the twin pause bars while running.
-    if paused {
-        draw_play_glyph(frame, cx, cy, BUTTON_GLYPH, texture_scale);
-    } else {
-        draw_pause_glyph(frame, cx, cy, BUTTON_GLYPH, texture_scale);
-    }
-}
-
-/// Joystick input-source toggle: shows the device currently driving the
-/// emulated port-2 joystick (a gamepad in `Gamepad` mode, a keyboard in
-/// `Keyboard` mode). Clicking it flips between the two, so the active source is
-/// always visible rather than hidden behind a key combination.
-fn draw_joystick_button(
-    frame: &mut [u8],
-    rect: Rect,
-    mode: JoystickInputMode,
-    hover: bool,
-    texture_scale: usize,
-) {
-    draw_button_base(frame, rect, hover, texture_scale);
-    match mode {
-        JoystickInputMode::Gamepad => draw_gamepad_glyph(frame, rect, texture_scale),
-        JoystickInputMode::Keyboard => draw_keyboard_glyph(frame, rect, texture_scale),
-    }
-}
-
-/// A small gamepad: a rounded green body with a recessed d-pad on the left and
-/// two action buttons on the right.
-fn draw_gamepad_glyph(frame: &mut [u8], rect: Rect, texture_scale: usize) {
-    let s = texture_scale;
-    let mut cell = |x: usize, y: usize, w: usize, h: usize, color: u32| {
-        fill_rect(
-            frame,
-            Rect {
-                x: rect.x + x * s,
-                y: rect.y + y * s,
-                w: w * s,
-                h: h * s,
-            },
-            color,
-            texture_scale,
-        );
-    };
-    // Body and the two grip bumps.
-    cell(4, 8, 14, 8, BUTTON_GLYPH);
-    cell(3, 13, 3, 3, BUTTON_GLYPH);
-    cell(16, 13, 3, 3, BUTTON_GLYPH);
-    // D-pad cross, cut into the body on the left.
-    cell(7, 9, 2, 5, BUTTON_EDGE_DARK);
-    cell(5, 11, 6, 2, BUTTON_EDGE_DARK);
-    // Two action buttons on the right.
-    cell(13, 10, 2, 2, BUTTON_EDGE_DARK);
-    cell(15, 12, 2, 2, BUTTON_EDGE_DARK);
-}
-
-/// A small keyboard: a recessed dark case holding two rows of green keys and a
-/// space bar.
-fn draw_keyboard_glyph(frame: &mut [u8], rect: Rect, texture_scale: usize) {
-    let s = texture_scale;
-    let mut cell = |x: usize, y: usize, w: usize, h: usize, color: u32| {
-        fill_rect(
-            frame,
-            Rect {
-                x: rect.x + x * s,
-                y: rect.y + y * s,
-                w: w * s,
-                h: h * s,
-            },
-            color,
-            texture_scale,
-        );
-    };
-    // Case.
-    cell(3, 6, 16, 11, BUTTON_EDGE_DARK);
-    // Two rows of keys.
-    for &kx in &[5, 8, 11, 14] {
-        cell(kx, 8, 2, 2, BUTTON_GLYPH);
-        cell(kx, 11, 2, 2, BUTTON_GLYPH);
-    }
-    // Space bar.
-    cell(7, 14, 8, 2, BUTTON_GLYPH);
-}
-
-/// The pause symbol: two short vertical bars flanking the centre.
-fn draw_pause_glyph(frame: &mut [u8], cx: usize, cy: usize, color: u32, texture_scale: usize) {
-    let bar_w = 2 * texture_scale;
-    let bar_h = 11 * texture_scale;
-    let gap = 3 * texture_scale;
-    let top = cy.saturating_sub(bar_h / 2);
-    let left = cx.saturating_sub(gap / 2 + bar_w);
-    let right = cx + gap / 2;
-    for x in [left, right] {
-        fill_rect(
-            frame,
-            Rect {
-                x,
-                y: top,
-                w: bar_w,
-                h: bar_h,
-            },
-            color,
-            texture_scale,
-        );
-    }
-}
-
-/// The play symbol: a right-pointing filled triangle.
-fn draw_play_glyph(frame: &mut [u8], cx: usize, cy: usize, color: u32, texture_scale: usize) {
-    let s = texture_scale as f32;
-    let half_h = 6.0 * s;
-    let width = 11.0 * s;
-    let left = cx as f32 - width / 2.0 + 1.0;
-    let cyf = cy as f32 + 0.5;
-    fill_triangle(
-        frame,
-        [
-            (left, cyf - half_h),
-            (left, cyf + half_h),
-            (left + width, cyf),
-        ],
-        color,
-        texture_scale,
-    );
-}
-
-/// The IEC power symbol: a near-closed ring broken at the top, with a
-/// vertical bar dropping through the gap toward the centre.
-fn draw_power_glyph(frame: &mut [u8], cx: usize, cy: usize, color: u32, texture_scale: usize) {
-    let scale = texture_scale as f32;
-    let ccx = cx as f32 + 0.5;
-    let ccy = cy as f32 + 0.5 + 0.5 * scale;
-    let radius = 5.5 * scale;
-    let stroke = 1.35 * scale;
-
-    // Ring, swept clockwise from just right of top all the way around to
-    // just left of top, leaving a gap centred on 12 o'clock.
-    let gap = 0.6_f32;
-    let top = -std::f32::consts::FRAC_PI_2;
-    let start = top + gap;
-    let end = top + std::f32::consts::TAU - gap;
-    let steps = 32;
-    let mut prev = (ccx + radius * start.cos(), ccy + radius * start.sin());
-    for step in 1..=steps {
-        let t = start + (end - start) * step as f32 / steps as f32;
-        let next = (ccx + radius * t.cos(), ccy + radius * t.sin());
-        draw_thick_line(
-            frame,
-            prev.0,
-            prev.1,
-            next.0,
-            next.1,
-            stroke,
-            color,
-            texture_scale,
-        );
-        prev = next;
-    }
-
-    // Vertical bar from above the ring down to its centre.
-    draw_thick_line(
-        frame,
-        ccx,
-        ccy - radius - 1.5 * scale,
-        ccx,
-        ccy - 0.5 * scale,
-        stroke,
-        color,
-        texture_scale,
-    );
-}
-
-fn draw_reset_glyph(frame: &mut [u8], cx: usize, cy: usize, texture_scale: usize) {
-    let svg_scale = 0.17 * texture_scale as f32;
-    let ox = cx as f32 - 8.5 * texture_scale as f32;
-    let oy = cy as f32 - 8.5 * texture_scale as f32;
-    let map = |x: f32, y: f32| (ox + x * svg_scale, oy + y * svg_scale);
-    let stroke = 1.35 * texture_scale as f32;
-
-    let (x0, y0) = map(50.0, 10.0);
-    let (x1, y1) = map(50.0, 45.0);
-    draw_thick_line(frame, x0, y0, x1, y1, stroke, BUTTON_GLYPH, texture_scale);
-
-    draw_cubic(
-        frame,
-        [(20.0, 29.0), (4.0, 52.0), (15.0, 90.0), (50.0, 90.0)],
-        svg_scale,
-        ox,
-        oy,
-        stroke,
-        BUTTON_GLYPH,
-        texture_scale,
-    );
-    draw_cubic(
-        frame,
-        [(50.0, 90.0), (85.0, 90.0), (100.0, 47.0), (74.0, 20.0)],
-        svg_scale,
-        ox,
-        oy,
-        stroke,
-        BUTTON_GLYPH,
-        texture_scale,
-    );
-
-    let arrow = [map(2.0, 21.0), map(31.0, 19.0), map(33.0, 48.0)];
-    fill_triangle(frame, arrow, BUTTON_GLYPH, texture_scale);
-}
-
-fn draw_cubic(
-    frame: &mut [u8],
-    p: [(f32, f32); 4],
-    scale: f32,
-    ox: f32,
-    oy: f32,
-    radius: f32,
-    color: u32,
-    texture_scale: usize,
-) {
-    let mut prev = (ox + p[0].0 * scale, oy + p[0].1 * scale);
-    for step in 1..=18 {
-        let t = step as f32 / 18.0;
-        let mt = 1.0 - t;
-        let x = mt * mt * mt * p[0].0
-            + 3.0 * mt * mt * t * p[1].0
-            + 3.0 * mt * t * t * p[2].0
-            + t * t * t * p[3].0;
-        let y = mt * mt * mt * p[0].1
-            + 3.0 * mt * mt * t * p[1].1
-            + 3.0 * mt * t * t * p[2].1
-            + t * t * t * p[3].1;
-        let next = (ox + x * scale, oy + y * scale);
-        draw_thick_line(
-            frame,
-            prev.0,
-            prev.1,
-            next.0,
-            next.1,
-            radius,
-            color,
-            texture_scale,
-        );
-        prev = next;
-    }
-}
-
-fn draw_thick_line(
-    frame: &mut [u8],
-    x0: f32,
-    y0: f32,
-    x1: f32,
-    y1: f32,
-    radius: f32,
-    color: u32,
-    texture_scale: usize,
-) {
-    let min_x = (x0.min(x1) - radius - 1.0).floor().max(0.0) as usize;
-    let max_x = (x0.max(x1) + radius + 1.0)
-        .ceil()
-        .min((texture_width(texture_scale) - 1) as f32) as usize;
-    let min_y = (y0.min(y1) - radius - 1.0).floor().max(0.0) as usize;
-    let max_y = (y0.max(y1) + radius + 1.0)
-        .ceil()
-        .min((texture_height(texture_scale) - 1) as f32) as usize;
-    let dx = x1 - x0;
-    let dy = y1 - y0;
-    let len2 = dx * dx + dy * dy;
-    for y in min_y..=max_y {
-        for x in min_x..=max_x {
-            let px = x as f32 + 0.5;
-            let py = y as f32 + 0.5;
-            let t = if len2 == 0.0 {
-                0.0
-            } else {
-                (((px - x0) * dx + (py - y0) * dy) / len2).clamp(0.0, 1.0)
-            };
-            let nearest_x = x0 + t * dx;
-            let nearest_y = y0 + t * dy;
-            let dist_x = px - nearest_x;
-            let dist_y = py - nearest_y;
-            let dist = (dist_x * dist_x + dist_y * dist_y).sqrt();
-            let coverage = (radius + 0.5 - dist).clamp(0.0, 1.0);
-            if coverage > 0.0 {
-                blend_pixel(frame, x, y, color, coverage, texture_scale);
-            }
-        }
-    }
-}
-
-fn fill_triangle(frame: &mut [u8], p: [(f32, f32); 3], color: u32, texture_scale: usize) {
-    let min_x = p
-        .iter()
-        .map(|(x, _)| *x)
-        .fold(f32::INFINITY, f32::min)
-        .floor()
-        .max(0.0) as usize;
-    let max_x = p
-        .iter()
-        .map(|(x, _)| *x)
-        .fold(f32::NEG_INFINITY, f32::max)
-        .ceil()
-        .min((texture_width(texture_scale) - 1) as f32) as usize;
-    let min_y = p
-        .iter()
-        .map(|(_, y)| *y)
-        .fold(f32::INFINITY, f32::min)
-        .floor()
-        .max(0.0) as usize;
-    let max_y = p
-        .iter()
-        .map(|(_, y)| *y)
-        .fold(f32::NEG_INFINITY, f32::max)
-        .ceil()
-        .min((texture_height(texture_scale) - 1) as f32) as usize;
-    let area = edge(p[0], p[1], p[2]);
-    if area == 0.0 {
-        return;
-    }
-    for y in min_y..=max_y {
-        for x in min_x..=max_x {
-            let mut hits = 0;
-            for sy in 0..3 {
-                for sx in 0..3 {
-                    let point = (
-                        x as f32 + (sx as f32 + 0.5) / 3.0,
-                        y as f32 + (sy as f32 + 0.5) / 3.0,
-                    );
-                    let w0 = edge(p[1], p[2], point);
-                    let w1 = edge(p[2], p[0], point);
-                    let w2 = edge(p[0], p[1], point);
-                    if (w0 >= 0.0 && w1 >= 0.0 && w2 >= 0.0)
-                        || (w0 <= 0.0 && w1 <= 0.0 && w2 <= 0.0)
-                    {
-                        hits += 1;
-                    }
-                }
-            }
-            if hits > 0 {
-                blend_pixel(frame, x, y, color, hits as f32 / 9.0, texture_scale);
-            }
-        }
-    }
-}
-
-fn edge(a: (f32, f32), b: (f32, f32), c: (f32, f32)) -> f32 {
-    (c.0 - a.0) * (b.1 - a.1) - (c.1 - a.1) * (b.0 - a.0)
-}
-
-/// Draw a transient overlay message near the bottom-left of the display
-/// region: a translucent panel with the text (plus a 1px drop shadow for
-/// legibility over arbitrary video). Operates on the presentation
-/// texture, so it is never captured in screenshots.
-/// Persistent "(*) REC" badge in the display's top-right corner while a
-/// video recording runs. Like the OSD it is drawn into the presentation
-/// texture after the frame is captured, so it is never recorded.
-fn draw_record_badge(frame: &mut [u8], texture_scale: usize) {
-    let s = texture_scale;
-    let px = 2 * s;
-    let pad = 4 * s;
-    let margin = 8 * s;
-    let dot_d = 8 * s;
-    let gap = 4 * s;
-
-    let text = "REC";
-    let text_w = font::text_width(text, px);
-    let text_h = font::text_height(px);
-    let box_w = dot_d + gap + text_w + 2 * pad;
-    let box_h = text_h + 2 * pad;
-    let box_x = (FB_WIDTH * s).saturating_sub(margin + box_w);
-    let box_y = margin;
-
-    fill_rect_blend(
-        frame,
-        Rect {
-            x: box_x,
-            y: box_y,
-            w: box_w,
-            h: box_h,
-        },
-        OSD_BG,
-        0.68,
-        s,
-    );
-    // Red record dot, centred on the text line.
-    let cx = (box_x + pad + dot_d / 2) as f32;
-    let cy = (box_y + box_h / 2) as f32;
-    let radius = dot_d as f32 / 2.0;
-    for y in box_y..box_y + box_h {
-        for x in box_x + pad..box_x + pad + dot_d {
-            let dx = x as f32 + 0.5 - cx;
-            let dy = y as f32 + 0.5 - cy;
-            if dx * dx + dy * dy <= radius * radius {
-                put_pixel(frame, x, y, RECORD_DOT, s);
-            }
-        }
-    }
-    let text_x = box_x + pad + dot_d + gap;
-    let text_y = box_y + pad;
-    font::draw_text(
-        frame,
-        texture_width(s),
-        texture_height(s),
-        text_x + s,
-        text_y + s,
-        text,
-        OSD_SHADOW,
-        px,
-    );
-    font::draw_text(
-        frame,
-        texture_width(s),
-        texture_height(s),
-        text_x,
-        text_y,
-        text,
-        OSD_TEXT,
-        px,
-    );
-}
-
-fn draw_osd(frame: &mut [u8], text: &str, texture_scale: usize) {
-    let s = texture_scale;
-    let px = 2 * s; // font pixel -> device pixels
-    let pad = 4 * s;
-    let margin = 8 * s;
-    let fw = texture_width(s);
-    let display_h = present_height() * s;
-
-    let text_w = font::text_width(text, px).min(fw.saturating_sub(2 * (margin + pad)));
-    let text_h = font::text_height(px);
-    let box_w = (text_w + 2 * pad).min(fw.saturating_sub(2 * margin));
-    let box_h = text_h + 2 * pad;
-    let box_x = margin;
-    let box_y = display_h.saturating_sub(margin + box_h);
-
-    fill_rect_blend(
-        frame,
-        Rect {
-            x: box_x,
-            y: box_y,
-            w: box_w,
-            h: box_h,
-        },
-        OSD_BG,
-        0.68,
-        s,
-    );
-    let text_x = box_x + pad;
-    let text_y = box_y + pad;
-    font::draw_text(
-        frame,
-        fw,
-        texture_height(s),
-        text_x + s,
-        text_y + s,
-        text,
-        OSD_SHADOW,
-        px,
-    );
-    font::draw_text(
-        frame,
-        fw,
-        texture_height(s),
-        text_x,
-        text_y,
-        text,
-        OSD_TEXT,
-        px,
-    );
-}
-
-/// Fill `rect` by alpha-blending `color` over the existing texture
-/// contents. Used for the semi-transparent overlay panel.
-pub(super) fn fill_rect_blend(
-    frame: &mut [u8],
-    rect: Rect,
-    color: u32,
-    alpha: f32,
-    texture_scale: usize,
-) {
-    let x1 = (rect.x + rect.w).min(texture_width(texture_scale));
-    let y1 = (rect.y + rect.h).min(texture_height(texture_scale));
-    for y in rect.y.min(texture_height(texture_scale))..y1 {
-        for x in rect.x.min(texture_width(texture_scale))..x1 {
-            blend_pixel(frame, x, y, color, alpha, texture_scale);
-        }
-    }
-}
-
-fn draw_text(frame: &mut [u8], x: usize, y: usize, text: &str, color: u32, texture_scale: usize) {
-    let mut cursor = x;
-    for ch in text.chars() {
-        if let Some(rows) = glyph(ch) {
-            draw_glyph(frame, cursor, y, rows, color, texture_scale);
-            cursor += 12 * texture_scale;
-        } else {
-            cursor += 6 * texture_scale;
-        }
-    }
-}
-
-fn draw_glyph(
-    frame: &mut [u8],
-    x: usize,
-    y: usize,
-    rows: [u8; 5],
-    color: u32,
-    texture_scale: usize,
-) {
-    let block = 2 * texture_scale;
-    for (row_idx, row) in rows.iter().enumerate() {
-        for col in 0..5 {
-            if row & (1 << (4 - col)) == 0 {
-                continue;
-            }
-            let px = x + col * block;
-            let py = y + row_idx * block;
-            fill_rect(
-                frame,
-                Rect {
-                    x: px,
-                    y: py,
-                    w: block,
-                    h: block,
-                },
-                color,
-                texture_scale,
-            );
-        }
-    }
-}
-
-fn glyph(ch: char) -> Option<[u8; 5]> {
-    match ch {
-        'C' => Some([0b01110, 0b10000, 0b10000, 0b10000, 0b01110]),
-        'D' => Some([0b11100, 0b10010, 0b10010, 0b10010, 0b11100]),
-        'F' => Some([0b11110, 0b10000, 0b11100, 0b10000, 0b10000]),
-        'H' => Some([0b10010, 0b10010, 0b11110, 0b10010, 0b10010]),
-        'L' => Some([0b10000, 0b10000, 0b10000, 0b10000, 0b11110]),
-        'O' => Some([0b01110, 0b10001, 0b10001, 0b10001, 0b01110]),
-        'P' => Some([0b11110, 0b10010, 0b11110, 0b10000, 0b10000]),
-        'R' => Some([0b11110, 0b10010, 0b11110, 0b10100, 0b10010]),
-        'V' => Some([0b10001, 0b10001, 0b01010, 0b01010, 0b00100]),
-        'W' => Some([0b10001, 0b10001, 0b10101, 0b10101, 0b01010]),
-        _ => None,
-    }
-}
-
-pub(super) fn draw_rect_bevel(
-    frame: &mut [u8],
-    rect: Rect,
-    light: u32,
-    dark: u32,
-    texture_scale: usize,
-) {
-    for inset in 0..texture_scale {
-        draw_hline_span(
-            frame,
-            rect.y + inset,
-            rect.x,
-            rect.x + rect.w,
-            light,
-            texture_scale,
-        );
-        draw_vline_span(
-            frame,
-            rect.x + inset,
-            rect.y,
-            rect.y + rect.h,
-            light,
-            texture_scale,
-        );
-        draw_hline_span(
-            frame,
-            rect.y + rect.h - 1 - inset,
-            rect.x,
-            rect.x + rect.w,
-            dark,
-            texture_scale,
-        );
-        draw_vline_span(
-            frame,
-            rect.x + rect.w - 1 - inset,
-            rect.y,
-            rect.y + rect.h,
-            dark,
-            texture_scale,
-        );
-    }
-}
-
-fn draw_hline(frame: &mut [u8], y: usize, color: u32, texture_scale: usize) {
-    draw_hline_span(
-        frame,
-        y,
-        0,
-        texture_width(texture_scale),
-        color,
-        texture_scale,
-    );
-}
-
-fn draw_hline_span(
-    frame: &mut [u8],
-    y: usize,
-    x0: usize,
-    x1: usize,
-    color: u32,
-    texture_scale: usize,
-) {
-    if y >= texture_height(texture_scale) {
-        return;
-    }
-    for x in x0.min(texture_width(texture_scale))..x1.min(texture_width(texture_scale)) {
-        put_pixel(frame, x, y, color, texture_scale);
-    }
-}
-
-fn draw_vline_span(
-    frame: &mut [u8],
-    x: usize,
-    y0: usize,
-    y1: usize,
-    color: u32,
-    texture_scale: usize,
-) {
-    if x >= texture_width(texture_scale) {
-        return;
-    }
-    for y in y0.min(texture_height(texture_scale))..y1.min(texture_height(texture_scale)) {
-        put_pixel(frame, x, y, color, texture_scale);
-    }
-}
-
-pub(super) fn fill_rect(frame: &mut [u8], rect: Rect, color: u32, texture_scale: usize) {
-    let x1 = (rect.x + rect.w).min(texture_width(texture_scale));
-    let y1 = (rect.y + rect.h).min(texture_height(texture_scale));
-    for y in rect.y.min(texture_height(texture_scale))..y1 {
-        for x in rect.x.min(texture_width(texture_scale))..x1 {
-            put_pixel(frame, x, y, color, texture_scale);
-        }
-    }
-}
-
-fn put_pixel(frame: &mut [u8], x: usize, y: usize, color: u32, texture_scale: usize) {
-    if x >= texture_width(texture_scale) || y >= texture_height(texture_scale) {
-        return;
-    }
-    let off = (y * texture_width(texture_scale) + x) * 4;
-    frame[off..off + 4].copy_from_slice(&color.to_le_bytes());
-}
-
-fn blend_pixel(frame: &mut [u8], x: usize, y: usize, color: u32, alpha: f32, texture_scale: usize) {
-    if alpha >= 1.0 {
-        put_pixel(frame, x, y, color, texture_scale);
-        return;
-    }
-    if x >= texture_width(texture_scale) || y >= texture_height(texture_scale) {
-        return;
-    }
-    let alpha = alpha.clamp(0.0, 1.0);
-    let off = (y * texture_width(texture_scale) + x) * 4;
-    let src = color.to_le_bytes();
-    for chan in 0..3 {
-        let dst = frame[off + chan] as f32;
-        let src = src[chan] as f32;
-        frame[off + chan] = (dst + (src - dst) * alpha).round() as u8;
-    }
-    frame[off + 3] = 0xFF;
 }
 
 impl App {
@@ -5915,7 +3378,8 @@ impl App {
         self.joystick_input_mode = cfg.joystick_input_mode;
         self.keyboard_joy_held = KeyboardJoystickHeld::default();
         self.about_machine_lines = crate::config::about_machine_lines(cfg);
-        self.deinterlacer = Deinterlacer::with_phosphor(crate::config::resolve_phosphor(cfg.phosphor));
+        self.deinterlacer =
+            Deinterlacer::with_phosphor(crate::config::resolve_phosphor(cfg.phosphor));
         self.ui.menu_open = false;
         self.ui.panel = None;
         self.powered_on = true;
@@ -7795,2993 +5259,17 @@ impl App {
     }
 }
 
-/// `[display] overscan = "tv"`: black out the deep-overscan margins like the
-/// bezel of a CRT. Demos routinely leave junk in the deep overscan (e.g. HAM
-/// streams that converge off-screen, as on the 9 Fingers title); a TV hides
-/// it and so does this mask. The emulated framebuffer itself always carries
-/// the full field; this runs on the presentation copy only.
-///
-/// The window is a realistic PAL TV visible area rather than the bare standard
-/// window: real sets show a margin of overscan, which intentional overscan
-/// displays rely on, while the deep-overscan junk the mask exists to hide sits
-/// further out. Default TV presentation keeps 24 lo-res pixels of horizontal
-/// overscan beside the standard PAL window; full overscan remains available
-/// through `Overscan::Full`. The mask is horizontal only: vertical border
-/// colour changes are part of the Denise output and can be intentional
-/// effects, so rows above or below the standard display remain as rendered.
-/// `h_shift` is any horizontal presentation shift already applied to the
-/// frame, so the bezel tracks the shifted picture instead of clipping its left
-/// edge.
-fn mask_present_frame_to_tv(fb: &mut [u32], h_shift: usize, _standard_top_row: usize) {
-    debug_assert!(fb.len() >= FB_PIXELS);
-    let black = rgba(0, 0, 0);
-    let (source_left, source_right) = tv_source_h_bounds();
-    let left = source_left.saturating_sub(h_shift);
-    let right = source_right.saturating_sub(h_shift).min(FB_WIDTH).max(left);
-    for row in fb.chunks_mut(FB_WIDTH) {
-        row[..left].fill(black);
-        if right < FB_WIDTH {
-            row[right..].fill(black);
-        }
-    }
-}
+mod host_input;
+mod present;
+mod statusbar;
+pub(super) use present::{scale_rect, texture_height, texture_width, Rect};
+pub(super) use statusbar::{draw_rect_bevel, fill_rect, fill_rect_blend};
 
-/// The framebuffer row carrying the standard window's first line after
-/// `center_present_frame_for_visible_start` has run: the centring shift
-/// plus however many overscan rows were already visible above it.
-fn standard_window_top_row(visible_start_vpos: u32) -> usize {
-    let overscan_rows_already_visible =
-        STANDARD_PAL_VISIBLE_START_VPOS.saturating_sub(visible_start_vpos) as usize;
-    presentation_source_y_offset(visible_start_vpos) + overscan_rows_already_visible
-}
-
-/// Whether horizontal recentring is enabled. On unless COPPERLINE_HCENTER is
-/// set to a falsey value (0/false/off/no), so full-overscan presentation can
-/// show the standard display exactly as rendered when debugging alignment.
-fn hcenter_enabled() -> bool {
-    match crate::envcfg::var("COPPERLINE_HCENTER") {
-        Some(v) => !matches!(
-            v.trim().to_ascii_lowercase().as_str(),
-            "0" | "false" | "off" | "no"
-        ),
-        None => true,
-    }
-}
-
-fn threaded_render_enabled() -> bool {
-    match crate::envcfg::var("COPPERLINE_THREADED_RENDER") {
-        Some(v) => !matches!(
-            v.trim().to_ascii_lowercase().as_str(),
-            "" | "0" | "false" | "off" | "no"
-        ),
-        None => true,
-    }
-}
-
-/// Shift the rendered frame left by `shift` framebuffer pixels, filling the
-/// vacated right columns with black. Used to recentre a standard display
-/// whose deep left overscan would otherwise push the picture right of
-/// centre. A no-op when `shift` is 0 (overscan frames).
-pub(crate) fn center_present_frame_horizontally(fb: &mut [u32], shift: usize) {
-    debug_assert!(fb.len() >= FB_PIXELS);
-    if shift == 0 {
-        return;
-    }
-    let shift = shift.min(FB_WIDTH);
-    let black = rgba(0, 0, 0);
-    for y in 0..FB_HEIGHT {
-        let row = &mut fb[y * FB_WIDTH..(y + 1) * FB_WIDTH];
-        row.copy_within(shift.., 0);
-        row[FB_WIDTH - shift..].fill(black);
-    }
-}
-
-pub(crate) fn center_present_frame_for_visible_start(fb: &mut [u32], visible_start_vpos: u32) {
-    debug_assert!(fb.len() >= FB_PIXELS);
-    let offset = presentation_source_y_offset(visible_start_vpos);
-    if offset == 0 {
-        return;
-    }
-
-    for y in (0..FB_HEIGHT - offset).rev() {
-        let src = y * FB_WIDTH;
-        let dst = (y + offset) * FB_WIDTH;
-        fb.copy_within(src..src + FB_WIDTH, dst);
-    }
-    fb[..offset * FB_WIDTH].fill(rgba(0, 0, 0));
-}
-
-fn presentation_source_y_offset(visible_start_vpos: u32) -> usize {
-    let standard_offset = FB_HEIGHT.saturating_sub(STANDARD_PAL_VISIBLE_LINES) / 2;
-    let overscan_rows_already_visible =
-        STANDARD_PAL_VISIBLE_START_VPOS.saturating_sub(visible_start_vpos) as usize;
-    standard_offset.saturating_sub(overscan_rows_already_visible)
-}
-
-fn presentation_h_shift_for(snapshot: &RenderRegisterSnapshot, overscan: Overscan) -> usize {
-    match overscan {
-        // TV mode is an aperture over the emulated framebuffer, matching the
-        // fixed source cutout used by reference emulators. Do not copy pixels
-        // sideways here: a standard hi-res screen already occupies the right
-        // edge of Copperline's 716-pixel cutout.
-        Overscan::Tv => 0,
-        Overscan::Full => bitplane::present_h_shift_for(snapshot),
-    }
-}
-
-fn tv_source_h_bounds() -> (usize, usize) {
-    let left = bitplane::STANDARD_VISIBLE_X0.saturating_sub(TV_HORIZONTAL_OVERSCAN_MARGIN);
-    let right = bitplane::STANDARD_VISIBLE_X0
-        .saturating_add(STANDARD_PAL_VISIBLE_WIDTH)
-        .saturating_add(TV_HORIZONTAL_OVERSCAN_MARGIN)
-        .min(FB_WIDTH)
-        .max(left);
-    (left, right)
-}
-
-fn should_render_emulated_frame(last_rendered: Option<u64>, current: u64) -> bool {
-    last_rendered != Some(current)
-}
-
-fn status_with_latched_fdd_track(
-    status: FrontPanelStatus,
-    last_fdd_track: &mut Option<u8>,
-) -> FrontPanelStatus {
-    let fdd_track = match status.fdd_track {
-        Some(track) => {
-            *last_fdd_track = Some(track);
-            Some(track)
-        }
-        None => *last_fdd_track,
-    };
-    FrontPanelStatus {
-        fdd_track,
-        ..status
-    }
-}
-
-fn take_integral_mouse_delta(value: &mut f64) -> i32 {
-    let whole = value.trunc();
-    if whole > i32::MAX as f64 {
-        *value = 0.0;
-        i32::MAX
-    } else if whole < i32::MIN as f64 {
-        *value = 0.0;
-        i32::MIN
-    } else {
-        *value -= whole;
-        whole as i32
-    }
-}
-
-/// Hex digit for a key, for the debugger's address entry box.
-/// Map a key to the character it types into the debugger entry box: digits, all
-/// letters (for register names and the EQ/NE/.../AND/IGN condition mnemonics and
-/// M<hex> memory operands), and Space. `push_entry_char` filters and uppercases.
-fn entry_char_for_key(code: KeyCode) -> Option<char> {
-    use KeyCode::*;
-    Some(match code {
-        Digit0 | Numpad0 => '0',
-        Digit1 | Numpad1 => '1',
-        Digit2 | Numpad2 => '2',
-        Digit3 | Numpad3 => '3',
-        Digit4 | Numpad4 => '4',
-        Digit5 | Numpad5 => '5',
-        Digit6 | Numpad6 => '6',
-        Digit7 | Numpad7 => '7',
-        Digit8 | Numpad8 => '8',
-        Digit9 | Numpad9 => '9',
-        KeyA => 'A',
-        KeyB => 'B',
-        KeyC => 'C',
-        KeyD => 'D',
-        KeyE => 'E',
-        KeyF => 'F',
-        KeyG => 'G',
-        KeyH => 'H',
-        KeyI => 'I',
-        KeyJ => 'J',
-        KeyK => 'K',
-        KeyL => 'L',
-        KeyM => 'M',
-        KeyN => 'N',
-        KeyO => 'O',
-        KeyP => 'P',
-        KeyQ => 'Q',
-        KeyR => 'R',
-        KeyS => 'S',
-        KeyT => 'T',
-        KeyU => 'U',
-        KeyV => 'V',
-        KeyW => 'W',
-        KeyX => 'X',
-        KeyY => 'Y',
-        KeyZ => 'Z',
-        Space => ' ',
-        _ => return None,
-    })
-}
-
-/// Format a calibration session for the panel: pad identity, one row per
-/// step with its captured binding, and a prompt for what to do next.
-fn build_calibration_view(session: &crate::gamepad::CalibrationSession) -> ui::CalibrationView {
-    use crate::gamepad::CalibrationSession;
-    let pad_line = if session.backend_missing() {
-        "No gamepad backend available on this host".to_string()
-    } else {
-        match (session.pad_name(), session.connected()) {
-            (Some(name), true) => format!("Controller: {name}"),
-            (Some(name), false) => format!("Reconnect {name}..."),
-            (None, _) => "Connect a controller...".to_string(),
-        }
-    };
-    let rows = (0..CalibrationSession::step_count())
-        .map(|index| ui::CalRow {
-            label: CalibrationSession::step_label(index),
-            binding: session.binding_text(index),
-            current: session.current_step() == Some(index),
-        })
-        .collect();
-    let status = if session.backend_missing() {
-        "Calibration needs a gamepad backend (not available headless).".to_string()
-    } else if session.done() {
-        if session.live_test().is_empty() {
-            "All steps captured. Push controls to test, then Save.".to_string()
-        } else {
-            format!("Testing: {}", session.live_test())
-        }
-    } else if !session.connected() {
-        "Waiting for a controller to be connected.".to_string()
-    } else if session.can_skip() {
-        "Push and hold the control, or Skip if the pad lacks it.".to_string()
-    } else {
-        "Push and hold the control on the pad.".to_string()
-    };
-    ui::CalibrationView {
-        pad_line,
-        rows,
-        status,
-    }
-}
-
-/// Translate a winit `KeyCode` to an Amiga raw scan code.
-/// Covers the alphanumeric block, common modifiers we treat as their
-/// nearest Amiga equivalents, function keys, and arrows. Anything not
-/// in the table returns None (the keypress is silently dropped).
-fn host_to_amiga_rawkey(code: KeyCode) -> Option<u8> {
-    use KeyCode::*;
-    Some(match code {
-        // Letters (row-by-row, Amiga's funny layout)
-        KeyA => 0x20,
-        KeyB => 0x35,
-        KeyC => 0x33,
-        KeyD => 0x22,
-        KeyE => 0x12,
-        KeyF => 0x23,
-        KeyG => 0x24,
-        KeyH => 0x25,
-        KeyI => 0x17,
-        KeyJ => 0x26,
-        KeyK => 0x27,
-        KeyL => 0x28,
-        KeyM => 0x37,
-        KeyN => 0x36,
-        KeyO => 0x18,
-        KeyP => 0x19,
-        KeyQ => 0x10,
-        KeyR => 0x13,
-        KeyS => 0x21,
-        KeyT => 0x14,
-        KeyU => 0x16,
-        KeyV => 0x34,
-        KeyW => 0x11,
-        KeyX => 0x32,
-        KeyY => 0x15,
-        KeyZ => 0x31,
-        // Top-row digits
-        Digit1 => 0x01,
-        Digit2 => 0x02,
-        Digit3 => 0x03,
-        Digit4 => 0x04,
-        Digit5 => 0x05,
-        Digit6 => 0x06,
-        Digit7 => 0x07,
-        Digit8 => 0x08,
-        Digit9 => 0x09,
-        Digit0 => 0x0A,
-        // Punctuation
-        Backquote => 0x00,
-        Minus => 0x0B,
-        Equal => 0x0C,
-        Backslash => 0x0D,
-        BracketLeft => 0x1A,
-        BracketRight => 0x1B,
-        Semicolon => 0x29,
-        Quote => 0x2A,
-        Comma => 0x38,
-        Period => 0x39,
-        Slash => 0x3A,
-        // International keys: the ISO 102nd key between left Shift and
-        // Z is Amiga rawkey $30; the Japanese Ro key sits in the same
-        // matrix position on layouts that have it.
-        IntlBackslash | IntlRo => 0x30,
-        // Control
-        Space => 0x40,
-        Enter => 0x44,
-        Backspace => 0x41,
-        Tab => 0x42,
-        Escape => 0x45,
-        Delete => 0x46,
-        // Amiga Help: F11 host-side (no dedicated host key exists).
-        F11 => 0x5F,
-        ShiftLeft => 0x60,
-        ShiftRight => 0x61,
-        CapsLock => 0x62,
-        // The Amiga keyboard has a single Ctrl key (left side); there is no
-        // right Ctrl. Map host ControlLeft to it. Host ControlRight has no
-        // Amiga counterpart, so alias it to Right Amiga ($67) alongside
-        // SuperRight -- many PC/laptop keyboards lack a right Super/Win key,
-        // leaving Right Amiga otherwise unreachable.
-        ControlLeft => 0x63,
-        AltLeft => 0x64,
-        AltRight => 0x65,
-        SuperLeft => 0x66,
-        SuperRight | ControlRight => 0x67,
-        // Arrows
-        ArrowUp => 0x4C,
-        ArrowDown => 0x4D,
-        ArrowRight => 0x4E,
-        ArrowLeft => 0x4F,
-        // Function keys
-        F1 => 0x50,
-        F2 => 0x51,
-        F3 => 0x52,
-        F4 => 0x53,
-        F5 => 0x54,
-        F6 => 0x55,
-        F7 => 0x56,
-        F8 => 0x57,
-        F9 => 0x58,
-        F10 => 0x59,
-        // Numpad
-        Numpad0 => 0x0F,
-        Numpad1 => 0x1D,
-        Numpad2 => 0x1E,
-        Numpad3 => 0x1F,
-        Numpad4 => 0x2D,
-        Numpad5 => 0x2E,
-        Numpad6 => 0x2F,
-        Numpad7 => 0x3D,
-        Numpad8 => 0x3E,
-        Numpad9 => 0x3F,
-        NumpadDecimal => 0x3C,
-        NumpadEnter => 0x43,
-        NumpadSubtract => 0x4A,
-        NumpadAdd => 0x5E,
-        NumpadMultiply => 0x5D,
-        NumpadDivide => 0x5C,
-        NumpadParenLeft => 0x5A,
-        NumpadParenRight => 0x5B,
-        _ => return None,
-    })
-}
-
-/// Parse an Amiga raw key as either a decimal/hex code or a common
-/// key name for scripted input.
-pub fn parse_amiga_key(s: &str) -> Option<u8> {
-    let trimmed = s.trim();
-    if let Some(raw) = parse_u8(trimmed) {
-        return Some(raw);
-    }
-    let name = trimmed.to_ascii_lowercase().replace(['-', '_', '+'], "");
-    Some(match name.as_str() {
-        "a" => 0x20,
-        "b" => 0x35,
-        "c" => 0x33,
-        "d" => 0x22,
-        "e" => 0x12,
-        "f" => 0x23,
-        "g" => 0x24,
-        "h" => 0x25,
-        "i" => 0x17,
-        "j" => 0x26,
-        "k" => 0x27,
-        "l" => 0x28,
-        "m" => 0x37,
-        "n" => 0x36,
-        "o" => 0x18,
-        "p" => 0x19,
-        "q" => 0x10,
-        "r" => 0x13,
-        "s" => 0x21,
-        "t" => 0x14,
-        "u" => 0x16,
-        "v" => 0x34,
-        "w" => 0x11,
-        "x" => 0x32,
-        "y" => 0x15,
-        "z" => 0x31,
-        "1" => 0x01,
-        "2" => 0x02,
-        "3" => 0x03,
-        "4" => 0x04,
-        "5" => 0x05,
-        "6" => 0x06,
-        "7" => 0x07,
-        "8" => 0x08,
-        "9" => 0x09,
-        "0" => 0x0A,
-        "space" => 0x40,
-        "backspace" | "bs" => 0x41,
-        "tab" => 0x42,
-        "enter" | "return" => 0x44,
-        "escape" | "esc" => 0x45,
-        "delete" | "del" => 0x46,
-        "up" | "arrowup" => 0x4C,
-        "down" | "arrowdown" => 0x4D,
-        "right" | "arrowright" => 0x4E,
-        "left" | "arrowleft" => 0x4F,
-        "f1" => 0x50,
-        "f2" => 0x51,
-        "f3" => 0x52,
-        "f4" => 0x53,
-        "f5" => 0x54,
-        "f6" => 0x55,
-        "f7" => 0x56,
-        "f8" => 0x57,
-        "f9" => 0x58,
-        "f10" => 0x59,
-        "lshift" | "leftshift" | "shift" => 0x60,
-        "rshift" | "rightshift" => 0x61,
-        "caps" | "capslock" => 0x62,
-        "ctrl" | "control" | "lctrl" | "leftctrl" | "leftcontrol" | "rctrl" | "rightctrl"
-        | "rightcontrol" => 0x63,
-        "lalt" | "leftalt" | "alt" => 0x64,
-        "ralt" | "rightalt" => 0x65,
-        "lami" | "leftami" | "lamiga" | "leftamiga" | "ami" | "amiga" | "cmd" | "command"
-        | "super" | "meta" => 0x66,
-        "rami" | "rightami" | "ramiga" | "rightamiga" => 0x67,
-        _ => return None,
-    })
-}
-
-fn parse_u8(s: &str) -> Option<u8> {
-    if let Some(rest) = s.strip_prefix("0x").or_else(|| s.strip_prefix("$")) {
-        u8::from_str_radix(rest, 16).ok()
-    } else {
-        s.parse().ok()
-    }
-}
-
-fn save_present_frame(
-    path: &std::path::Path,
-    present_fb: &[u32],
-    src_rows: usize,
-    overscan: Overscan,
-    standard_tv_aperture: bool,
-) -> anyhow::Result<()> {
-    if crate::envcfg::flag("COPPERLINE_SHOT_RAW") {
-        return screenshot::save(
-            path,
-            &present_fb[..src_rows * FB_WIDTH],
-            FB_WIDTH as u32,
-            src_rows as u32,
-        );
-    }
-
-    if overscan == Overscan::Tv && standard_tv_aperture {
-        return screenshot::save_cropped_clamped(
-            path,
-            present_fb,
-            FB_WIDTH,
-            src_rows,
-            TV_PAL_PRESENT_SOURCE_X,
-            TV_PAL_PRESENT_SOURCE_Y,
-            TV_PAL_PRESENT_WIDTH,
-            TV_PAL_PRESENT_HEIGHT,
-        );
-    }
-
-    screenshot::save_scaled_y(
-        path,
-        present_fb,
-        FB_WIDTH as u32,
-        src_rows as u32,
-        present_height() as u32,
-    )
-}
-
-fn is_standard_pal_presentation(geometry: FrameGeometry, src_rows: usize) -> bool {
-    !geometry.programmable && geometry.frame_lines >= 312 && src_rows == OUT_HEIGHT
-}
-
-fn uses_standard_pal_tv_aperture(
-    geometry: FrameGeometry,
-    src_rows: usize,
-    snapshot: &RenderRegisterSnapshot,
-) -> bool {
-    is_standard_pal_presentation(geometry, src_rows)
-        && bitplane::uses_standard_horizontal_content(snapshot)
-}
+pub use host_input::parse_amiga_key;
+use host_input::*;
+pub(crate) use present::center_present_frame_for_visible_start;
+use present::*;
+use statusbar::*;
 
 #[cfg(test)]
-mod tests {
-    use super::ui::{Panel, UiControl};
-    use super::{
-        bar_layout, center_present_frame_for_visible_start, center_present_frame_horizontally,
-        control_at, copperline_icon_image, copperline_logo_image, copy_present_frame,
-        copy_window_present_frame, draw_status_bar, fdd_track_counter_rect, fdd_track_digit_rect,
-        host_shortcut_modifier_pressed, host_to_amiga_rawkey, joystick_mode_uses_keyboard,
-        joystick_toggle_rect, keyboard_joystick_key_for, led_row_rect, mask_present_frame_to_tv,
-        paint_test_screen, parse_amiga_key, pause_button_rect, power_button_rect, present_height,
-        presentation_h_shift_for, presentation_source_y_offset, raw_device_qualifier_family_held,
-        raw_device_qualifier_rawkey, rawkey_is_held, rawkey_transition_is_duplicate,
-        reboot_button_rect, repeated_main_key_should_drop, rgba, shot_button_rect,
-        should_render_emulated_frame, standard_window_top_row, status_with_latched_fdd_track,
-        take_integral_mouse_delta, texture_height, texture_width, tv_aperture_source_row,
-        tv_source_h_bounds, volume_percent_from_pos, volume_slider_track_rect, BarControl,
-        DriveBar, JoystickInputMode, KeyboardJoystickHeld, KeyboardJoystickKey, MediaBar,
-        StatusBarView, ToolPanelKind, AMIGA_RAWKEY_LEFT_ALT, AMIGA_RAWKEY_LEFT_SHIFT,
-        AMIGA_RAWKEY_RIGHT_ALT, AMIGA_RAWKEY_RIGHT_SHIFT, BUTTON_GLYPH, BUTTON_GLYPH_DISABLED,
-        CD_BODY, CD_LED_OFF, CD_LED_ON, DISK_BODY, DISK_BODY_SHADOW, DISK_LABEL, FDD_LED_OFF,
-        FDD_LED_ON, HDD_LED_OFF, HDD_LED_ON, POWER_GLYPH_OFF, POWER_GLYPH_ON, POWER_LED_OFF,
-        POWER_LED_ON, STANDARD_PAL_VISIBLE_LINES, STANDARD_PAL_VISIBLE_START_VPOS, STATUS_BG,
-        TRACK_SEGMENT_OFF, TRACK_SEGMENT_ON, TV_PAL_LIVE_PAD_X, TV_PAL_PRESENT_HEIGHT,
-        TV_PAL_PRESENT_SOURCE_X, TV_PAL_PRESENT_SOURCE_Y, TV_PAL_PRESENT_WIDTH, VOLUME_FILL,
-        VOLUME_GLYPH_X,
-    };
-    use crate::audio::{AudioSink, NullSink};
-    use crate::bus::{FrontPanelStatus, RenderRegisterSnapshot};
-    use crate::config::{Overscan, WarpSpeed};
-    use crate::video::{FB_HEIGHT, FB_PIXELS, FB_WIDTH};
-    use std::cell::RefCell;
-    use std::path::PathBuf;
-    use std::rc::Rc;
-    use winit::event::{ElementState, RawKeyEvent};
-    use winit::keyboard::{KeyCode, ModifiersState, PhysicalKey};
-
-    /// A typical session: DF0 connected with a disk in, no CD drive.
-    fn single_drive_media() -> MediaBar {
-        let mut drives = [DriveBar::default(); 4];
-        drives[0] = DriveBar {
-            connected: true,
-            inserted: true,
-            multi: false,
-        };
-        MediaBar { drives, cd: None }
-    }
-
-    fn media(connected: usize, cd: Option<bool>) -> MediaBar {
-        let mut drives = [DriveBar::default(); 4];
-        for drive in drives.iter_mut().take(connected) {
-            *drive = DriveBar {
-                connected: true,
-                inserted: true,
-                multi: false,
-            };
-        }
-        MediaBar { drives, cd }
-    }
-
-    fn view(status: FrontPanelStatus, powered_on: bool, paused: bool) -> StatusBarView {
-        StatusBarView {
-            status,
-            powered_on,
-            paused,
-            media: single_drive_media(),
-            joystick_input_mode: JoystickInputMode::Gamepad,
-            hover: None,
-        }
-    }
-
-    #[test]
-    fn host_mapping_includes_amiga_modifiers() {
-        assert_eq!(host_to_amiga_rawkey(KeyCode::ControlLeft), Some(0x63));
-        assert_eq!(host_to_amiga_rawkey(KeyCode::AltLeft), Some(0x64));
-        assert_eq!(host_to_amiga_rawkey(KeyCode::AltRight), Some(0x65));
-        assert_eq!(host_to_amiga_rawkey(KeyCode::SuperLeft), Some(0x66));
-        assert_eq!(host_to_amiga_rawkey(KeyCode::SuperRight), Some(0x67));
-        // The Amiga has no right Ctrl, so host ControlRight doubles as a
-        // Right Amiga ($67) alias for keyboards without a right Super key.
-        assert_eq!(host_to_amiga_rawkey(KeyCode::ControlRight), Some(0x67));
-    }
-
-    #[test]
-    fn host_repeat_filter_accepts_unheld_amiga_qualifier_press() {
-        let mut held = [false; 128];
-
-        assert!(!repeated_main_key_should_drop(
-            &held,
-            KeyCode::ShiftRight,
-            ElementState::Pressed,
-            true,
-            false
-        ));
-
-        held[AMIGA_RAWKEY_RIGHT_SHIFT as usize] = true;
-        assert!(repeated_main_key_should_drop(
-            &held,
-            KeyCode::ShiftRight,
-            ElementState::Pressed,
-            true,
-            false
-        ));
-
-        assert!(repeated_main_key_should_drop(
-            &held,
-            KeyCode::F12,
-            ElementState::Pressed,
-            true,
-            false
-        ));
-        assert!(!repeated_main_key_should_drop(
-            &held,
-            KeyCode::ShiftRight,
-            ElementState::Pressed,
-            false,
-            false
-        ));
-        assert!(!repeated_main_key_should_drop(
-            &held,
-            KeyCode::ArrowRight,
-            ElementState::Pressed,
-            true,
-            true
-        ));
-    }
-
-    #[test]
-    fn raw_device_qualifier_filter_is_limited_to_amiga_modifier_lines() {
-        assert_eq!(
-            raw_device_qualifier_rawkey(KeyCode::ShiftLeft),
-            Some(AMIGA_RAWKEY_LEFT_SHIFT)
-        );
-        assert_eq!(
-            raw_device_qualifier_rawkey(KeyCode::ShiftRight),
-            Some(AMIGA_RAWKEY_RIGHT_SHIFT)
-        );
-        assert_eq!(
-            raw_device_qualifier_rawkey(KeyCode::AltLeft),
-            Some(AMIGA_RAWKEY_LEFT_ALT)
-        );
-        assert_eq!(
-            raw_device_qualifier_rawkey(KeyCode::AltRight),
-            Some(AMIGA_RAWKEY_RIGHT_ALT)
-        );
-        assert_eq!(raw_device_qualifier_rawkey(KeyCode::KeyS), None);
-        assert_eq!(raw_device_qualifier_rawkey(KeyCode::ArrowRight), None);
-    }
-
-    #[test]
-    fn raw_device_qualifier_family_reports_physical_side_state() {
-        let mut held = [false; 128];
-        assert!(!raw_device_qualifier_family_held(
-            &held,
-            AMIGA_RAWKEY_LEFT_ALT,
-            AMIGA_RAWKEY_RIGHT_ALT
-        ));
-
-        held[AMIGA_RAWKEY_LEFT_ALT as usize] = true;
-        assert!(raw_device_qualifier_family_held(
-            &held,
-            AMIGA_RAWKEY_LEFT_ALT,
-            AMIGA_RAWKEY_RIGHT_ALT
-        ));
-
-        held[AMIGA_RAWKEY_LEFT_ALT as usize] = false;
-        held[AMIGA_RAWKEY_RIGHT_ALT as usize] = true;
-        assert!(raw_device_qualifier_family_held(
-            &held,
-            AMIGA_RAWKEY_LEFT_ALT,
-            AMIGA_RAWKEY_RIGHT_ALT
-        ));
-    }
-
-    #[test]
-    fn amiga_qualifier_transitions_ignore_duplicate_host_events() {
-        let mut held = [false; 128];
-
-        assert!(rawkey_transition_is_duplicate(
-            &held,
-            AMIGA_RAWKEY_LEFT_SHIFT,
-            false
-        ));
-        assert!(!rawkey_transition_is_duplicate(
-            &held,
-            AMIGA_RAWKEY_LEFT_SHIFT,
-            true
-        ));
-
-        held[AMIGA_RAWKEY_LEFT_SHIFT as usize] = true;
-        assert!(rawkey_transition_is_duplicate(
-            &held,
-            AMIGA_RAWKEY_LEFT_SHIFT,
-            true
-        ));
-        assert!(!rawkey_transition_is_duplicate(
-            &held,
-            AMIGA_RAWKEY_LEFT_SHIFT,
-            false
-        ));
-    }
-
-    #[test]
-    fn aggregate_modifier_release_clears_held_amiga_qualifiers() {
-        let mut app = test_app();
-        for rawkey in [
-            AMIGA_RAWKEY_LEFT_SHIFT,
-            AMIGA_RAWKEY_RIGHT_SHIFT,
-            AMIGA_RAWKEY_LEFT_ALT,
-            AMIGA_RAWKEY_RIGHT_ALT,
-        ] {
-            app.handle_amiga_key_event(rawkey, true);
-            assert!(rawkey_is_held(&app.held_rawkeys, rawkey));
-        }
-
-        app.update_host_modifiers(ModifiersState::SHIFT);
-        assert!(rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_LEFT_SHIFT));
-        assert!(rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_RIGHT_SHIFT));
-        assert!(!rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_LEFT_ALT));
-        assert!(!rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_RIGHT_ALT));
-
-        app.update_host_modifiers(ModifiersState::empty());
-        assert!(!rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_LEFT_SHIFT));
-        assert!(!rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_RIGHT_SHIFT));
-    }
-
-    #[test]
-    fn raw_device_alt_hold_blocks_altgr_aggregate_cleanup() {
-        let mut app = test_app();
-        app.main_window_focused = true;
-
-        app.handle_raw_device_key_event(RawKeyEvent {
-            physical_key: PhysicalKey::Code(KeyCode::AltLeft),
-            state: ElementState::Pressed,
-        });
-        app.update_host_modifiers(ModifiersState::ALT);
-        assert!(rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_LEFT_ALT));
-
-        app.update_host_modifiers(ModifiersState::empty());
-        assert!(rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_LEFT_ALT));
-
-        app.handle_raw_device_key_event(RawKeyEvent {
-            physical_key: PhysicalKey::Code(KeyCode::AltRight),
-            state: ElementState::Pressed,
-        });
-        assert!(rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_LEFT_ALT));
-        assert!(rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_RIGHT_ALT));
-
-        app.handle_raw_device_key_event(RawKeyEvent {
-            physical_key: PhysicalKey::Code(KeyCode::AltLeft),
-            state: ElementState::Released,
-        });
-        assert!(!rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_LEFT_ALT));
-        assert!(rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_RIGHT_ALT));
-    }
-
-    #[test]
-    fn raw_device_release_clears_one_side_while_aggregate_modifier_remains() {
-        let mut app = test_app();
-        app.main_window_focused = true;
-
-        app.handle_raw_device_key_event(RawKeyEvent {
-            physical_key: PhysicalKey::Code(KeyCode::ShiftLeft),
-            state: ElementState::Pressed,
-        });
-        app.handle_raw_device_key_event(RawKeyEvent {
-            physical_key: PhysicalKey::Code(KeyCode::ShiftRight),
-            state: ElementState::Pressed,
-        });
-        assert!(rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_LEFT_SHIFT));
-        assert!(rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_RIGHT_SHIFT));
-
-        app.handle_raw_device_key_event(RawKeyEvent {
-            physical_key: PhysicalKey::Code(KeyCode::ShiftLeft),
-            state: ElementState::Released,
-        });
-        app.update_host_modifiers(ModifiersState::SHIFT);
-        assert!(!rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_LEFT_SHIFT));
-        assert!(rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_RIGHT_SHIFT));
-
-        app.handle_raw_device_key_event(RawKeyEvent {
-            physical_key: PhysicalKey::Code(KeyCode::ShiftRight),
-            state: ElementState::Released,
-        });
-        assert!(!rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_RIGHT_SHIFT));
-    }
-
-    #[test]
-    fn raw_device_alt_right_respects_keyboard_joystick_ownership() {
-        let mut app = test_app();
-        app.main_window_focused = true;
-        app.set_joystick_input_mode(JoystickInputMode::Keyboard);
-
-        app.handle_raw_device_key_event(RawKeyEvent {
-            physical_key: PhysicalKey::Code(KeyCode::AltRight),
-            state: ElementState::Pressed,
-        });
-        assert!(app.keyboard_joy_held.fire_right_alt);
-        assert!(!rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_RIGHT_ALT));
-
-        app.handle_raw_device_key_event(RawKeyEvent {
-            physical_key: PhysicalKey::Code(KeyCode::AltRight),
-            state: ElementState::Released,
-        });
-        assert!(!app.keyboard_joy_held.fire_right_alt);
-        assert!(!rawkey_is_held(&app.held_rawkeys, AMIGA_RAWKEY_RIGHT_ALT));
-    }
-
-    #[test]
-    fn keyboard_joystick_mapping_matches_fsuae_controls() {
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::ArrowUp),
-            Some(KeyboardJoystickKey::Up)
-        );
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::ArrowDown),
-            Some(KeyboardJoystickKey::Down)
-        );
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::ArrowLeft),
-            Some(KeyboardJoystickKey::Left)
-        );
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::ArrowRight),
-            Some(KeyboardJoystickKey::Right)
-        );
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::ControlRight),
-            Some(KeyboardJoystickKey::FireRightCtrl)
-        );
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::AltRight),
-            Some(KeyboardJoystickKey::FireRightAlt)
-        );
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::KeyC),
-            Some(KeyboardJoystickKey::Red)
-        );
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::KeyX),
-            Some(KeyboardJoystickKey::Blue)
-        );
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::KeyD),
-            Some(KeyboardJoystickKey::Green)
-        );
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::KeyS),
-            Some(KeyboardJoystickKey::Yellow)
-        );
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::Enter),
-            Some(KeyboardJoystickKey::Play)
-        );
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::KeyZ),
-            Some(KeyboardJoystickKey::Rewind)
-        );
-        assert_eq!(
-            keyboard_joystick_key_for(KeyCode::KeyA),
-            Some(KeyboardJoystickKey::Forward)
-        );
-        assert_eq!(keyboard_joystick_key_for(KeyCode::ControlLeft), None);
-    }
-
-    #[test]
-    fn keyboard_joystick_fire_aliases_release_independently() {
-        let mut held = KeyboardJoystickHeld::default();
-        held.set(KeyboardJoystickKey::FireRightCtrl, true);
-        held.set(KeyboardJoystickKey::Red, true);
-        assert!(held.joystick_state().fire);
-
-        held.set(KeyboardJoystickKey::FireRightCtrl, false);
-        assert!(held.joystick_state().fire);
-
-        held.set(KeyboardJoystickKey::Red, false);
-        assert!(!held.joystick_state().fire);
-    }
-
-    #[test]
-    fn joystick_input_mode_toggles_between_two_explicit_modes() {
-        // The toggle flips directly between the two modes; there is no hidden
-        // auto-detect state, so the keyboard mapping is engaged exactly when
-        // (and only when) the mode is Keyboard.
-        assert_eq!(
-            JoystickInputMode::Gamepad.next(),
-            JoystickInputMode::Keyboard
-        );
-        assert_eq!(
-            JoystickInputMode::Keyboard.next(),
-            JoystickInputMode::Gamepad
-        );
-
-        assert!(joystick_mode_uses_keyboard(JoystickInputMode::Keyboard));
-        assert!(!joystick_mode_uses_keyboard(JoystickInputMode::Gamepad));
-    }
-
-    #[test]
-    fn joystick_toggle_clears_worst_case_media() {
-        // The toggle sits at a fixed x just left of the volume glyph. The
-        // widest media layout (four floppies plus a CD) must not reach it, and
-        // it must stay left of the volume control's hit area.
-        let toggle = joystick_toggle_rect();
-        let layout = bar_layout(&media(4, Some(true)));
-        let media_right = layout
-            .cd_eject
-            .into_iter()
-            .chain(layout.drive_eject.into_iter().flatten())
-            .map(|r| r.x + r.w)
-            .max()
-            .unwrap();
-        assert!(
-            media_right <= toggle.x,
-            "media right edge {media_right} overlaps joystick toggle at {}",
-            toggle.x
-        );
-        assert!(toggle.x + toggle.w <= VOLUME_GLYPH_X);
-    }
-
-    #[test]
-    fn joystick_toggle_is_hit_tested_and_draws_each_mode() {
-        let layout = bar_layout(&single_drive_media());
-        let toggle = joystick_toggle_rect();
-        let center = (
-            (toggle.x + toggle.w / 2) as i32,
-            (toggle.y + toggle.h / 2) as i32,
-        );
-        assert_eq!(control_at(center, &layout), Some(BarControl::Joystick));
-
-        // Each mode lights the green glyph somewhere in the button (gamepad
-        // body vs. keyboard keys), so the two states are visually distinct.
-        let scale = 1;
-        for mode in [JoystickInputMode::Gamepad, JoystickInputMode::Keyboard] {
-            let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-            let mut v = view(FrontPanelStatus::default(), true, false);
-            v.joystick_input_mode = mode;
-            draw_status_bar(&mut frame, &v, scale);
-            let lit = (toggle.y..toggle.y + toggle.h).any(|y| {
-                (toggle.x..toggle.x + toggle.w)
-                    .any(|x| pixel(&frame, x, y, scale) == BUTTON_GLYPH.to_le_bytes())
-            });
-            assert!(lit, "joystick toggle drew no glyph for {mode:?}");
-        }
-    }
-
-    #[test]
-    fn host_shortcut_modifier_uses_platform_convention() {
-        #[cfg(target_os = "macos")]
-        {
-            assert!(host_shortcut_modifier_pressed(ModifiersState::SUPER));
-            assert!(!host_shortcut_modifier_pressed(ModifiersState::ALT));
-        }
-
-        #[cfg(not(target_os = "macos"))]
-        {
-            assert!(host_shortcut_modifier_pressed(ModifiersState::ALT));
-            assert!(!host_shortcut_modifier_pressed(ModifiersState::SUPER));
-        }
-    }
-
-    #[test]
-    fn named_key_parser_accepts_modifiers_and_raw_codes() {
-        assert_eq!(parse_amiga_key("ctrl"), Some(0x63));
-        assert_eq!(parse_amiga_key("left-alt"), Some(0x64));
-        assert_eq!(parse_amiga_key("ralt"), Some(0x65));
-        assert_eq!(parse_amiga_key("lami"), Some(0x66));
-        assert_eq!(parse_amiga_key("right-amiga"), Some(0x67));
-        assert_eq!(parse_amiga_key("0x04"), Some(0x04));
-        assert_eq!(parse_amiga_key("$04"), Some(0x04));
-        assert_eq!(parse_amiga_key("unknown-key"), None);
-    }
-
-    #[test]
-    fn renderer_runs_once_per_emulated_frame() {
-        assert!(should_render_emulated_frame(None, 0));
-        assert!(!should_render_emulated_frame(Some(12), 12));
-        assert!(should_render_emulated_frame(Some(12), 13));
-    }
-
-    #[test]
-    fn warp_burst_decouples_emulation_from_the_vsync_present() {
-        let mut app = test_app();
-        // test_app builds an unpaced (warp) emulator. Default warp level is
-        // Max: retire many frames per presented frame, bounded by a wall-clock
-        // budget so the loop still presents at vsync.
-        app.warp_speed = WarpSpeed::Max;
-        let (cap, budget) = app.warp_burst_plan(false);
-        assert!(cap > 1, "warp Max must skip output frames, got cap {cap}");
-        assert!(budget.is_some(), "Max bounds the burst by wall-clock time");
-
-        // A fixed level retires exactly its multiplier in frames, with no time
-        // bound -- predictable speed = level x refresh rate.
-        app.warp_speed = WarpSpeed::X4;
-        assert_eq!(app.warp_burst_plan(false), (4, None));
-
-        // Headless capture renders every frame, so the burst must not engage
-        // even though the core is unpaced.
-        assert_eq!(app.warp_burst_plan(true), (1, None));
-
-        // Real-time pacing presents one frame per loop regardless of level.
-        app.emu.set_paced(true);
-        assert_eq!(app.warp_burst_plan(false), (1, None));
-    }
-
-    #[test]
-    fn cycle_warp_speed_walks_the_levels() {
-        let mut app = test_app();
-        app.warp_speed = WarpSpeed::X8;
-        app.cycle_warp_speed();
-        assert_eq!(app.warp_speed, WarpSpeed::X16);
-        app.cycle_warp_speed();
-        assert_eq!(app.warp_speed, WarpSpeed::Max);
-        app.cycle_warp_speed();
-        assert_eq!(app.warp_speed, WarpSpeed::X2);
-    }
-
-    #[test]
-    fn mouse_delta_integrator_keeps_fractional_remainder() {
-        let mut delta = 0.75;
-        assert_eq!(take_integral_mouse_delta(&mut delta), 0);
-        assert_eq!(delta, 0.75);
-
-        delta += 0.5;
-        assert_eq!(take_integral_mouse_delta(&mut delta), 1);
-        assert_eq!(delta, 0.25);
-
-        delta -= 1.5;
-        assert_eq!(take_integral_mouse_delta(&mut delta), -1);
-        assert_eq!(delta, -0.25);
-    }
-
-    #[test]
-    fn status_bar_draws_hdd_led_only_on_ide_machines() {
-        let scale = 1;
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        // Row 2 of 3 is where the HDD LED sits on an IDE machine.
-        let hdd = led_row_rect(2, 3);
-
-        // No IDE port: the HDD row stays status-bar background.
-        draw_status_bar(
-            &mut frame,
-            &view(FrontPanelStatus::default(), true, false),
-            scale,
-        );
-        assert_eq!(
-            pixel(&frame, hdd.x + hdd.w / 2, hdd.y + hdd.h / 2, scale),
-            STATUS_BG.to_le_bytes()
-        );
-
-        draw_status_bar(
-            &mut frame,
-            &view(
-                FrontPanelStatus {
-                    hdd_led: Some(false),
-                    ..FrontPanelStatus::default()
-                },
-                true,
-                false,
-            ),
-            scale,
-        );
-        assert_eq!(
-            pixel(&frame, hdd.x + hdd.w / 2, hdd.y + hdd.h / 2, scale),
-            HDD_LED_OFF.to_le_bytes()
-        );
-
-        draw_status_bar(
-            &mut frame,
-            &view(
-                FrontPanelStatus {
-                    hdd_led: Some(true),
-                    ..FrontPanelStatus::default()
-                },
-                true,
-                false,
-            ),
-            scale,
-        );
-        assert_eq!(
-            pixel(&frame, hdd.x + hdd.w / 2, hdd.y + hdd.h / 2, scale),
-            HDD_LED_ON.to_le_bytes()
-        );
-    }
-
-    #[test]
-    fn status_bar_draws_power_and_fdd_led_states() {
-        let scale = 1;
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        draw_status_bar(
-            &mut frame,
-            &view(
-                FrontPanelStatus {
-                    power_led_on: true,
-                    fdd_led_on: false,
-                    fdd_track: Some(5),
-                    hdd_led: None,
-                    cd_led: None,
-                    output_volume_percent: 100,
-                },
-                true,
-                false,
-            ),
-            scale,
-        );
-
-        let power = led_row_rect(0, 2);
-        let fdd = led_row_rect(1, 2);
-        assert_eq!(
-            pixel(&frame, power.x + power.w / 2, power.y + power.h / 2, scale),
-            POWER_LED_ON.to_le_bytes()
-        );
-        assert_eq!(
-            pixel(&frame, fdd.x + fdd.w / 2, fdd.y + fdd.h / 2, scale),
-            FDD_LED_OFF.to_le_bytes()
-        );
-        let hundreds = fdd_track_digit_rect(0);
-        let ones = fdd_track_digit_rect(2);
-        assert_eq!(
-            pixel(
-                &frame,
-                hundreds.x + hundreds.w / 2,
-                hundreds.y + hundreds.h / 2,
-                scale
-            ),
-            TRACK_SEGMENT_OFF.to_le_bytes()
-        );
-        assert_eq!(
-            pixel(&frame, ones.x + ones.w / 2, ones.y + ones.h / 2, scale),
-            TRACK_SEGMENT_ON.to_le_bytes()
-        );
-
-        draw_status_bar(
-            &mut frame,
-            &view(
-                FrontPanelStatus {
-                    power_led_on: false,
-                    fdd_led_on: true,
-                    fdd_track: Some(42),
-                    hdd_led: None,
-                    cd_led: None,
-                    output_volume_percent: 100,
-                },
-                true,
-                false,
-            ),
-            scale,
-        );
-        assert_eq!(
-            pixel(&frame, fdd.x + fdd.w / 2, fdd.y + fdd.h / 2, scale),
-            FDD_LED_ON.to_le_bytes()
-        );
-    }
-
-    #[test]
-    fn status_bar_extinguishes_power_led_when_host_power_is_off() {
-        let scale = 1;
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        draw_status_bar(
-            &mut frame,
-            &view(
-                FrontPanelStatus {
-                    power_led_on: true,
-                    fdd_led_on: false,
-                    fdd_track: Some(0),
-                    hdd_led: None,
-                    cd_led: None,
-                    output_volume_percent: 100,
-                },
-                false,
-                false,
-            ),
-            scale,
-        );
-
-        let power = led_row_rect(0, 2);
-        assert_eq!(
-            pixel(&frame, power.x + power.w / 2, power.y + power.h / 2, scale),
-            POWER_LED_OFF.to_le_bytes()
-        );
-    }
-
-    #[test]
-    fn status_bar_stacks_fdd_led_under_power_led() {
-        let power = led_row_rect(0, 2);
-        let fdd = led_row_rect(1, 2);
-        let track = fdd_track_counter_rect();
-        let layout = bar_layout(&single_drive_media());
-
-        assert_eq!(fdd.x, power.x);
-        assert_eq!(fdd.w, power.w);
-        assert!(fdd.y >= power.y + power.h);
-        assert!(track.x >= power.x + power.w);
-        assert!(layout.drive_load[0].unwrap().x >= track.x + track.w);
-    }
-
-    #[test]
-    fn status_bar_power_button_glyph_tracks_power_state() {
-        let scale = 1;
-        let button = power_button_rect();
-        // A pixel squarely on the glyph's vertical bar (button centre
-        // column, a few rows above centre), where line coverage is full.
-        let gx = button.x + button.w / 2;
-        let gy = button.y + button.h / 2 - 3;
-
-        for (powered_on, expected) in [(true, POWER_GLYPH_ON), (false, POWER_GLYPH_OFF)] {
-            let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-            draw_status_bar(
-                &mut frame,
-                &view(
-                    FrontPanelStatus {
-                        power_led_on: powered_on,
-                        fdd_led_on: false,
-                        fdd_track: Some(0),
-                        hdd_led: None,
-                        cd_led: None,
-                        output_volume_percent: 100,
-                    },
-                    powered_on,
-                    false,
-                ),
-                scale,
-            );
-            assert_eq!(pixel(&frame, gx, gy, scale), expected.to_le_bytes());
-        }
-    }
-
-    #[test]
-    fn test_screen_paints_colour_bars_over_a_grey_wedge() {
-        use crate::video::{FB_HEIGHT, FB_PIXELS, FB_WIDTH};
-        let mut fb = vec![0u32; FB_PIXELS];
-        paint_test_screen(&mut fb);
-
-        // Top region: leftmost bar is grey, rightmost is blue.
-        assert_eq!(fb[0], rgba(192, 192, 192));
-        assert_eq!(fb[FB_WIDTH - 1], rgba(0, 0, 192));
-
-        // Bottom region: grey wedge runs from black at the left to white
-        // at the right.
-        let bottom = (FB_HEIGHT - 1) * FB_WIDTH;
-        assert_eq!(fb[bottom], rgba(0, 0, 0));
-        assert_eq!(fb[bottom + FB_WIDTH - 1], rgba(255, 255, 255));
-    }
-
-    #[test]
-    fn embedded_brand_assets_decode_with_transparent_edges() {
-        let logo = copperline_logo_image().expect("embedded logo PNG");
-        assert_eq!((logo.width, logo.height), (620, 128));
-        assert_eq!(logo.rgba[3], 0);
-        assert!(logo.rgba.chunks_exact(4).any(|px| px[3] == 0xFF));
-
-        let icon = copperline_icon_image().expect("embedded icon PNG");
-        assert_eq!((icon.width, icon.height), (256, 256));
-        assert_eq!(icon.rgba[3], 0);
-        assert!(icon.rgba.chunks_exact(4).any(|px| px[3] == 0xFF));
-    }
-
-    #[test]
-    fn test_screen_blits_copperline_logo_over_colour_bars() {
-        let mut fb = vec![0u32; FB_PIXELS];
-        paint_test_screen(&mut fb);
-
-        let logo = copperline_logo_image().expect("embedded logo PNG");
-        let (idx, px) = logo
-            .rgba
-            .chunks_exact(4)
-            .enumerate()
-            .find(|(_, px)| px[3] == 0xFF)
-            .expect("opaque logo pixel");
-        let x = FB_WIDTH.saturating_sub(logo.width) / 2 + idx % logo.width;
-        let y = (FB_HEIGHT * 4 / 5).saturating_sub(logo.height) / 2 + idx / logo.width;
-
-        assert_eq!(
-            fb[y * FB_WIDTH + x],
-            rgba(px[0] as u32, px[1] as u32, px[2] as u32)
-        );
-    }
-
-    #[test]
-    fn power_button_sits_left_of_reboot_without_overlap() {
-        let power = power_button_rect();
-        let reboot = reboot_button_rect();
-        let volume = volume_slider_track_rect();
-        assert!(power.x + power.w <= reboot.x);
-        assert!(power.x >= volume.x + volume.w);
-        assert_eq!(power.y, reboot.y);
-        assert_eq!(power.h, reboot.h);
-    }
-
-    #[test]
-    fn pause_button_sits_left_of_power_without_overlap() {
-        let pause = pause_button_rect();
-        let power = power_button_rect();
-        let volume = volume_slider_track_rect();
-        assert!(pause.x + pause.w <= power.x);
-        assert!(pause.x >= volume.x + volume.w);
-        assert_eq!(pause.y, power.y);
-        assert_eq!(pause.h, power.h);
-    }
-
-    #[test]
-    fn status_bar_pause_button_glyph_tracks_pause_state() {
-        let scale = 1;
-        let button = pause_button_rect();
-        // Centre column, a few rows above centre: on the play triangle's
-        // body when paused and between the pause bars when running.
-        let cx = button.x + button.w / 2;
-        let cy = button.y + button.h / 2;
-
-        // Paused: a play triangle fills the centre column.
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        draw_status_bar(
-            &mut frame,
-            &view(
-                FrontPanelStatus {
-                    power_led_on: true,
-                    fdd_led_on: false,
-                    fdd_track: Some(0),
-                    hdd_led: None,
-                    cd_led: None,
-                    output_volume_percent: 100,
-                },
-                true,
-                true,
-            ),
-            scale,
-        );
-        assert_eq!(pixel(&frame, cx, cy, scale), BUTTON_GLYPH.to_le_bytes());
-
-        // Running: the gap between the two pause bars leaves the centre
-        // column on the button face.
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        draw_status_bar(
-            &mut frame,
-            &view(
-                FrontPanelStatus {
-                    power_led_on: true,
-                    fdd_led_on: false,
-                    fdd_track: Some(0),
-                    hdd_led: None,
-                    cd_led: None,
-                    output_volume_percent: 100,
-                },
-                true,
-                false,
-            ),
-            scale,
-        );
-        assert_ne!(pixel(&frame, cx, cy, scale), BUTTON_GLYPH.to_le_bytes());
-    }
-
-    #[test]
-    fn status_bar_draws_disk_image_button_next_to_track_counter() {
-        let scale = 1;
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        draw_status_bar(
-            &mut frame,
-            &view(
-                FrontPanelStatus {
-                    power_led_on: true,
-                    fdd_led_on: true,
-                    fdd_track: Some(5),
-                    hdd_led: None,
-                    cd_led: None,
-                    output_volume_percent: 50,
-                },
-                true,
-                false,
-            ),
-            scale,
-        );
-
-        let layout = bar_layout(&single_drive_media());
-        let button = layout.drive_load[0].unwrap();
-        let track = fdd_track_counter_rect();
-        assert!(button.x >= track.x + track.w);
-        assert_eq!(
-            pixel(&frame, button.x + 5, button.y + 11, scale),
-            DISK_BODY.to_le_bytes()
-        );
-        assert_eq!(
-            pixel(&frame, button.x + button.w / 2, button.y + 15, scale),
-            DISK_LABEL.to_le_bytes()
-        );
-        // The drive number 0 is written on the disk label (top-left of
-        // the 3x5 digit).
-        assert_eq!(
-            pixel(&frame, button.x + 12, button.y + 12, scale),
-            DISK_BODY_SHADOW.to_le_bytes()
-        );
-    }
-
-    #[test]
-    fn status_bar_draws_swap_and_eject_buttons_with_enable_states() {
-        let scale = 1;
-        let mut bar = single_drive_media();
-        let layout = bar_layout(&bar);
-        let swap = layout.drive_swap[0].unwrap();
-        let eject = layout.drive_eject[0].unwrap();
-        assert!(swap.x >= layout.drive_load[0].unwrap().x + 22);
-        assert!(eject.x >= swap.x + swap.w);
-
-        // Single disk, inserted: swap is dim, eject is live.
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        let mut v = view(FrontPanelStatus::default(), true, false);
-        draw_status_bar(&mut frame, &v, scale);
-        assert_eq!(
-            pixel(&frame, swap.x + 5, swap.y + 8, scale),
-            BUTTON_GLYPH_DISABLED.to_le_bytes()
-        );
-        assert_eq!(
-            pixel(&frame, eject.x + 5, eject.y + 15, scale),
-            BUTTON_GLYPH.to_le_bytes()
-        );
-
-        // Playlist queued, no disk in: swap is live, eject is dim.
-        bar.drives[0].multi = true;
-        bar.drives[0].inserted = false;
-        v.media = bar;
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        draw_status_bar(&mut frame, &v, scale);
-        assert_eq!(
-            pixel(&frame, swap.x + 5, swap.y + 8, scale),
-            BUTTON_GLYPH.to_le_bytes()
-        );
-        assert_eq!(
-            pixel(&frame, eject.x + 5, eject.y + 15, scale),
-            BUTTON_GLYPH_DISABLED.to_le_bytes()
-        );
-    }
-
-    #[test]
-    fn status_bar_draws_cd_buttons_only_on_cd_machines() {
-        assert!(bar_layout(&media(1, None)).cd_load.is_none());
-        assert!(bar_layout(&media(1, None)).cd_eject.is_none());
-
-        let bar = media(1, Some(true));
-        let layout = bar_layout(&bar);
-        let cd_load = layout.cd_load.unwrap();
-        let cd_eject = layout.cd_eject.unwrap();
-        assert!(cd_load.x >= layout.drive_eject[0].unwrap().x + 16);
-        assert!(cd_eject.x >= cd_load.x + cd_load.w);
-
-        let scale = 1;
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        let v = StatusBarView {
-            status: FrontPanelStatus::default(),
-            powered_on: true,
-            paused: false,
-            media: bar,
-            joystick_input_mode: JoystickInputMode::Gamepad,
-            hover: None,
-        };
-        draw_status_bar(&mut frame, &v, scale);
-        // The disc body below the hub.
-        assert_eq!(
-            pixel(&frame, cd_load.x + 11, cd_load.y + 17, scale),
-            CD_BODY.to_le_bytes()
-        );
-        assert_eq!(
-            pixel(&frame, cd_eject.x + 5, cd_eject.y + 15, scale),
-            BUTTON_GLYPH.to_le_bytes()
-        );
-    }
-
-    #[test]
-    fn status_bar_draws_cd_led_on_cd_machines() {
-        let scale = 1;
-
-        // CDTV/CD32 without IDE: rows are PWR, FDD, CD at the classic
-        // three-row spacing.
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        let mut v = view(
-            FrontPanelStatus {
-                cd_led: Some(true),
-                ..FrontPanelStatus::default()
-            },
-            true,
-            false,
-        );
-        v.media = media(1, Some(true));
-        draw_status_bar(&mut frame, &v, scale);
-        let cd = led_row_rect(2, 3);
-        assert_eq!(
-            pixel(&frame, cd.x + cd.w / 2, cd.y + cd.h / 2, scale),
-            CD_LED_ON.to_le_bytes()
-        );
-
-        // With IDE as well, all four rows pack tighter and the CD LED is
-        // the last row, still inside the bar.
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        let mut v = view(
-            FrontPanelStatus {
-                hdd_led: Some(false),
-                cd_led: Some(false),
-                ..FrontPanelStatus::default()
-            },
-            true,
-            false,
-        );
-        v.media = media(1, Some(true));
-        draw_status_bar(&mut frame, &v, scale);
-        let cd = led_row_rect(3, 4);
-        assert!(cd.y + cd.h <= present_height() + super::STATUS_BAR_HEIGHT);
-        assert_eq!(
-            pixel(&frame, cd.x + cd.w / 2, cd.y + cd.h / 2, scale),
-            CD_LED_OFF.to_le_bytes()
-        );
-
-        // No CD drive: the CD row is absent and that area stays bar
-        // background.
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        draw_status_bar(
-            &mut frame,
-            &view(FrontPanelStatus::default(), true, false),
-            scale,
-        );
-        let cd = led_row_rect(2, 3);
-        assert_eq!(
-            pixel(&frame, cd.x + cd.w / 2, cd.y + cd.h / 2, scale),
-            STATUS_BG.to_le_bytes()
-        );
-    }
-
-    #[test]
-    fn bar_layout_stacks_three_or_more_drives_two_up() {
-        // One or two drives sit in a single full-height row.
-        let flat = bar_layout(&media(2, Some(true)));
-        let df0 = flat.drive_load[0].unwrap();
-        let df1 = flat.drive_load[1].unwrap();
-        assert_eq!(df0.y, df1.y);
-        assert_eq!(df0.h, super::STATUS_CONTROL_H);
-
-        // Three or four drives stack in a two-column grid: DF2 sits
-        // below DF0 in shorter buttons.
-        let stacked = bar_layout(&media(4, Some(true)));
-        let df0 = stacked.drive_load[0].unwrap();
-        let df1 = stacked.drive_load[1].unwrap();
-        let df2 = stacked.drive_load[2].unwrap();
-        let df3 = stacked.drive_load[3].unwrap();
-        assert_eq!(df0.y, df1.y);
-        assert_eq!(df2.y, df3.y);
-        assert_eq!(df0.x, df2.x);
-        assert_eq!(df1.x, df3.x);
-        assert!(df2.y >= df0.y + df0.h);
-        assert!(df0.h < super::STATUS_CONTROL_H);
-        // The grid clears the track counter on the left and the volume
-        // control on the right, CD cluster included.
-        assert!(df0.x >= fdd_track_counter_rect().x + fdd_track_counter_rect().w);
-        let cd_eject = stacked.cd_eject.unwrap();
-        assert!(cd_eject.x + cd_eject.w <= VOLUME_GLYPH_X);
-        // Stacked buttons stay inside the status bar.
-        assert!(df2.y + df2.h <= present_height() + super::STATUS_BAR_HEIGHT);
-    }
-
-    #[test]
-    fn control_at_maps_media_and_screenshot_buttons() {
-        let layout = bar_layout(&media(2, Some(false)));
-        let centre = |r: super::Rect| ((r.x + r.w / 2) as i32, (r.y + r.h / 2) as i32);
-
-        assert_eq!(
-            control_at(centre(layout.drive_load[0].unwrap()), &layout),
-            Some(BarControl::DriveLoad(0))
-        );
-        assert_eq!(
-            control_at(centre(layout.drive_swap[1].unwrap()), &layout),
-            Some(BarControl::DriveSwap(1))
-        );
-        assert_eq!(
-            control_at(centre(layout.drive_eject[1].unwrap()), &layout),
-            Some(BarControl::DriveEject(1))
-        );
-        assert_eq!(
-            control_at(centre(layout.cd_load.unwrap()), &layout),
-            Some(BarControl::CdLoad)
-        );
-        assert_eq!(
-            control_at(centre(layout.cd_eject.unwrap()), &layout),
-            Some(BarControl::CdEject)
-        );
-        assert_eq!(
-            control_at(centre(shot_button_rect()), &layout),
-            Some(BarControl::Screenshot)
-        );
-        assert_eq!(
-            control_at(centre(pause_button_rect()), &layout),
-            Some(BarControl::Pause)
-        );
-        // No drive 2 connected: the space right of its would-be cluster
-        // is empty bar.
-        assert_eq!(control_at((2, 2), &layout), None);
-    }
-
-    #[test]
-    fn status_bar_draws_volume_control_and_maps_pointer_position() {
-        let scale = 1;
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        draw_status_bar(
-            &mut frame,
-            &view(
-                FrontPanelStatus {
-                    power_led_on: true,
-                    fdd_led_on: true,
-                    fdd_track: Some(5),
-                    hdd_led: None,
-                    cd_led: None,
-                    output_volume_percent: 50,
-                },
-                true,
-                false,
-            ),
-            scale,
-        );
-
-        let track = volume_slider_track_rect();
-        assert_eq!(volume_percent_from_pos((track.x as i32, track.y as i32)), 0);
-        assert_eq!(
-            volume_percent_from_pos(((track.x + track.w - 1) as i32, track.y as i32)),
-            100
-        );
-        assert_eq!(
-            pixel(&frame, track.x + track.w / 4, track.y + track.h / 2, scale),
-            VOLUME_FILL.to_le_bytes()
-        );
-    }
-
-    #[test]
-    fn status_bar_latches_fdd_track_when_no_drive_is_selected() {
-        let mut last_fdd_track = None;
-        let status = status_with_latched_fdd_track(
-            FrontPanelStatus {
-                power_led_on: true,
-                fdd_led_on: true,
-                fdd_track: Some(42),
-                hdd_led: None,
-                cd_led: None,
-                output_volume_percent: 100,
-            },
-            &mut last_fdd_track,
-        );
-        assert_eq!(status.fdd_track, Some(42));
-        assert_eq!(last_fdd_track, Some(42));
-
-        let status = status_with_latched_fdd_track(
-            FrontPanelStatus {
-                power_led_on: true,
-                fdd_led_on: false,
-                fdd_track: None,
-                hdd_led: None,
-                cd_led: None,
-                output_volume_percent: 100,
-            },
-            &mut last_fdd_track,
-        );
-        assert_eq!(status.fdd_track, Some(42));
-    }
-
-    #[test]
-    fn status_bar_draws_at_hidpi_texture_scale() {
-        let scale = 2;
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-
-        draw_status_bar(
-            &mut frame,
-            &view(
-                FrontPanelStatus {
-                    power_led_on: true,
-                    fdd_led_on: true,
-                    fdd_track: Some(159),
-                    hdd_led: None,
-                    cd_led: None,
-                    output_volume_percent: 100,
-                },
-                true,
-                true,
-            ),
-            scale,
-        );
-
-        let power = led_row_rect(0, 2);
-        assert_eq!(
-            pixel(
-                &frame,
-                (power.x + power.w / 2) * scale,
-                (power.y + power.h / 2) * scale,
-                scale
-            ),
-            POWER_LED_ON.to_le_bytes()
-        );
-        let ones = fdd_track_digit_rect(2);
-        assert_eq!(
-            pixel(
-                &frame,
-                (ones.x + ones.w / 2) * scale,
-                (ones.y + ones.h / 2) * scale,
-                scale
-            ),
-            TRACK_SEGMENT_ON.to_le_bytes()
-        );
-    }
-
-    #[test]
-    fn present_frame_copy_scales_texture_rows_at_hidpi() {
-        use crate::video::deinterlace::{OUT_HEIGHT, OUT_PIXELS};
-        let scale = 2;
-        let mut src = vec![0u32; OUT_PIXELS];
-        src[0] = 0x1122_3344;
-        src[1] = 0x5566_7788;
-        src[(OUT_HEIGHT - 1) * FB_WIDTH] = 0xAABB_CCDD;
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-
-        copy_present_frame(&src, OUT_HEIGHT, &mut frame, scale);
-
-        // The top output row samples the top source row exactly (the
-        // centre-aligned position clamps at the edge), and horizontal
-        // duplication carries each source pixel across the HiDPI pair.
-        assert_eq!(pixel(&frame, 0, 0, scale), src[0].to_le_bytes());
-        assert_eq!(pixel(&frame, 1, 0, scale), src[0].to_le_bytes());
-        assert_eq!(pixel(&frame, 2, 0, scale), src[1].to_le_bytes());
-        // The bottom output row resolves to the last woven source line.
-        assert_eq!(
-            pixel(&frame, 0, present_height() * scale - 1, scale),
-            src[(OUT_HEIGHT - 1) * FB_WIDTH].to_le_bytes()
-        );
-    }
-
-    #[test]
-    fn tv_window_copy_centres_reference_aperture_in_live_texture() {
-        use crate::video::deinterlace::{OUT_HEIGHT, OUT_PIXELS};
-        let scale = 1;
-        let mut src = vec![0u32; OUT_PIXELS];
-        let row_y = TV_PAL_PRESENT_SOURCE_Y;
-        let standard_left = crate::video::bitplane::STANDARD_VISIBLE_X0;
-        let standard_right = standard_left + 320 * 2 - 1;
-        let left_marker = 0x1122_3344u32;
-        let right_marker = 0x5566_7788u32;
-        let left_edge = 0x99AA_BBCCu32;
-        let right_edge = 0xDDEE_FF00u32;
-
-        src[row_y * FB_WIDTH + TV_PAL_PRESENT_SOURCE_X] = left_edge;
-        src[row_y * FB_WIDTH + FB_WIDTH - 1] = right_edge;
-        src[row_y * FB_WIDTH + standard_left] = left_marker;
-        src[row_y * FB_WIDTH + standard_right] = right_marker;
-
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        copy_window_present_frame(&src, OUT_HEIGHT, &mut frame, scale, Overscan::Tv, true);
-
-        let dst_standard_left = TV_PAL_LIVE_PAD_X + (standard_left - TV_PAL_PRESENT_SOURCE_X);
-        let dst_standard_right = dst_standard_left + 320 * 2 - 1;
-        assert_eq!(dst_standard_left, FB_WIDTH - 1 - dst_standard_right);
-        assert_eq!(
-            pixel(&frame, dst_standard_left, 0, scale),
-            left_marker.to_le_bytes()
-        );
-        assert_eq!(
-            pixel(&frame, dst_standard_right, 0, scale),
-            right_marker.to_le_bytes()
-        );
-        assert_eq!(pixel(&frame, 0, 0, scale), left_edge.to_le_bytes());
-        assert_eq!(
-            pixel(&frame, FB_WIDTH - 1, 0, scale),
-            right_edge.to_le_bytes()
-        );
-    }
-
-    #[test]
-    fn tv_window_copy_preserves_true_overscan_fetches() {
-        use crate::video::deinterlace::{OUT_HEIGHT, OUT_PIXELS};
-        let scale = 1;
-        let mut src = vec![0u32; OUT_PIXELS];
-        let left_overscan = 0x1122_3344u32;
-        let standard_crop_edge = 0x5566_7788u32;
-
-        src[0] = left_overscan;
-        src[TV_PAL_PRESENT_SOURCE_X] = standard_crop_edge;
-
-        let mut frame = vec![0u8; texture_width(scale) * texture_height(scale) * 4];
-        copy_window_present_frame(&src, OUT_HEIGHT, &mut frame, scale, Overscan::Tv, false);
-
-        assert_eq!(pixel(&frame, 0, 0, scale), left_overscan.to_le_bytes());
-        assert_ne!(pixel(&frame, 0, 0, scale), standard_crop_edge.to_le_bytes());
-    }
-
-    #[test]
-    fn square_pixel_canvas_maps_woven_rows_one_to_one() {
-        use crate::video::deinterlace::OUT_HEIGHT;
-        use crate::video::PRESENT_HEIGHT_SQUARE;
-        // The square-pixel canvas is exactly the woven field: every
-        // scanline is one output row, so a standard 320x256 PAL display
-        // occupies precisely 640x512 output pixels.
-        assert_eq!(PRESENT_HEIGHT_SQUARE, OUT_HEIGHT);
-        for y in 0..OUT_HEIGHT {
-            assert_eq!(
-                crate::screenshot::scaled_source_row(y, OUT_HEIGHT, PRESENT_HEIGHT_SQUARE),
-                y
-            );
-        }
-    }
-
-    #[test]
-    fn tv_aperture_row_mapping_pads_square_bezel_and_covers_43() {
-        use crate::video::{PRESENT_HEIGHT_SQUARE, PRESENT_HEIGHT_TV};
-        // 4:3 canvas: no bezel rows; the 540 aperture rows map onto all
-        // 537 output rows exactly as before the square-pixel option.
-        assert_eq!(tv_aperture_source_row(0, PRESENT_HEIGHT_TV, 1), Some(0));
-        assert_eq!(
-            tv_aperture_source_row(PRESENT_HEIGHT_TV - 1, PRESENT_HEIGHT_TV, 1),
-            Some(TV_PAL_PRESENT_HEIGHT - 1)
-        );
-        // Square canvas: black bezel bands centre the aperture and its
-        // rows map 1:1.
-        let pad = (PRESENT_HEIGHT_SQUARE - TV_PAL_PRESENT_HEIGHT) / 2;
-        assert_eq!(tv_aperture_source_row(0, PRESENT_HEIGHT_SQUARE, 1), None);
-        assert_eq!(
-            tv_aperture_source_row(pad - 1, PRESENT_HEIGHT_SQUARE, 1),
-            None
-        );
-        assert_eq!(
-            tv_aperture_source_row(pad, PRESENT_HEIGHT_SQUARE, 1),
-            Some(0)
-        );
-        assert_eq!(
-            tv_aperture_source_row(pad + TV_PAL_PRESENT_HEIGHT - 1, PRESENT_HEIGHT_SQUARE, 1),
-            Some(TV_PAL_PRESENT_HEIGHT - 1)
-        );
-        assert_eq!(
-            tv_aperture_source_row(pad + TV_PAL_PRESENT_HEIGHT, PRESENT_HEIGHT_SQUARE, 1),
-            None
-        );
-        // HiDPI: bezel and 1:1 mapping scale with the texture factor.
-        assert_eq!(
-            tv_aperture_source_row(2 * pad - 1, PRESENT_HEIGHT_SQUARE, 2),
-            None
-        );
-        assert_eq!(
-            tv_aperture_source_row(2 * pad, PRESENT_HEIGHT_SQUARE, 2),
-            Some(0)
-        );
-        assert_eq!(
-            tv_aperture_source_row(2 * pad + 1, PRESENT_HEIGHT_SQUARE, 2),
-            Some(0)
-        );
-        assert_eq!(
-            tv_aperture_source_row(2 * pad + 2, PRESENT_HEIGHT_SQUARE, 2),
-            Some(1)
-        );
-    }
-
-    #[test]
-    fn present_row_selection_covers_every_source_line_at_hidpi() {
-        use crate::video::deinterlace::OUT_HEIGHT;
-        // The live window writes a HiDPI texture before the OS compositor
-        // scales it. At that output size, every woven source row should be
-        // represented by one or more whole texture rows without mixing
-        // neighbouring Amiga scanlines.
-        let out_rows = present_height() * 2;
-        let mut hits = vec![0usize; OUT_HEIGHT];
-        let mut prev = 0usize;
-        for y in 0..out_rows {
-            let src_y = crate::screenshot::scaled_source_row(y, OUT_HEIGHT, out_rows);
-            assert!(src_y < OUT_HEIGHT);
-            assert!(src_y >= prev);
-            hits[src_y] += 1;
-            prev = src_y;
-        }
-        for (y, count) in hits.iter().enumerate() {
-            assert!(
-                *count > 0,
-                "source row {y} dropped from the presentation entirely"
-            );
-            assert!(
-                *count <= 2,
-                "source row {y} has unexpectedly thick presentation coverage: {count}"
-            );
-        }
-    }
-
-    #[test]
-    fn standard_pal_frames_get_vertical_presentation_margin() {
-        let standard_offset = presentation_source_y_offset(STANDARD_PAL_VISIBLE_START_VPOS);
-
-        assert!(standard_offset > 0);
-        assert_eq!(
-            presentation_source_y_offset(STANDARD_PAL_VISIBLE_START_VPOS - standard_offset as u32),
-            0
-        );
-    }
-
-    #[test]
-    fn horizontal_centering_shifts_left_and_blacks_the_right() {
-        let mut fb = vec![rgba(0, 0, 0); FB_PIXELS];
-        let marker = rgba(0x12, 0x34, 0x56);
-        // A marker 30px in on the first row, and one at the right edge.
-        fb[30] = marker;
-        fb[FB_WIDTH - 1] = rgba(0x99, 0x88, 0x77);
-
-        center_present_frame_horizontally(&mut fb, 26);
-
-        // Content moved left by 26: x=30 -> x=4.
-        assert_eq!(fb[4], marker);
-        assert_eq!(fb[30], rgba(0, 0, 0));
-        // The right 26 columns are blacked out.
-        for x in (FB_WIDTH - 26)..FB_WIDTH {
-            assert_eq!(fb[x], rgba(0, 0, 0));
-        }
-    }
-
-    #[test]
-    fn horizontal_centering_is_a_noop_for_zero_shift() {
-        let mut fb = vec![rgba(0, 0, 0); FB_PIXELS];
-        let marker = rgba(1, 2, 3);
-        fb[100] = marker;
-        center_present_frame_horizontally(&mut fb, 0);
-        assert_eq!(fb[100], marker);
-    }
-
-    #[test]
-    fn tv_presentation_keeps_standard_hires_framebuffer_origin() {
-        let snapshot = RenderRegisterSnapshot {
-            bplcon0: 0x0200,
-            diwstrt: 0x0581,
-            diwstop: 0x40C1,
-            ddfstrt: 0x003C,
-            ddfstop: 0x00D0,
-            ..RenderRegisterSnapshot::default()
-        };
-
-        assert_eq!(presentation_h_shift_for(&snapshot, Overscan::Tv), 0);
-    }
-
-    #[test]
-    fn tv_pal_crop_centres_standard_display_in_aperture() {
-        let standard_left = crate::video::bitplane::STANDARD_VISIBLE_X0;
-        let standard_right = standard_left + 640;
-
-        assert_eq!(TV_PAL_PRESENT_WIDTH, 692);
-        assert_eq!(TV_PAL_PRESENT_HEIGHT, 540);
-        assert_eq!(standard_left - TV_PAL_PRESENT_SOURCE_X, 26);
-        assert_eq!(
-            TV_PAL_PRESENT_WIDTH - (standard_right - TV_PAL_PRESENT_SOURCE_X),
-            26
-        );
-        assert_eq!(TV_PAL_PRESENT_SOURCE_Y, 18);
-    }
-
-    #[test]
-    fn standard_pal_frame_centering_preserves_horizontal_margin() {
-        let offset = presentation_source_y_offset(STANDARD_PAL_VISIBLE_START_VPOS);
-        let mut fb = vec![rgba(0, 0, 0); FB_PIXELS];
-        let marker = rgba(0x12, 0x34, 0x56);
-        fb[32] = marker;
-
-        center_present_frame_for_visible_start(&mut fb, STANDARD_PAL_VISIBLE_START_VPOS);
-
-        assert_eq!(fb[0], rgba(0, 0, 0));
-        assert_eq!(fb[(offset * FB_WIDTH) + 31], rgba(0, 0, 0));
-        assert_eq!(fb[(offset * FB_WIDTH) + 32], marker);
-    }
-
-    #[test]
-    fn tv_overscan_mask_blacks_margins_and_keeps_the_tv_window() {
-        let marker = rgba(0x12, 0x34, 0x56);
-        let mut fb = vec![marker; FB_PIXELS];
-
-        // A standard screen after vertical presentation: window top at
-        // framebuffer row 14 (the standard centring offset), with the TV
-        // aperture anchored to the emulated framebuffer.
-        let std_top = standard_window_top_row(STANDARD_PAL_VISIBLE_START_VPOS);
-        let shift = 0;
-        mask_present_frame_to_tv(&mut fb, shift, std_top);
-
-        let (source_left, source_right) = tv_source_h_bounds();
-        let left = source_left - shift;
-        let right = source_right - shift;
-        let mid_row = std_top + 100;
-        assert_eq!(right, FB_WIDTH);
-        assert_eq!(fb[mid_row * FB_WIDTH + left - 1], rgba(0, 0, 0));
-        assert_eq!(fb[mid_row * FB_WIDTH + left], marker);
-        assert_eq!(fb[mid_row * FB_WIDTH + right - 1], marker);
-        assert_eq!(fb[mid_row * FB_WIDTH + FB_WIDTH - 1], marker);
-        // The deep left overscan margin stays hidden.
-        assert_eq!(fb[mid_row * FB_WIDTH], rgba(0, 0, 0));
-        // Vertical border rows remain visible; the TV mask only hides the
-        // deep horizontal margins.
-        assert_eq!(fb[(std_top - 1) * FB_WIDTH + left - 1], rgba(0, 0, 0));
-        assert_eq!(fb[(std_top - 1) * FB_WIDTH + left], marker);
-        let bottom = std_top + STANDARD_PAL_VISIBLE_LINES - 1;
-        assert_eq!(fb[bottom * FB_WIDTH + left - 1], rgba(0, 0, 0));
-        assert_eq!(fb[bottom * FB_WIDTH + left], marker);
-    }
-
-    #[test]
-    fn tv_mask_preserves_visible_left_overscan_margin_without_shifting() {
-        let std_top = standard_window_top_row(STANDARD_PAL_VISIBLE_START_VPOS);
-        let mid_row = std_top + 100;
-        let (source_left, _) = tv_source_h_bounds();
-        let marker = rgba(0x12, 0x34, 0x56);
-        let hidden_marker = rgba(0x98, 0x76, 0x54);
-        let mut fb = vec![rgba(0, 0, 0); FB_PIXELS];
-
-        fb[mid_row * FB_WIDTH + source_left - 1] = hidden_marker;
-        fb[mid_row * FB_WIDTH + source_left] = marker;
-
-        mask_present_frame_to_tv(&mut fb, 0, std_top);
-
-        assert_eq!(fb[mid_row * FB_WIDTH + source_left - 1], rgba(0, 0, 0));
-        assert_eq!(fb[mid_row * FB_WIDTH + source_left], marker);
-    }
-
-    #[test]
-    fn tv_overscan_mask_tracks_the_centering_shift() {
-        let marker = rgba(0x12, 0x34, 0x56);
-        let mut fb = vec![marker; FB_PIXELS];
-
-        // A standard display shifted left 16px for centring: the bezel
-        // moves with it, so the window's left edge is not clipped.
-        let std_top = standard_window_top_row(STANDARD_PAL_VISIBLE_START_VPOS);
-        mask_present_frame_to_tv(&mut fb, 16, std_top);
-
-        let (source_left, source_right) = tv_source_h_bounds();
-        let left = source_left - 16;
-        let right = source_right - 16;
-        let mid_row = std_top + 100;
-        assert_eq!(fb[mid_row * FB_WIDTH + left - 1], rgba(0, 0, 0));
-        assert_eq!(fb[mid_row * FB_WIDTH + left], marker);
-        assert_eq!(fb[mid_row * FB_WIDTH + right - 1], marker);
-        assert_eq!(fb[mid_row * FB_WIDTH + right], rgba(0, 0, 0));
-        assert_eq!(fb[mid_row * FB_WIDTH + FB_WIDTH - 1], rgba(0, 0, 0));
-    }
-
-    #[test]
-    fn tv_overscan_mask_preserves_vertical_border_rows() {
-        let marker = rgba(0x12, 0x34, 0x56);
-        let mut fb = vec![marker; FB_PIXELS];
-        let std_top = standard_window_top_row(STANDARD_PAL_VISIBLE_START_VPOS);
-        let shift = 0;
-        let (source_left, source_right) = tv_source_h_bounds();
-        let left = source_left - shift;
-        let right = source_right - shift;
-        let bottom = std_top + STANDARD_PAL_VISIBLE_LINES - 1;
-
-        mask_present_frame_to_tv(&mut fb, shift, std_top);
-
-        assert_eq!(right, FB_WIDTH);
-        assert_eq!(fb[bottom * FB_WIDTH + left - 1], rgba(0, 0, 0));
-        assert_eq!(fb[bottom * FB_WIDTH + left], marker);
-        assert_eq!(fb[bottom * FB_WIDTH + right - 1], marker);
-    }
-
-    #[test]
-    fn tv_overscan_mask_tracks_overscan_visible_starts() {
-        // A deep-overscan frame (visible start above the standard window):
-        // the centring shift is consumed, the standard window sits lower in
-        // the framebuffer, and the TV window follows it down.
-        let visible_start = STANDARD_PAL_VISIBLE_START_VPOS - 16;
-        let std_top = standard_window_top_row(visible_start);
-        assert_eq!(std_top, 16);
-
-        let marker = rgba(0x12, 0x34, 0x56);
-        let mut fb = vec![marker; FB_PIXELS];
-        mask_present_frame_to_tv(&mut fb, 0, std_top);
-
-        let (left, _) = tv_source_h_bounds();
-        assert_eq!(
-            fb[std_top.saturating_sub(1) * FB_WIDTH + left - 1],
-            rgba(0, 0, 0)
-        );
-        assert_eq!(fb[std_top.saturating_sub(1) * FB_WIDTH + left], marker);
-        let bottom = std_top + STANDARD_PAL_VISIBLE_LINES - 1;
-        assert_eq!(fb[bottom * FB_WIDTH + left - 1], rgba(0, 0, 0));
-        assert_eq!(fb[bottom * FB_WIDTH + left], marker);
-    }
-
-    #[test]
-    fn reboot_button_hit_rect_is_bounded_to_status_bar_button() {
-        let button = reboot_button_rect();
-
-        assert!(button.contains(((button.x + 1) as i32, (button.y + 1) as i32)));
-        assert!(button.contains((
-            (button.x + button.w - 1) as i32,
-            (button.y + button.h - 1) as i32
-        )));
-        assert!(!button.contains(((button.x - 1) as i32, button.y as i32)));
-        assert!(!button.contains((button.x as i32, (button.y - 1) as i32)));
-    }
-
-    fn pixel(frame: &[u8], x: usize, y: usize, scale: usize) -> [u8; 4] {
-        frame[(y * texture_width(scale) + x) * 4..(y * texture_width(scale) + x) * 4 + 4]
-            .try_into()
-            .unwrap()
-    }
-
-    /// An interactive App around a minimal machine: a NOP-sled ROM with
-    /// reset vectors pointing into it, no audio, unpaced. Lets the
-    /// debugger window's actions and view builders run against the real
-    /// emulator without a host window.
-    fn test_app() -> super::App {
-        test_app_with_audio(Box::new(NullSink))
-    }
-
-    fn test_app_with_audio(audio: Box<dyn AudioSink>) -> super::App {
-        use crate::chipset::paula::Paula;
-        use crate::config::{CpuModel, PacingBudget};
-        use crate::emulator::Emulator;
-        use crate::floppy::FloppyController;
-        use crate::memory::{Memory, ROM_BASE, ROM_SIZE};
-        use crate::serial::StdoutSink;
-
-        let mut rom = vec![0u8; ROM_SIZE];
-        let pc = ROM_BASE as u32 + 8;
-        rom[0..4].copy_from_slice(&0x0007_FFFEu32.to_be_bytes());
-        rom[4..8].copy_from_slice(&pc.to_be_bytes());
-        // NOP sled for the rest of the test program.
-        for word in rom[8..4096].chunks_exact_mut(2) {
-            word.copy_from_slice(&0x4E71u16.to_be_bytes());
-        }
-        let mem = Memory {
-            chip_ram: vec![0u8; 512 * 1024],
-            slow_ram: Vec::new(),
-            rom,
-            overlay: true,
-            zorro: crate::zorro::ZorroChain::default(),
-            extended_rom: Vec::new(),
-            extended_rom_base: 0,
-            wcs: Vec::new(),
-            wcs_write_protected: false,
-        };
-        let bus = crate::bus::Bus::new(
-            mem,
-            Paula::new(Box::new(StdoutSink::new()), audio),
-            FloppyController::default(),
-        );
-        let emu = Emulator::new(bus, CpuModel::M68000, false, PacingBudget::Cycles, 2, false)
-            .expect("test emulator");
-        super::App::new(
-            emu,
-            true,
-            None,
-            None,
-            None,
-            Vec::new(),
-            Vec::new(),
-            Vec::new(),
-            Vec::new(),
-            Vec::new(),
-            None,
-            std::array::from_fn(|_| Vec::new()),
-            [true; 4],
-            crate::config::Overscan::Full,
-            0.0,
-            crate::config::WarpSpeed::Max,
-            crate::config::JoystickInputMode::Gamepad,
-            vec!["Machine: test".to_string()],
-            crate::config::RawConfig::default(),
-        )
-    }
-
-    #[test]
-    fn launcher_panel_edits_machine_setup() {
-        use crate::config::MachineModel;
-        use crate::video::launcher::{LauncherField, LauncherTab};
-
-        let mut app = test_app();
-        app.open_launcher();
-        assert!(matches!(app.ui.panel, Some(Panel::Launcher(_))));
-
-        // Pick a machine, switch tabs, and flip a toggle through the same
-        // control dispatch the mouse uses.
-        app.activate_ui_control(UiControl::LauncherModel(MachineModel::A1200));
-        app.activate_ui_control(UiControl::LauncherTab(LauncherTab::Cpu));
-        app.activate_ui_control(UiControl::LauncherToggle(LauncherField::Fpu));
-
-        let state = match &app.ui.panel {
-            Some(Panel::Launcher(state)) => state,
-            _ => panic!("launcher closed unexpectedly"),
-        };
-        assert_eq!(state.setup.model(), Some(MachineModel::A1200));
-        assert_eq!(state.tab, LauncherTab::Cpu);
-        // The A1200's profile defaults (AGA, EC020, 2M chip) plus the FPU we
-        // toggled on are what a save would emit.
-        let raw = state.setup.to_raw();
-        assert_eq!(raw.machine.profile.as_deref(), Some("A1200"));
-        assert_eq!(raw.cpu.fpu, Some(true));
-        assert!(state.setup.build_config().is_ok());
-    }
-
-    #[test]
-    fn launcher_run_keeps_panel_open_on_error() {
-        use crate::video::launcher::LauncherField;
-
-        let mut app = test_app();
-        app.powered_on = false;
-        app.open_launcher();
-        // A floppy image that does not exist fails config validation.
-        if let Some(Panel::Launcher(state)) = app.ui.panel.as_mut() {
-            state
-                .setup
-                .set_path(LauncherField::Df0Image, PathBuf::from("/no/such/disk.adf"));
-        }
-        app.launcher_run();
-        match &app.ui.panel {
-            Some(Panel::Launcher(state)) => assert!(
-                state.status.as_ref().is_some_and(|s| s.error),
-                "expected an error status to keep the user in the launcher"
-            ),
-            _ => panic!("launcher should stay open on a validation error"),
-        }
-        assert!(
-            !app.powered_on,
-            "a failed Run must not power the machine on"
-        );
-    }
-
-    #[test]
-    fn state_load_closes_launcher_and_powers_restored_machine() {
-        let path = std::env::temp_dir().join(format!(
-            "copperline-launcher-state-load-{}.clstate",
-            std::process::id()
-        ));
-        let mut app = test_app();
-        app.emu.save_state(&path).expect("save test state");
-
-        app.power_off();
-        let parked_present = app.present_fb.clone();
-        app.open_launcher();
-        assert!(matches!(app.ui.panel, Some(Panel::Launcher(_))));
-
-        assert!(app.load_state_from_path(&path));
-        assert!(app.ui.panel.is_none(), "state load should dismiss launcher");
-        assert!(app.powered_on);
-        assert!(!app.cpu_halted);
-        assert!(
-            app.present_fb == parked_present,
-            "load itself should not invent a rendered frame"
-        );
-
-        for _ in 0..3 {
-            app.emu.step_frame().expect("step restored frame");
-            if app.finish_render_for_current_frame() {
-                break;
-            }
-        }
-        assert_ne!(
-            app.present_fb, parked_present,
-            "restored machine should render over the parked test screen"
-        );
-
-        let _ = std::fs::remove_file(&path);
-    }
-
-    struct SuspensionSink {
-        states: Rc<RefCell<Vec<bool>>>,
-    }
-
-    impl AudioSink for SuspensionSink {
-        fn push(&mut self, _left: f32, _right: f32) {}
-
-        fn flush(&mut self) {}
-
-        fn set_live_output_suspended(&mut self, suspended: bool) {
-            self.states.borrow_mut().push(suspended);
-        }
-    }
-
-    #[test]
-    fn host_pause_states_suspend_live_audio_output() {
-        let states = Rc::new(RefCell::new(Vec::new()));
-        let mut app = test_app_with_audio(Box::new(SuspensionSink {
-            states: Rc::clone(&states),
-        }));
-
-        app.toggle_pause();
-        assert_eq!(states.borrow().last(), Some(&true));
-        app.toggle_pause();
-        assert_eq!(states.borrow().last(), Some(&false));
-
-        app.power_off();
-        assert_eq!(states.borrow().last(), Some(&true));
-        app.toggle_power();
-        assert_eq!(states.borrow().last(), Some(&false));
-
-        app.open_debugger();
-        assert_eq!(states.borrow().last(), Some(&true));
-        app.debugger_toggle_run();
-        assert_eq!(states.borrow().last(), Some(&false));
-    }
-
-    #[test]
-    fn host_io_audio_suspension_restores_current_run_state() {
-        let states = Rc::new(RefCell::new(Vec::new()));
-        let mut app = test_app_with_audio(Box::new(SuspensionSink {
-            states: Rc::clone(&states),
-        }));
-
-        app.suspend_live_audio_for_host_io();
-        app.finish_host_io_pause();
-        assert_eq!(states.borrow().as_slice(), &[true, false]);
-
-        app.toggle_pause();
-        app.suspend_live_audio_for_host_io();
-        app.finish_host_io_pause();
-        assert_eq!(states.borrow().last(), Some(&true));
-    }
-
-    #[test]
-    fn restoring_over_placeholder_detected_only_for_the_silent_config_screen() {
-        // The exact configuration-screen placeholder: powered off, launcher
-        // open, NullSink installed (as build_placeholder_machine produces).
-        let mut app = test_app();
-        app.power_off();
-        app.open_launcher();
-        assert!(
-            app.restoring_over_placeholder(),
-            "powered-off launcher with a null sink is the placeholder"
-        );
-
-        // A live (non-null) sink behind the launcher is a real running session
-        // re-opening the config screen: its audio must not be torn out.
-        let states = Rc::new(RefCell::new(Vec::new()));
-        let mut live = test_app_with_audio(Box::new(SuspensionSink {
-            states: Rc::clone(&states),
-        }));
-        live.power_off();
-        live.open_launcher();
-        assert!(
-            !live.restoring_over_placeholder(),
-            "a real audio sink behind the launcher is not the placeholder"
-        );
-
-        // A null sink but powered on, or with no launcher open, is not the
-        // pre-boot placeholder either.
-        let mut powered = test_app();
-        powered.open_launcher();
-        assert!(powered.powered_on);
-        assert!(!powered.restoring_over_placeholder());
-
-        let mut no_launcher = test_app();
-        no_launcher.power_off();
-        assert!(no_launcher.ui.panel.is_none());
-        assert!(!no_launcher.restoring_over_placeholder());
-    }
-
-    #[test]
-    fn state_load_over_running_session_keeps_its_live_audio_sink() {
-        // Re-opening the config screen over a running machine and loading a
-        // state must not replace the live audio sink (the regression guard for
-        // the placeholder-upgrade path). Uses a probe sink so no real audio
-        // device is touched.
-        let path = std::env::temp_dir().join(format!(
-            "copperline-running-state-load-{}.clstate",
-            std::process::id()
-        ));
-        let states = Rc::new(RefCell::new(Vec::new()));
-        let mut app = test_app_with_audio(Box::new(SuspensionSink {
-            states: Rc::clone(&states),
-        }));
-        app.emu.save_state(&path).expect("save test state");
-        app.open_launcher();
-        assert!(app.powered_on);
-        assert!(!app.restoring_over_placeholder());
-
-        assert!(app.load_state_from_path(&path));
-
-        // The probe sink is still the installed one: a suspension toggle still
-        // reaches it. (A replacement CpalSink would have dropped the probe.)
-        states.borrow_mut().clear();
-        app.suspend_live_audio_for_host_io();
-        app.finish_host_io_pause();
-        assert_eq!(
-            states.borrow().as_slice(),
-            &[true, false],
-            "live audio sink should survive a state load over a running session"
-        );
-
-        let _ = std::fs::remove_file(&path);
-    }
-
-    /// End-to-end recording through the app: start, run emulated frames
-    /// through the same render/capture path as the event loop, stop, and
-    /// check the resulting AVI carries the frames and matching audio.
-    /// COPPERLINE_RECORDER_KEEP=1 keeps the file for playback checks.
-    #[test]
-    fn recording_captures_emulated_frames_with_audio() {
-        let mut app = test_app();
-        let path = std::env::temp_dir().join(format!(
-            "copperline-app-recording-{}.avi",
-            std::process::id()
-        ));
-        for warmup_step in 0..4 {
-            app.emu.step_frame().expect("step frame");
-            let rendered = if app.render_worker.is_some() {
-                app.finish_render_for_current_frame()
-            } else {
-                app.render_emulated_frame_if_needed()
-            };
-            if rendered {
-                break;
-            }
-            assert!(
-                warmup_step < 3,
-                "fixture should produce an initial renderable frame"
-            );
-        }
-
-        app.start_recording_to(path.clone());
-        assert!(app.recorder.is_some(), "recorder should be active");
-
-        let frames_to_record = 5;
-        let mut rendered_frames = 0;
-        let mut step_quanta = 0;
-        while rendered_frames < frames_to_record {
-            app.emu.step_frame().expect("step frame");
-            let rendered = if app.render_worker.is_some() {
-                app.finish_render_for_current_frame()
-            } else {
-                app.render_emulated_frame_if_needed()
-            };
-            app.capture_recorder_output(rendered);
-            if rendered {
-                rendered_frames += 1;
-            }
-            step_quanta += 1;
-            assert!(
-                step_quanta <= frames_to_record * 2,
-                "fixture should keep producing renderable frames"
-            );
-        }
-        app.stop_recording();
-        assert!(app.recorder.is_none());
-        // Stopping again is a no-op, and Paula's tap is off.
-        app.stop_recording();
-        assert!(app.emu.bus_mut().paula.take_captured_audio().is_empty());
-
-        let data = std::fs::read(&path).expect("recording file exists");
-        if crate::envcfg::flag("COPPERLINE_RECORDER_KEEP") {
-            eprintln!("kept {}", path.display());
-        } else {
-            std::fs::remove_file(&path).unwrap();
-        }
-        assert_eq!(&data[0..4], b"RIFF");
-        assert_eq!(&data[8..12], b"AVI ");
-        assert_eq!(&data[112..116], b"ZMBV");
-        // avih dwTotalFrames at offset 48 (see recorder::build_header).
-        let frames = u32::from_le_bytes(data[48..52].try_into().unwrap());
-        assert_eq!(frames, frames_to_record);
-        // Audio stream length (samples) at offset 264 should cover the
-        // same emulated interval: ~882 mixer samples per PAL frame or
-        // ~735 per NTSC frame (the fixture machine's standard).
-        let audio_len = u32::from_le_bytes(data[264..268].try_into().unwrap());
-        let per_frame = audio_len as f64 / frames_to_record as f64;
-        assert!(
-            (700.0..=920.0).contains(&per_frame),
-            "audio samples per frame {per_frame}"
-        );
-    }
-
-    #[test]
-    fn debugger_window_pauses_steps_and_restores_run_state() {
-        let mut app = test_app();
-        assert!(!app.paused);
-
-        // Opening pauses; the memory view starts at the PC's page.
-        app.toggle_debugger();
-        assert!(app.paused);
-        let pc_before = app.emu.machine.pc();
-        match app.debugger_panel.as_ref() {
-            Some(panel) => {
-                assert_eq!(panel.mem_addr, pc_before & 0x00FF_FFF0);
-            }
-            _ => panic!("debugger panel should be open"),
-        }
-
-        // Step executes exactly one instruction (a 2-byte NOP).
-        app.debugger_step();
-        assert_eq!(app.emu.machine.pc(), pc_before.wrapping_add(2));
-
-        // Run to a nearby address lands exactly there.
-        let target = pc_before.wrapping_add(10);
-        if let Some(panel) = app.debugger_panel.as_mut() {
-            panel.entry = format!("{target:X}");
-        }
-        app.debugger_run_to();
-        assert_eq!(app.emu.machine.pc() & 0x00FF_FFFF, target & 0x00FF_FFFF);
-
-        // Step Frame advances emulated time by one whole frame.
-        let frame_before = app.emu.bus().emulated_frames();
-        app.debugger_step_frame();
-        assert!(app.emu.bus().emulated_frames() > frame_before);
-
-        // Closing restores the pre-debugger (running) state.
-        app.toggle_debugger();
-        assert!(app.debugger_panel.is_none());
-        assert!(!app.paused);
-
-        // Run pressed inside the debugger survives closing it.
-        app.toggle_debugger();
-        assert!(app.paused);
-        app.debugger_toggle_run();
-        assert!(!app.paused);
-        app.close_tool_panel(ToolPanelKind::Debugger);
-        assert!(!app.paused);
-    }
-
-    #[test]
-    fn debugger_and_frame_analyzer_can_stay_open_together() {
-        let mut app = test_app();
-        assert!(!app.paused);
-
-        app.open_debugger();
-        assert!(app.paused);
-        assert!(app.debugger_panel.is_some());
-        assert!(app.frame_analyzer_panel.is_none());
-
-        app.open_frame_analyzer();
-        assert!(app.paused);
-        assert!(app.debugger_panel.is_some());
-        assert!(app.frame_analyzer_panel.is_some());
-
-        app.close_tool_panel(ToolPanelKind::FrameAnalyzer);
-        assert!(app.paused, "debugger should keep the machine paused");
-        assert!(app.debugger_panel.is_some());
-        assert!(app.frame_analyzer_panel.is_none());
-
-        app.close_tool_panel(ToolPanelKind::Debugger);
-        assert!(!app.paused);
-        assert!(app.debugger_panel.is_none());
-
-        let mut app = test_app();
-        app.open_frame_analyzer();
-        assert!(app.paused);
-        app.open_debugger();
-        assert!(app.debugger_panel.is_some());
-        assert!(app.frame_analyzer_panel.is_some());
-
-        app.close_tool_panel(ToolPanelKind::Debugger);
-        assert!(app.paused, "analyzer should keep the machine paused");
-        assert!(app.debugger_panel.is_none());
-        assert!(app.frame_analyzer_panel.is_some());
-
-        app.close_tool_panel(ToolPanelKind::FrameAnalyzer);
-        assert!(!app.paused);
-        assert!(app.frame_analyzer_panel.is_none());
-    }
-
-    #[test]
-    fn debugger_views_reflect_machine_state() {
-        let mut app = test_app();
-        app.open_debugger();
-
-        let pc = app.emu.machine.pc();
-        for tab in super::ui::DEBUG_TABS {
-            if let Some(panel) = app.debugger_panel.as_mut() {
-                panel.tab = tab;
-            }
-            let Some(panel) = app.debugger_panel.as_ref() else {
-                unreachable!()
-            };
-            let view = app.build_debugger_view(panel);
-            assert!(!view.lines.is_empty());
-            match tab {
-                super::ui::DebugTab::Cpu => {
-                    assert!(view.lines[0].text.contains(&format!("PC {pc:08X}")));
-                    // The disassembly cursor line is highlighted and decodes
-                    // the NOP sled.
-                    let cursor = view
-                        .lines
-                        .iter()
-                        .find(|line| line.highlight)
-                        .expect("a highlighted PC line");
-                    assert!(cursor.text.contains("NOP"), "{}", cursor.text);
-                }
-                super::ui::DebugTab::Chipset => {
-                    assert!(view.lines.iter().any(|l| l.text.starts_with("DMACON")));
-                    assert!(view.lines.iter().any(|l| l.text.starts_with("INTENA")));
-                    assert!(view.lines.iter().any(|l| l.text.contains("COLOR00")));
-                }
-                super::ui::DebugTab::Copper => {
-                    assert!(view.lines[0].text.contains("COP1LC"));
-                }
-                super::ui::DebugTab::Audio => {
-                    // Text is mirrored into `lines` for the fallback/invariant.
-                    assert!(view.lines[0].text.starts_with("DMACON"));
-                    assert!(view.lines[0].text.contains("ADKCON"));
-                    assert!(view.lines.iter().any(|l| l.text.starts_with("AUD0")));
-                    assert!(view.lines.iter().any(|l| l.text.starts_with("AUD3")));
-                    // The structured view drives the graphical layout: four
-                    // Paula channels plus a CD row.
-                    let audio = view.audio.as_ref().expect("audio scope view");
-                    assert!(audio.header.starts_with("DMACON"));
-                    assert_eq!(audio.channels.len(), 4);
-                    assert!(audio.channels[0].text[0].text.starts_with("AUD0"));
-                    assert!(audio.cd.text[0].text.contains("CD-DA"));
-                }
-                super::ui::DebugTab::Memory => {
-                    // The hex dump shows the NOP sled at the PC's ROM page.
-                    assert!(view.lines.iter().any(|l| l.text.contains("4E 71")));
-                }
-                super::ui::DebugTab::Break => {
-                    assert!(view.lines.iter().any(|l| l.text == "Breakpoints:"));
-                    assert!(view.lines.iter().any(|l| l.text == "  (none)"));
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn audio_tab_mute_buttons_toggle_paula_mutes() {
-        let mut app = test_app();
-        app.open_debugger();
-        if let Some(panel) = app.debugger_panel.as_mut() {
-            panel.tab = super::ui::DebugTab::Audio;
-        }
-        // Clicking a channel's mute button toggles that Paula channel's mute
-        // through the full dispatch path.
-        assert!(!app.emu.bus().paula.channel_muted(1));
-        app.activate_ui_control(super::ui::UiControl::DebugAudioMute(1));
-        assert!(app.emu.bus().paula.channel_muted(1));
-        app.activate_ui_control(super::ui::UiControl::DebugAudioMute(1));
-        assert!(!app.emu.bus().paula.channel_muted(1));
-        // Index 4 is the CD-DA mute.
-        assert!(!app.emu.bus().paula.cd_muted());
-        app.activate_ui_control(super::ui::UiControl::DebugAudioMute(4));
-        assert!(app.emu.bus().paula.cd_muted());
-    }
-
-    #[test]
-    fn interactive_breakpoint_pauses_and_reopens_the_debugger() {
-        let mut app = test_app();
-        app.open_debugger();
-
-        // Toggle a breakpoint a few instructions ahead via the entry box.
-        let target = app.emu.machine.pc().wrapping_add(8) & 0x00FF_FFFF;
-        if let Some(panel) = app.debugger_panel.as_mut() {
-            panel.tab = super::ui::DebugTab::Break;
-            panel.entry = format!("{target:X}");
-        }
-        app.activate_ui_control(super::ui::UiControl::DebugBreakToggle);
-        assert!(app.emu.machine.ui_breaks().is_breakpoint(target));
-
-        // The Break tab lists it.
-        if let Some(panel) = app.debugger_panel.as_ref() {
-            let view = app.build_debugger_view(panel);
-            assert!(view
-                .lines
-                .iter()
-                .any(|l| l.text.contains(&format!("${target:06X}"))));
-        }
-
-        // Close the panel (machine resumes) and run a frame: the hit
-        // pauses the machine at the breakpoint, before it executes.
-        app.close_tool_panel(ToolPanelKind::Debugger);
-        assert!(!app.paused);
-        app.emu.step_frame().expect("frame");
-        assert!(app.surface_debug_stop());
-        assert!(app.paused);
-        assert_eq!(app.emu.machine.pc() & 0x00FF_FFFF, target);
-        assert!(app.debugger_panel.is_some());
-        assert!(app
-            .last_debug_stop
-            .as_deref()
-            .is_some_and(|s| s.contains("Breakpoint")));
-
-        // Resuming does not immediately re-trip the same breakpoint.
-        app.debugger_toggle_run();
-        assert!(app.last_debug_stop.is_none());
-        app.emu.step_frame().expect("frame");
-        assert_ne!(app.emu.machine.pc() & 0x00FF_FFFF, target);
-
-        // Toggling the same address again removes the breakpoint.
-        if let Some(panel) = app.debugger_panel.as_mut() {
-            panel.entry = format!("{target:X}");
-        }
-        app.activate_ui_control(super::ui::UiControl::DebugBreakToggle);
-        assert!(!app.emu.machine.ui_breaks().is_breakpoint(target));
-    }
-
-    #[test]
-    fn opening_the_debugger_arms_reverse_and_step_reconstructs() {
-        let mut app = test_app();
-        // Opening the debugger auto-arms the reverse snapshot ring.
-        app.open_debugger();
-        assert!(app.emu.time_travel_enabled());
-        if let Some(panel) = app.debugger_panel.as_ref() {
-            assert!(
-                app.build_debugger_view(panel).reverse_available,
-                "reverse controls should be enabled once armed"
-            );
-        }
-
-        // Advance enough frames that several snapshots accrue.
-        for _ in 0..16 {
-            app.debugger_step_frame();
-        }
-        let pos_before = app.emu.retired_instructions();
-        let pc_before = app.emu.machine.pc();
-        assert!(pos_before > 0, "the NOP sled retired instructions");
-
-        // Reverse Step moves the position strictly backward.
-        app.activate_ui_control(super::ui::UiControl::DebugReverseStep);
-        let pos_after = app.emu.retired_instructions();
-        assert_eq!(pos_after, pos_before - 1, "stepped back exactly one");
-
-        // Replaying forward to the original position reconstructs the PC.
-        for _ in 0..16 {
-            app.debugger_step_frame();
-        }
-        assert!(app.emu.retired_instructions() >= pos_before);
-        let frame_before = app.emu.bus().emulated_frames();
-        let pos_before_frame = app.emu.retired_instructions();
-        assert!(frame_before > 0, "frame history should have advanced");
-
-        // Reverse Frame moves to the previous Agnus frame counter value.
-        app.activate_ui_control(super::ui::UiControl::DebugReverseFrame);
-        assert_eq!(
-            app.emu.bus().emulated_frames(),
-            frame_before - 1,
-            "stepped back exactly one emulated video frame"
-        );
-        assert!(
-            app.emu.retired_instructions() < pos_before_frame,
-            "reverse frame should move to an earlier instruction boundary"
-        );
-
-        // And reverse-continue with no breakpoints is a no-op (reports, does
-        // not move): position is unchanged afterward.
-        let pos = app.emu.retired_instructions();
-        app.activate_ui_control(super::ui::UiControl::DebugReverseRun);
-        assert_eq!(app.emu.retired_instructions(), pos);
-        let _ = pc_before;
-    }
-
-    #[test]
-    fn interactive_watchpoint_stops_when_the_word_changes() {
-        let mut app = test_app();
-        // Map chip RAM at $0 so the watched word is real memory.
-        app.emu.machine.disable_overlay();
-        let addr = 0x0000_1000u32;
-        assert!(app.emu.machine.ui_toggle_watch(addr));
-
-        // Unchanged memory: a full frame runs without stopping.
-        app.emu.step_frame().expect("frame");
-        assert!(!app.surface_debug_stop());
-
-        // Change the watched word (as any non-CPU bus master would); the
-        // next executed instruction notices and stops the machine.
-        app.emu.bus_mut().mem.chip_ram[addr as usize] = 0xAB;
-        app.emu.step_frame().expect("frame");
-        assert!(app.surface_debug_stop());
-        assert!(app.paused);
-        assert!(app
-            .last_debug_stop
-            .as_deref()
-            .is_some_and(|s| s.contains("Watch $001000")));
-    }
-
-    #[test]
-    fn chipset_register_watch_stops_on_a_cpu_write() {
-        let mut app = test_app();
-        // Replace part of the NOP sled with MOVE.W #$8020,$DFF096
-        // (DMACON), a few instructions ahead of the PC so the already
-        // prefetched words are not affected.
-        let pc = app.emu.machine.pc();
-        let off = (pc as usize & 0x7FFFF) + 8;
-        let mov: [u16; 4] = [0x33FC, 0x8020, 0x00DF, 0xF096];
-        for (k, word) in mov.iter().enumerate() {
-            app.emu.bus_mut().mem.rom[off + k * 2..off + k * 2 + 2]
-                .copy_from_slice(&word.to_be_bytes());
-        }
-
-        // Watch DMACON via the entry box, accepting the full address form.
-        app.open_debugger();
-        if let Some(panel) = app.debugger_panel.as_mut() {
-            panel.tab = super::ui::DebugTab::Break;
-            panel.entry = "DFF096".to_string();
-        }
-        app.activate_ui_control(super::ui::UiControl::DebugRegToggle);
-        assert_eq!(app.emu.machine.ui_breaks().reg_watches, [0x096]);
-        app.debugger_toggle_run();
-        app.close_tool_panel(ToolPanelKind::Debugger);
-
-        app.emu.step_frame().expect("frame");
-        assert!(app.surface_debug_stop());
-        assert!(app.paused);
-        let stop = app.last_debug_stop.as_deref().unwrap();
-        assert!(stop.contains("DMACON"), "{stop}");
-        assert!(stop.contains("8020"), "{stop}");
-        assert!(stop.contains("cpu write"), "{stop}");
-    }
-
-    #[test]
-    fn modal_panel_swallows_amiga_key_presses() {
-        let mut app = test_app();
-        app.ui.panel = Some(Panel::About);
-
-        // Escape closes the panel.
-        assert!(app.ui_handle_key(KeyCode::Escape));
-        assert!(app.ui.panel.is_none());
-
-        // Hex entry: digits accumulate, Enter commits to the memory view.
-        app.open_debugger();
-        if let Some(panel) = app.debugger_panel.as_mut() {
-            panel.entry_active = true;
-            panel.tab = super::ui::DebugTab::Memory;
-        }
-        for key in [
-            KeyCode::KeyC,
-            KeyCode::Digit0,
-            KeyCode::Digit0,
-            KeyCode::Digit1,
-        ] {
-            assert!(app.ui_handle_key(key));
-        }
-        assert!(app.ui_handle_key(KeyCode::Enter));
-        match app.debugger_panel.as_ref() {
-            Some(panel) => {
-                assert_eq!(panel.entry, "C001");
-                assert_eq!(panel.mem_addr, 0xC000);
-                assert!(!panel.entry_active);
-            }
-            _ => panic!("debugger panel should be open"),
-        }
-    }
-
-    #[test]
-    fn frame_analyzer_cursor_keys_move_selected_slot() {
-        let mut app = test_app();
-        app.open_frame_analyzer();
-        app.frame_analyzer_step_frame();
-        assert!(app.emu.bus().frame_bus_trace().is_some());
-        assert!(app.ui_key_accepts_repeat(Some(ToolPanelKind::FrameAnalyzer), KeyCode::ArrowRight));
-        assert!(!app.ui_key_accepts_repeat(Some(ToolPanelKind::FrameAnalyzer), KeyCode::KeyR));
-
-        let (start_hpos, start_vpos) = match app.frame_analyzer_panel.as_ref() {
-            Some(panel) => (panel.selected_hpos, panel.selected_vpos),
-            _ => panic!("frame analyzer panel should be open"),
-        };
-
-        assert!(app.ui_handle_key(KeyCode::ArrowRight));
-        assert!(app.ui_handle_key(KeyCode::ArrowDown));
-        match app.frame_analyzer_panel.as_ref() {
-            Some(panel) => {
-                assert_eq!(panel.selected_hpos, start_hpos + 1);
-                assert_eq!(panel.selected_vpos, start_vpos + 1);
-            }
-            _ => panic!("frame analyzer panel should be open"),
-        }
-
-        assert!(app.ui_handle_key(KeyCode::ArrowLeft));
-        assert!(app.ui_handle_key(KeyCode::ArrowUp));
-        match app.frame_analyzer_panel.as_ref() {
-            Some(panel) => {
-                assert_eq!(panel.selected_hpos, start_hpos);
-                assert_eq!(panel.selected_vpos, start_vpos);
-            }
-            _ => panic!("frame analyzer panel should be open"),
-        }
-
-        if let Some(panel) = app.frame_analyzer_panel.as_mut() {
-            panel.selected_hpos = 0;
-            panel.selected_vpos = 0;
-        }
-        assert!(app.ui_handle_key(KeyCode::ArrowLeft));
-        assert!(app.ui_handle_key(KeyCode::ArrowUp));
-        match app.frame_analyzer_panel.as_ref() {
-            Some(panel) => {
-                assert_eq!(panel.selected_hpos, 0);
-                assert_eq!(panel.selected_vpos, 0);
-            }
-            _ => panic!("frame analyzer panel should be open"),
-        }
-
-        let (max_hpos, max_vpos) = app
-            .emu
-            .bus()
-            .frame_bus_trace()
-            .map(|trace| {
-                (
-                    trace.cols.saturating_sub(1).min(u16::MAX as usize) as u16,
-                    trace.rows.saturating_sub(1).min(u16::MAX as usize) as u16,
-                )
-            })
-            .unwrap();
-        if let Some(panel) = app.frame_analyzer_panel.as_mut() {
-            panel.selected_hpos = max_hpos;
-            panel.selected_vpos = max_vpos;
-        }
-        assert!(app.ui_handle_key(KeyCode::ArrowRight));
-        assert!(app.ui_handle_key(KeyCode::ArrowDown));
-        match app.frame_analyzer_panel.as_ref() {
-            Some(panel) => {
-                assert_eq!(panel.selected_hpos, max_hpos);
-                assert_eq!(panel.selected_vpos, max_vpos);
-            }
-            _ => panic!("frame analyzer panel should be open"),
-        }
-    }
-
-    #[test]
-    fn debugger_keys_step_and_pin_disassembly() {
-        let mut app = test_app();
-        app.open_debugger();
-
-        // S steps one instruction while the entry box is unfocused.
-        let pc_before = app.emu.machine.pc();
-        assert!(app.ui_handle_key(KeyCode::KeyS));
-        assert_eq!(app.emu.machine.pc(), pc_before.wrapping_add(2));
-
-        // R toggles run; the explicit choice survives closing the panel.
-        assert!(app.paused);
-        assert!(app.ui_handle_key(KeyCode::KeyR));
-        assert!(!app.paused);
-        assert!(app.ui_handle_key(KeyCode::KeyR));
-        assert!(app.paused);
-
-        // On the CPU tab, Enter pins the disassembly origin to the typed
-        // address; an empty box follows the PC again.
-        if let Some(panel) = app.debugger_panel.as_mut() {
-            panel.entry_active = true;
-            panel.entry = "FC0010".to_string();
-        }
-        assert!(app.ui_handle_key(KeyCode::Enter));
-        match app.debugger_panel.as_ref() {
-            Some(panel) => {
-                assert_eq!(panel.disasm_addr, Some(0xFC0010));
-                let view = app.build_debugger_view(panel);
-                // Each disasm line carries a one-char breakpoint marker prefix.
-                assert!(view.lines.iter().any(|l| l.text.contains("00FC0010")));
-            }
-            _ => panic!("debugger panel should be open"),
-        }
-        if let Some(panel) = app.debugger_panel.as_mut() {
-            panel.entry_active = true;
-            panel.entry.clear();
-        }
-        assert!(app.ui_handle_key(KeyCode::Enter));
-        match app.debugger_panel.as_ref() {
-            Some(panel) => assert_eq!(panel.disasm_addr, None),
-            _ => panic!("debugger panel should be open"),
-        }
-
-        // While the entry box is focused, S types the register-name letter
-        // 'S' (for SR/SP) into the box instead of stepping.
-        if let Some(panel) = app.debugger_panel.as_mut() {
-            panel.entry_active = true;
-            panel.entry.clear();
-        }
-        let pc_before = app.emu.machine.pc();
-        assert!(app.ui_handle_key(KeyCode::KeyS));
-        assert_eq!(app.emu.machine.pc(), pc_before);
-        assert_eq!(
-            app.debugger_panel.as_ref().map(|p| p.entry.as_str()),
-            Some("S")
-        );
-    }
-
-    #[test]
-    fn debugger_poke_writes_memory_and_registers() {
-        let mut app = test_app();
-        app.open_debugger();
-        // Map chip RAM at $0 so the low test address is writable RAM, not the
-        // boot ROM overlay.
-        app.emu.machine.disable_overlay();
-
-        // Memory tab: "ADDR VALUE" writes a word into chip RAM.
-        if let Some(panel) = app.debugger_panel.as_mut() {
-            panel.tab = super::ui::DebugTab::Memory;
-            panel.entry = "2000 BEEF".to_string();
-        }
-        app.debugger_poke();
-        assert_eq!(
-            app.emu.machine.debug_read_memory(0x2000, 2),
-            vec![0xBE, 0xEF]
-        );
-
-        // CPU tab: "REG VALUE" sets a register.
-        if let Some(panel) = app.debugger_panel.as_mut() {
-            panel.tab = super::ui::DebugTab::Cpu;
-            panel.entry = "D3 12345678".to_string();
-        }
-        app.debugger_poke();
-        assert_eq!(app.emu.machine.d(3), 0x1234_5678);
-    }
-}
+mod tests;
